Create Excel Document

This function relies on the openxlsx package for creating Excel documents.

as_excel(
  ...,
  sheet_names = NULL,
  autofilter = TRUE,
  autowidth = TRUE,
  rows_zebra = TRUE,
  cols_zebra = FALSE,
  save = identical(deparse(substitute(...)), "."),
  file = NULL,
  freeze_top_row = TRUE,
  overwrite = FALSE
)

save_excel(xl, file = NULL, overwrite = FALSE, ...)

Arguments

...

data objects, use named items for multiple tabs (see Examples) zie Examples

sheet_names

sheet names

autofilter

create autofilter on columns in first row

autowidth

automatically adjust columns widths

rows_zebra

create banded rows

cols_zebra

create banded columns

save

save the file directly, defaults to TRUE if used after a pipe (%>%)

file

file location to save Excel document to, defaults to a random filename

freeze_top_row

freeze the first row of the sheet

overwrite

overwrite existing file

xl

Excel object, as created with the openxlsx package

Examples

# does not save immediately, but returns the workbook object
xl <- as_excel("this is a sheet" = mtcars,
               "another sheet" = anscombe)
#> Warning: In read_secret(): environmental variable 'secrets_file' not set
xl
#> A Workbook object.
#>  
#> Worksheets:
#>  Sheet 1: "this is a sheet"
#>  
#> 	Custom column widths (column: width)
#> 	  1: 22.8, 2: 8, 3: 8, 4: 8, 5: 8, 6: 8, 7: 8, 8: 8, 9: 8, 10: 8, 11: 8, 12: 8 
#>  
#> 
#>  Sheet 2: "another sheet"
#>  
#> 	Custom column widths (column: width)
#> 	  1: 8, 2: 8, 3: 8, 4: 8, 5: 8, 6: 8, 7: 8, 8: 8 
#>  
#> 
#>  
#>  Worksheet write order: 1, 2
#>  Active Sheet 1: "this is a sheet" 
#> 	Position: 1
#> 

# save with save_excel() or export.excel()
export.excel(xl)
#> Data exported as '/tmp/RtmpQIJChk/file2d85167b9281/reference/xl.xlsx' (12 kB).
         
if (FALSE) {
# saves immediately
if (require("dplyr")) {
  mtcars %>%
    as_excel()
}
}