Download CBS-data

Download data from CBS Open data Statline.

cbs_topics()

cbs_search(topic, max_print = 25)

cbs_download(identifier, clean_cols = TRUE)

cbs_moreinfo(identifier)

Arguments

topic

topics to search for

max_print

maximum number of subjects to print

identifier

tracking number (1 to max_print) in cbs_search()

clean_cols

Clean column names.

Details

cbs_topics() retrieves all topics.

cbs_search() searches for a specific subject.

cbs_download() downloads tables. Input has to be a CBS Identifier (printed in red in cbs_search()), or a tracking number of cbs_search(), or the result of cbs_search().

cbs_moreinfo() gives a detailed explanation for the table. Input can also be a dataset downloaded with cbs_download().

Examples

cbs_search("Inwoners")
#> 1. Bevolkingskernen 2006; steden en dorpen in Nederland (71699ned)
#> Periode: 2006
#> Grootte: 2.192 x 81
#> Aangepast: dinsdag 29 januari 2013
#> Inwoners- en woninggegevens naar bevolkingskernen
#> Kenmerken van bevolking en woningen naar bevolkingskernen
#> 
#> 2. Doodsoorzaken; maand en jaar van overlijden, 1995-2012 (71594ned)
#> Periode: 1995 januari - 2012 december
#> Grootte: 234 x 14
#> Aangepast: vrijdag 12 december 2014
#> Overleden inwoners van Nederland naar primaire doodsoorzaak per maand 
#> en jaar van overlijden
#> 
#> 3. Doodsoorzaken; niet-natuurlijke dood (inwoners), div. kenmerken, 1996-2011 (37683)
#> Periode: 1996 - 2011
#> Grootte: 480 x 109
#> Aangepast: maandag 19 mei 2014
#> Personen overleden door een niet - natuurlijke dood: zelfdoding, moord, 
#> verkeers-/bedrijfs-/privé-ongevallen en overig/onbekend
#> 
#> 4. Faillissementen naar gemeenten met meer dan 100 000 inwoners, 1993 - 1999 (7495fail)
#> Periode: 1993 - 1999.
#> Grootte: 175 x 5
#> Aangepast: dinsdag 6 maart 2001
#> Per gemeenten met meer dan 100 000 inwoners; uitgesproken 
#> faillissementen totaal, over bedrijven en instellingen en overige 
#> 
#> 5. Maatstaven Financiële-verhoudingswet (Fvw) 1997-2006 (81880NED)
#> Periode: 1997 - 2006
#> Grootte: 14.660 x 68
#> Aangepast: woensdag 30 januari 2013
#> Maatstaven Financiële verhoudingswet (fvw) 1997-2006
#> Gemeentelijke cijfers inwoners, woningen, dichtheid, centrumfunctie
#> 
#> 6. Gemeenten; inwoners en uitkeringen, 1999-2009, regio-indeling 2009  (80860ned)
#> Periode: 1999-2009
#> Grootte: 5.038 x 27
#> Aangepast: donderdag 11 april 2013
#> Kerncijfers over inwoners en personen met uitkeringen,
#> naar landsdelen, provincies en gemeenten en regio-indeling 2009
#> 
#> 7. Geregistreerde criminaliteit; politieregio en soort misdrijf, 1994-2007 (37932)
#> Periode: 1994 - 2007.
#> Grootte: 4.312 x 13
#> Aangepast: vrijdag 31 juli 2009
#> Geregistreerde misdrijven en opgehelderde misdrijven (totaal en per 
#> 100.000 inwoners 12 - 79 jaar), verdachten en omvang politieregio naar 
#> 
#> 8. Geregistreerde criminaliteit; soort misdrijf, 1999-2007 (70871ned)
#> Periode: 1999 - 2007.
#> Grootte: 324 x 20
#> Aangepast: vrijdag 31 juli 2009
#> Geregistreerde misdrijven(totaal en per 100.000 inwoners
#> 12 - 79 jaar) naar type misdrijf volgens wetboek.
#> 
#> 9. Indicatoren jeugdzorg in natura; gemeenten, 2015-2020 (82972NED)
#> Periode: 2015-2020
#> Grootte: 9.766 x 7
#> Aangepast: vrijdag 29 oktober 2021
#> Enkele indicatoren over jeugdzorggebruik in natura per gemeente
#> clustering van de gemeenten naar inwonersaantal
#> 
#> 10. Inkomens van particuliere huishoudens naar herkomst en regio; 2000 (70710ned)
#> Periode: 2000
#> Grootte: 6.300 x 12
#> Aangepast: vrijdag 19 december 2003
#> Inkomens van particuliere huishoudens naar herkomst en generatie, 
#> regio (landsdelen, provincies en gemeenten > 100 000 inwoners)
#> 
#> 11. Monitoren van de ruimtelijke ontwikkelingen in Nederland, 2006 (71717ned)
#> Periode: 2006
#> Grootte: 44.880 x 7
#> Aangepast: vrijdag 30 juli 2010
#> Inwoners en woningen naar risico- en beleidsgebieden
#> Risico- en beleidsgebieden, gemeenten
#> 
#> 12. Monitor Nota Ruimte 2008; ruimtelijke ontwikkelingen naar regio, 2000-2008 (80385ned)
#> Periode: 2008
#> Grootte: 3.282 x 80
#> Aangepast: dinsdag 17 augustus 2010
#> Inwoners en woningen naar contouren en regio 2008.
#> Risico- en beleidscontouren, gemeenten
#> 
#> 13. Openbare bibliotheken, collectie naar grootte werkgebied, 1995 (7075WERK)
#> Periode: 1995
#> Grootte: 8 x 49
#> Aangepast: maandag 9 februari 2004
#> Collectie, uitleningen, exploitatieoverzicht naar inwonersklasse
#> en werkgebied. 
#> 
#> 14. Overledenen; moord en doodslag; pleeglocatie Nederland, 1996-2018 (81453NED)
#> Periode: 1996-2018
#> Grootte: 483 x 14
#> Aangepast: woensdag 29 juli 2020
#> Overledenen door moord of doodslag die in Nederland plaatsvond (inwoners 
#> en niet-inwoners), pleeglocatie, wijze van uitvoering
#> 
#> 15. Overledenen; ongevallen, inwoners van Nederland, 1996-2017 (81452NED)
#> Periode: 1996-2017
#> Grootte: 660 x 31
#> Aangepast: woensdag 20 november 2019
#> Nederlandse inwoners overleden door dodelijke ongevallen, tijdens vervoer
#> en brand, door accidentele val, vergiftiging en verdrinking
#> 
#> 16. Verdachten; herkomst, geslacht, generatie en soort misdrijf 1999-2010 (80317ned)
#> Periode: 1999-2010
#> Grootte: 65.664 x 7
#> Aangepast: maandag 22 augustus 2011
#> Aangehouden verdachten van misdrijven, aantal en per 1 000 inwoners
#> naar herkomst, geslacht, generatie allochtonen en soort misdrijf
#> 
#> 17. Verdachten; huishoudinkomen en soort misdrijf 1999-2008 (80320ned)
#> Periode: 1999-2008
#> Grootte: 320 x 5
#> Aangepast: dinsdag 5 april 2011
#> Aangehouden verdachten van misdrijven, aantal en per 1 000 inwoners
#> naar huishoudinkomen en soort misdrijf
#> 
#> 18. Verdachten; geslacht, leeftijd, herkomstgroepering en misdrijf 1999-2011 (80316ned)
#> Periode: 1999-2011
#> Grootte: 24.960 x 7
#> Aangepast: dinsdag 7 mei 2013
#> Aangehouden verdachten van misdrijven, aantal en per 1 000 inwoners
#> naar geslacht, leeftijd, herkomst en soort misdrijf
#> 
#> 19. Verdachten; geslacht, leeftijd, herkomstgr. en pleegcarrière 1999-2010 (80318ned)
#> Periode: 1999-2010
#> Grootte: 11.520 x 7
#> Aangepast: maandag 22 augustus 2011
#> Aangehouden verdachten van misdrijven, aantal en per 1 000 inwoners
#> naar geslacht, leeftijd, herkomstgroepering en pleegcarrière
#> 
#> 20. Verdachten; stedelijkheid woongemeente en soort misdrijf 1999-2008 (80322ned)
#> Periode: 1999-2008
#> Grootte: 560 x 5
#> Aangepast: dinsdag 5 april 2011
#> Aangehouden verdachten van misdrijven, aantal en per 1 000 inwoners
#> naar stedelijkheid woongemeente en soort misdrijf
#> 
#> 21. Verdachten; soort inkomen en misdrijf 1999-2008 (80321ned)
#> Periode: 1999-2008
#> Grootte: 240 x 5
#> Aangepast: dinsdag 5 april 2011
#> Aangehouden verdachten van misdrijven, aantal en per 1 000 inwoners 
#> naar soort inkomen (werk of uitkering) en soort misdrijf.
#> 
#> 22. Volkstelling 1899; indeeling naar den leeftijd en den burgerlijken staat (7379vt)
#> Periode: 1899
#> Grootte: 104.598 x 11
#> Aangepast: vrijdag 4 mei 2018
#> Volkstelling 1899. Aantallen personen naar leeftijd, burgerlijke staat 
#> en geslacht voor gemeenten met meer dan 5000 inwoners.
#> 
#> 23. Volkstelling; plaatselijke indeeling, 1899 (7380vt)
#> Periode: 1899
#> Grootte: 17.876 x 15
#> Aangepast: donderdag 2 september 1999
#> Volkstelling 1899. Aantallen woningen en inwoners op wijkniveau.
#> Wijkniveau
#> 
#> 24. Wijken en buurten; inwoners en uitkeringen, 1999-2009, regio-indeling 2009  (80859ned)
#> Periode: 1999-2009
#> Grootte: 160.138 x 27
#> Aangepast: donderdag 11 april 2013
#> Kerncijfers over inwoners en personen met uitkeringen
#> uitgesplitst naar gemeenten, wijken en buurten
#> 
#> 25. Bevolking; ontwikkeling in gemeenten met 100 000 of meer inwoners (70748ned)
#> Periode: 1 januari 2000 - 1 januari 2021
#> Grootte: 217.668 x 88
#> Aangepast: donderdag 14 oktober 2021
#> Bevolking, huishoudens, levendgeborenen, overledenen, verhuisde personen
#> en bevolkingsgroei in gemeenten met 100 000 of meer inwoners
#> 
#> ...39

x <- cbs_download(2) # 2nd hit of cbs_search()
str(x)
#> tibble [234 × 14] (S3: tbl_df/tbl/data.frame)
#>  $ perioden                               : chr [1:234] "1995MM01" "1995MM02" "1995MM03" "1995MM04" ...
#>   ..- attr(*, "labels")= Named chr [1:234] "1995MM01" "1995MM02" "1995MM03" "1995MM04" ...
#>   .. ..- attr(*, "names")= chr [1:234] "1995 januari" "1995 februari" "1995 maart" "1995 april" ...
#>   ..- attr(*, "label")= chr "Perioden"
#>   ..- attr(*, "is_time")= logi TRUE
#>  $ totaal_alle_overledenen                : int [1:234] 11974 10106 12069 11412 11180 10508 11288 10928 10199 11030 ...
#>   ..- attr(*, "label")= chr "Totaal alle overledenen"
#>  $ totaal_nieuwvormingen                  : int [1:234] 3228 2768 3253 2946 3158 2989 3150 3112 3048 3183 ...
#>   ..- attr(*, "label")= chr "Totaal nieuwvormingen"
#>  $ kwaadaardige_nv_van_luchtpijp_en_long  : int [1:234] 759 657 735 747 757 661 761 722 674 731 ...
#>   ..- attr(*, "label")= chr "Kwaadaardige nv. van luchtpijp en long"
#>  $ kwaadaardige_nieuwvorming_van_borst    : int [1:234] 311 235 338 278 305 269 288 295 264 312 ...
#>   ..- attr(*, "label")= chr "Kwaadaardige nieuwvorming van borst"
#>  $ overige_nieuwvormingen                 : int [1:234] 2158 1876 2180 1921 2096 2059 2101 2095 2110 2140 ...
#>   ..- attr(*, "label")= chr "Overige nieuwvormingen"
#>  $ psychische_stoornis_en_zktn_zenuwstel  : int [1:234] 565 426 514 503 498 453 533 582 459 477 ...
#>   ..- attr(*, "label")= chr "Psychische stoornis. en zktn zenuwstel.."
#>  $ totaal_ziekten_van_hart_en_vaatstelsel : int [1:234] 4745 4005 4706 4440 4304 4039 4129 3995 3856 4218 ...
#>   ..- attr(*, "label")= chr "Totaal ziekten van hart en vaatstelsel"
#>  $ acuut_hartinfarct                      : int [1:234] 1395 1240 1433 1361 1303 1262 1227 1172 1140 1327 ...
#>   ..- attr(*, "label")= chr "Acuut hartinfarct"
#>  $ hersenvaatletsels                      : int [1:234] 1106 936 1125 1040 1054 933 1037 1019 923 970 ...
#>   ..- attr(*, "label")= chr "Hersenvaatletsels"
#>  $ overige_ziekten_van_hart_en_vaatstelsel: int [1:234] 2244 1829 2148 2039 1947 1844 1865 1804 1793 1921 ...
#>   ..- attr(*, "label")= chr "Overige ziekten van hart en vaatstelsel"
#>  $ ziekten_van_de_ademhalingsorganen      : int [1:234] 1083 873 1211 1166 990 856 1100 960 730 907 ...
#>   ..- attr(*, "label")= chr "Ziekten van de ademhalingsorganen"
#>  $ wegverkeersongevallen                  : int [1:234] 112 92 105 91 101 112 120 102 84 100 ...
#>   ..- attr(*, "label")= chr "Wegverkeersongevallen"
#>  $ overige_doodsoorzaken                  : int [1:234] 2241 1942 2280 2266 2129 2059 2256 2177 2022 2145 ...
#>   ..- attr(*, "label")= chr "Overige doodsoorzaken"
#>  - attr(*, "identifier")= chr "71594ned"

cbs_moreinfo(x)
#> Doodsoorzaken; maand en jaar van overlijden, 1995-2012 (71594ned)
#> Deze tabel bevat cijfers over de doodsoorzaken van overleden inwoners van Nederland. De cijfers zijn uitgesplitst per maand en jaar van overlijden naar een beperkt aantal doodsoorzaken. De cijfers zijn gebaseerd op minimaal 95% van de totale populatie.
#> 
#> Gegevens beschikbaar van januari 1995 tot en met december 2012
#> 
#> Status van de cijfers:
#> De gegevens zijn definitief.
#> 
#> Wijzigingen per 12 december 2014:
#> Geen, deze tabel is stopgezet en wordt opgevolgd door tabel Overledenen; doodsoorzaak, kwartaal en jaar overlijden (zie paragraaf 3). 
#>  
#> Nieuwe cijfers:
#> Niet meer van toepassing. 
#> 
cbs_moreinfo(2)
#> Doodsoorzaken; maand en jaar van overlijden, 1995-2012 (71594ned)
#> Deze tabel bevat cijfers over de doodsoorzaken van overleden inwoners van Nederland. De cijfers zijn uitgesplitst per maand en jaar van overlijden naar een beperkt aantal doodsoorzaken. De cijfers zijn gebaseerd op minimaal 95% van de totale populatie.
#> 
#> Gegevens beschikbaar van januari 1995 tot en met december 2012
#> 
#> Status van de cijfers:
#> De gegevens zijn definitief.
#> 
#> Wijzigingen per 12 december 2014:
#> Geen, deze tabel is stopgezet en wordt opgevolgd door tabel Overledenen; doodsoorzaak, kwartaal en jaar overlijden (zie paragraaf 3). 
#>  
#> Nieuwe cijfers:
#> Niet meer van toepassing. 
#>