Dates around Today

These objects are imported from other packages. Follow the links below to see their documentation.

lubridate

now, today

yesterday(ref = today())

tomorrow(ref = today())

last_week(ref = today(), only_start_end = FALSE)

this_week(ref = today(), only_start_end = FALSE)

next_week(ref = today(), only_start_end = FALSE)

last_month(ref = today(), only_start_end = FALSE)

this_month(ref = today(), only_start_end = FALSE)

next_month(ref = today(), only_start_end = FALSE)

last_quarter(ref = today(), only_start_end = FALSE)

this_quarter(ref = today(), only_start_end = FALSE)

next_quarter(ref = today(), only_start_end = FALSE)

last_year(ref = today(), only_start_end = FALSE)

this_year(ref = today(), only_start_end = FALSE)

next_year(ref = today(), only_start_end = FALSE)

start_of_last_week(ref = today(), day = 1)

end_of_last_week(ref = today(), day = 7)

start_of_this_week(ref = today(), day = 1)

end_of_this_week(ref = today(), day = 7)

start_of_last_month(ref = today())

end_of_last_month(ref = today())

start_of_this_month(ref = today())

end_of_this_month(ref = today())

start_of_next_month(ref = today())

end_of_next_month(ref = today())

start_of_last_quarter(ref = today())

end_of_last_quarter(ref = today())

start_of_this_quarter(ref = today())

end_of_this_quarter(ref = today())

start_of_next_quarter(ref = today())

end_of_next_quarter(ref = today())

start_of_last_year(ref = today())

end_of_last_year(ref = today())

start_of_this_year(ref = today())

end_of_this_year(ref = today())

start_of_next_year(ref = today())

end_of_next_year(ref = today())

week2date(wk, yr = year(Sys.Date()), day = 1)

nth_monday(ref = today(), n = 1)

nth_tuesday(ref = today(), n = 1)

nth_wednesday(ref = today(), n = 1)

nth_thursday(ref = today(), n = 1)

nth_friday(ref = today(), n = 1)

nth_saturday(ref = today(), n = 1)

nth_sunday(ref = today(), n = 1)

year(ref = today())

Arguments

ref
reference date (defaults to today)

only_start_end
logical to indicate whether only the first and last value of the resulting vector should be returned

day
day to return (0 are 7 are Sunday, 1 is Monday, etc.)

wk
week to search for

yr
year to search for, defaults to current year

n
relative number of weeks

Details

All functions return a vector of dates, except for yesterday(), today(), tomorrow(), week2date(), and the start_of_*(), end_of_*() and nth_*() functions; these return 1 date.

Week ranges always start on Mondays and end on Sundays.

year() always returns an integer.

Examples

today() %in% this_month()

#> [1] TRUE

next_week()

#> [1] "2021-10-25" "2021-10-26" "2021-10-27" "2021-10-28" "2021-10-29"
#> [6] "2021-10-30" "2021-10-31"

next_week(only_start_end = TRUE)

#> [1] "2021-10-25" "2021-10-31"

library(dplyr, warn.conflicts = FALSE)

# 2nd Monday of last month:
last_month() %>% nth_monday(2)

#> [1] "2021-09-13"

df <- data.frame(date = sample(seq.Date(start_of_last_year(),
                                        end_of_this_year(),
                                        by = "day"),
                               size = 500))
df$time <- as.POSIXct(paste(df$date, "12:00:00"))

# these are equal:
df %>%
  filter(date %>% between(start_of_last_week(),
                          end_of_last_week()))

#>         date                time
#> 1 2021-10-11 2021-10-11 12:00:00
#> 2 2021-10-13 2021-10-13 12:00:00
#> 3 2021-10-17 2021-10-17 12:00:00
#> 4 2021-10-15 2021-10-15 12:00:00
#> 5 2021-10-16 2021-10-16 12:00:00

df %>%
  filter(date %in% last_week())

#>         date                time
#> 1 2021-10-11 2021-10-11 12:00:00
#> 2 2021-10-13 2021-10-13 12:00:00
#> 3 2021-10-17 2021-10-17 12:00:00
#> 4 2021-10-15 2021-10-15 12:00:00
#> 5 2021-10-16 2021-10-16 12:00:00

# be sure to transform times to dates in certain filters
df %>%
  filter(as.Date(time) %>% between(start_of_last_week(),
                                   end_of_last_week()))

#>         date                time
#> 1 2021-10-11 2021-10-11 12:00:00
#> 2 2021-10-13 2021-10-13 12:00:00
#> 3 2021-10-17 2021-10-17 12:00:00
#> 4 2021-10-15 2021-10-15 12:00:00
#> 5 2021-10-16 2021-10-16 12:00:00

df %>%
  filter(as.Date(time) %in% last_week())

#>         date                time
#> 1 2021-10-11 2021-10-11 12:00:00
#> 2 2021-10-13 2021-10-13 12:00:00
#> 3 2021-10-17 2021-10-17 12:00:00
#> 4 2021-10-15 2021-10-15 12:00:00
#> 5 2021-10-16 2021-10-16 12:00:00

  
if (FALSE) {
data <- certedb_getmmb(dates = last_week(only_start_end = TRUE))
data <- certedb_getmmb(where = db$o.ontvangstdatum %in% last_week())
data <- certedb_getmmb(where = db$dlt.modbac_datumtijd > start_of_this_month())
}