Import Data and Plots

These functions can be used to import data. They work closely with the certeprojects package and support Trello card numbers.

import(
  filename,
  card_number = project_get_current_id(ask = FALSE),
  auto_transform = FALSE,
  ...
)

import.R(filename, card_number = project_get_current_id(ask = FALSE), ...)

import.excel(
  filename,
  card_number = project_get_current_id(ask = FALSE),
  sheet = 1,
  auto_transform = TRUE,
  datenames = "nl",
  dateformat = "yyyy-mm-dd",
  timeformat = "HH:MM",
  decimal.mark = ",",
  big.mark = "",
  timezone = "UTC",
  na = c("", "NULL", "NA", "<NA>"),
  ...
)

import.csv(
  filename,
  card_number = project_get_current_id(ask = FALSE),
  auto_transform = TRUE,
  datenames = "nl",
  dateformat = "yyyy-mm-dd",
  timeformat = "HH:MM",
  decimal.mark = ",",
  big.mark = "",
  timezone = "UTC",
  na = c("", "NULL", "NA", "<NA>"),
  ...
)

import.csv2(
  filename,
  card_number = project_get_current_id(ask = FALSE),
  auto_transform = TRUE,
  datenames = "nl",
  dateformat = "yyyy-mm-dd",
  timeformat = "HH:MM",
  decimal.mark = ",",
  big.mark = "",
  timezone = "UTC",
  na = c("", "NULL", "NA", "<NA>"),
  ...
)

import.tsv(
  filename,
  card_number = project_get_current_id(ask = FALSE),
  auto_transform = TRUE,
  datenames = "nl",
  dateformat = "yyyy-mm-dd",
  timeformat = "HH:MM",
  decimal.mark = ",",
  big.mark = "",
  timezone = "UTC",
  na = c("", "NULL", "NA", "<NA>"),
  ...
)

import.spss(
  filename,
  card_number = project_get_current_id(ask = FALSE),
  auto_transform = FALSE,
  datenames = "en",
  dateformat = "yyyy-mm-dd",
  timeformat = "HH:MM",
  decimal.mark = ".",
  big.mark = "",
  timezone = "UTC",
  na = c("", "NULL", "NA", "<NA>"),
  ...
)

import.clipboard(
  sep = "\t",
  header = TRUE,
  startrow = 1,
  auto_transform = TRUE,
  datenames = "nl",
  dateformat = "yyyy-mm-dd",
  timeformat = "HH:MM",
  decimal.mark = ",",
  big.mark = "",
  timezone = "UTC",
  na = c("", "NULL", "NA", "<NA>"),
  ...
)

Arguments

filename
the full path of the file to be imported, will be parsed to a character

card_number
a Trello card number

auto_transform
transform the imported data with auto_transform()

...
arguments passed on to methods

sheet
Excel sheet to import, defaults to first sheet

datenames
language of the date names, such as weekdays and months

dateformat
expected date format, will be coerced with cleaner::format_datetime()

timeformat
expected time format, will be coerced with cleaner::format_datetime()

decimal.mark
separator for decimal numbers

big.mark
separator for thousands

timezone
expected time zone

na
values to interpret as NA

sep
character to separate values in a row

header
use first row as header

startrow
first row to start importing

Details

The import() function uses the rio package to guess the required import function.

The import.clipboard() function requires the clipr package to be installed.

Examples

#