POST Data (Encrypted)

Securely transfer any R object with full attribute preservation and cross-platform JSON compatibility. This uses the same approach as the mmbi.epi package of the UMCG.

post_data(
  object,
  url,
  authorization_header = NULL,
  compress = TRUE,
  encrypt = TRUE,
  key = read_secret("tools.encryption_password")
)

create_json_body(
  object,
  compress = TRUE,
  encrypt = TRUE,
  key = read_secret("tools.encryption_password")
)

Arguments

object

Any object of any size, preferably a data set

url

A character string specifying the target URL for the HTTP POST request. Must include the full scheme (e.g., "https://" or "http://"), hostname, and path.

authorization_header

A character string specifying the value of the Authorization header to include in the POST request, e.g. "Bearer <token>". Use NULL to omit the header.

compress

Should the object be compressed/decompressed? At least allowed: "gzip" (or TRUE), "bzip2", "xz", see base::memCompress(). Use FALSE to not compress/decompress.

encrypt

Should the object be encrypted/decrypted? This applies AES-GCM via openssl::aes_gcm_encrypt(), providing authenticated encryption. This guarantees both confidentiality and integrity: the file cannot be read without the correct key, and any tampering will be detected automatically during decryption. The initialization vector (iv) will be a length-12 random raw vector.

key

A character to be used as the encryption key. Internally, this is converted using openssl::sha256() to ensure a raw high-entropy key of length 32, suitable for AES-GCM encryption. The default is the system environment variable: mmbi_epi_encryption_key.

Examples

if (FALSE) { # \dontrun{

post_data(iris,
          url = "https://some-server:8000/post",
          compress = TRUE,
          encrypt = TRUE)
} # }

# MANUAL WAY -----------------------------------------------------------

# use create_json_body() to make a(n encrypted) JSON of an object

iris_json <- iris |> create_json_body(compress = TRUE, encrypt = TRUE)
#> Warning: In read_secret(): environmental variable 'secrets_file' not set

# curl -X POST https://some-server:8000/post
#      -H "Content-Type: application/json"
#      -d '...'

# replace the "..." with the outcome of create_json_body():
iris_json
#> {"data":"CQv87WiCDgM2wE6yRW1DxBl+lQV5n0rzDurjEsJRccxu5/J0R3Ap1+oHoRmIgT3J56VrTHBG\nLm0EJBRn/7iEiT4Rg0FvbeJpqL6t9EvCJuT+fh8eygyxryuDbsurZlKR19LySG8q2NWDLLmm\nIuQqRz2arhmWK/r0+OOco0Cx8/kyYJxfO6+q2UMPkX51aiQW0OUpvYRhVeuikApnl9bSQgAB\ntxnTZ53in1VEJ/py9NWrNypjdbIOloMqjWkdwtoqrvap55tdxZQTuFB7UHxwGsxRjfatxZga\nUzU1lJ0ab3XXj//JQX7Gq90ww7ImJ4AaE9JDErDcCfZxqEQHWKwuv9nd315/JfKgFDz/7Oxr\nn4lkyeNviZ5Lkd4PETzHUgiYy2yTFEMm2IU6OA8/1u400cgEmV8H94oKooyZtQsrDeI4EdEX\nklSw6ydprLBDlZMXSDUFoq/bVk8AJgN8F+iVlmNQh6Jkk3bOPUt/HhwvSLIbUK9IkHixLQx/\nacK/ZFXNZNBwfS/RF8oqlBcJ+Vl3fZA+/3WlESSwql3KaMPmL7dkJbcFUeobVE2HqfmA12mb\nr2myJ4zn1z/oYimIUec5mbLq06B8YsBSULOrC1iXdyjTJzOUp6xg9M5KELzm4gPe9B4WuT5k\nhpVVCFtWA7YHhm6DJFeDsBOmP1onSlp6W7zbIrnwoO63L9IpM6E5STPCeTT5kWEWvyDLkCEy\nbRnPkzjrr50TA8FD3faal0e7uHiUw8ZwkMvoO1BI9UIMNz+OHn2NUe84fm3x3jzRNwT6NhTZ\nCjeY7qW0/zVGXdR8TpMxU+2X6op40H8aeJMjekTxTq5nyi9/HzxyjiLfPQFkAEwmzVLNtcbU\nG/jU96PGjXGBMpO/KPqZNRwXL/LsEGb+cF8OvTICUx4UUIw9yhGygauIoBtWV3IzH+INZ3LE\nufWxqNR8vmsr53M/WTC0TFaJYBwtWk1XBai3OMgN3bOFqRiTmzN600EJ97211XetMClmvqhI\n13uHHFtA/g9JrzGBg5Rdhj9qgqLp9jM1fMMNsPsUgaR0uN5ocuvBZUkNuApsGhStC0dztiNU\nfq27afgTKglO5lJg5DTGCig1S5e9pm2ZxJC5XjkrqfJfzGMO22fXcpMta19S9ChW6ahP1aE9\n9apCnv2B+vnlmKYkL/fD6vK1dEY9DtrB90SI2DQKosQrLLwJMKDDPNpbwPa/jBbpe6i1ucQp\nBxeikHJKtR7lyn4GYa8ThSzM0/xHKO6XRtJkTlpC1ugxFoZHPPnoh+gbFRivJosb3KKC1d76\nQmWEF5Nw7CiSOokQmEm9Q0icwA9eDX9oD/vB20Fn1XpfL6+wo7TjBA9+EtnDtcugkEVMUOpl\nWY6BecAQC4+URrCECZXmZSUQ+NnoAKiVWJFvEuobN6VxHEh/yMCS8siq43lhB6yx8NUPoQj8\nuqIAOxW2u+wmHc4d+eDhXIQCN5V7EZt9V9rtpJnfA1bIszW0NlEOis2y/uLKa+Q+T4GSA1/r\nvYYL629kblWZxHrxPqLXFO9qa8JVnc3TuxfxjRyqqg/8FqahaRdyfenc9/u0ArghWHybsHRp\n1LIi97l6kVbedORz4Pu5nCAmc0ZGf7MeLecw1hMO5h3FD38iAM4xO4v8/3uRXqzsYYW5AzSy\ndNu+hDdMEEZTgq1OTVEE0QlSNWAtq/DCLAhC1Pk1A63HZ+oOOpyJX5CW/ArOzVz+7plVcT7s\neNivvg2oXi4doIOFIAR/9KWMK+nNN/5YWH1J/igLihxHHwPufuCzr+hCxItgE1/I4NG2e1UL\nhFQOvr/gZ+Xs+CVn1RVOrUymvS5oap4n946GS+vBe3Lbx0FrAGG++/UNQtbp8qA1Tj3XyH28\nsukeb3J+93uqsym+6z7BoprHl0zuGkkiv2gN5rPv9HS1MngxOq7gMEBD2KbaLnn74mCR7bF5\nJLrOzfyl4H7q1Bq6DJSHyJH0xyQEcTnlz/oxoXpS583w","iv":"4d/EVJ1b161aMDby"}