Read Certe Secret From File

This function reads from a local or remote YAML file, as set in the environmental variable "secrets_file".

read_secret(property, file = Sys.getenv("secrets_file"))

Arguments

property
the property to read, case-sensitive

file
Either a path to a file, a connection, or literal data (either a single string or a raw vector).

Files ending in .gz, .bz2, .xz, or .zip will be automatically uncompressed. Files starting with http://, https://, ftp://, or ftps:// will be automatically downloaded. Remote gz files can also be automatically downloaded and decompressed.

Literal data is most useful for examples and tests. To be recognised as a path, it must be wrapped with I(), be a string containing at least one new line, or be a vector containing at least one string with a new line.

Using a value of clipboard() will read from the system clipboard.

Details

This function will return a vector of length > 1 if the property value contains comma's. The property name and value have to be separated with a colon (:).

The file will be read using readr::read_lines(), which allows almost any path or remote connection.

Examples

# for this example, create a temporary 'secrets' file
my_secrets_file <- tempfile(fileext = ".yaml")
Sys.setenv(secrets_file = my_secrets_file)
writeLines(c("tenant_id: 8fb3c03060e02e89",
             "default_users: user_1, user_2"),
           my_secrets_file)

read_secret("tenant_id")

#> [1] "8fb3c03060e02e89"

read_secret("default_users")

#> [1] "user_1" "user_2"