Format Table as Flextable

Format a data.frame as flextable() with Certe style, bold headers and Dutch number formats. This function can also transform existing flextable and gtsummary objects to allow the formatting provided in this tbl_flextable() function.

tbl_flextable(
  x,
  row.names = rownames(x),
  row.names.bold = TRUE,
  rows.italic = NULL,
  rows.bold = NULL,
  rows.height = NULL,
  rows.fill = NULL,
  rows.fill.picker = "certeblauw5",
  rows.zebra = FALSE,
  row.total = FALSE,
  row.total.name = "Totaal",
  row.total.function = sum,
  row.total.widths = NULL,
  row.total.bold = TRUE,
  row.extra.header = list(values = NULL, widths = 1),
  row.extra.footer = list(values = NULL, widths = 1),
  column.names = colnames(x),
  column.names.bold = TRUE,
  columns.width = NULL,
  columns.percent = NULL,
  columns.italic = NULL,
  columns.bold = NULL,
  columns.fill = NULL,
  columns.fill.picker = "certeblauw5",
  columns.zebra = FALSE,
  column.total = FALSE,
  column.total.name = "Totaal",
  column.total.function = sum,
  column.total.bold = TRUE,
  align = "c",
  align.part = "all",
  caption = "",
  na = "",
  logicals = c("X", ""),
  round.numbers = 2,
  round.percent = 1,
  format.dates = "d mmm yyyy",
  decimal.mark = dec_mark(),
  big.mark = big_mark(),
  font.family = "Calibri",
  font.size = 9,
  font.size.header = font.size + 1,
  values.colour = NULL,
  values.colour.picker = "certeroze",
  values.fill = NULL,
  values.fill.picker = "certeroze3",
  values.bold = NULL,
  values.italic = NULL,
  autofit = is.null(columns.width) & is.null(rows.height),
  autofit.fullpage = TRUE,
  autofit.fullpage.width = 16,
  vline = NULL,
  vline.border = fp_border_default("black"),
  vline.part = c("body", "footer"),
  print = FALSE,
  ...
)

Arguments

x

a data.frame or a flextable object or a gtsummary object

row.names

row names to be displayed. Will be 1:nrow(x) if set to TRUE, but can be a vector of values.

row.names.bold

display row names in bold

rows.italic

column indexes of rows in italics

rows.bold

column indexes of rows in bold

rows.height

height of the rows in centimetres

rows.fill

the column indices of rows to be shaded

rows.fill.picker

the colour for values in rows.fill, is evaluated with colourpicker()

rows.zebra

banded rows in the body - equivalent to rows.fill = seq(2, nrow(x), 2)

row.total

add a row total (at the bottom of the table)

row.total.name

name of the row total

row.total.function

function used to calculate all numeric values per column (non-numeric columns are skipped)

row.total.widths

cell width in row total

row.total.bold

bold formatting of row total

row.extra.header

an extra header to be displayed above the table

row.extra.footer

an extra footer to show below the table

column.names

column names to be displayed. Can also be a named vector where the names are existing columns, or indices of columns. When this vector is smaller than ncol(x), only the first length(column.names) are replaced. When this vector is longer than ncol(x), all column names are replaced

column.names.bold

display column names in bold

columns.width

width of columns. For autofit.fullpage = TRUE, these are proportions to autofit.fullpage.width. For autofit.fullpage = FALSE, these are centimeters

columns.percent

display the column indices as percentages using format2() - example: columns.percent = c(2, 3)

columns.italic

column indices of columns to be displayed in italics

columns.bold

column indices of columns in bold

columns.fill

the column indices of rows to be shaded

columns.fill.picker

the colour for values in columns.fill, is evaluated with colourpicker()

columns.zebra

banded columns - equivalent to columns.fill = seq(2, ncol(x), 2)

column.total

adding a column total (to the right of the table)

column.total.name

name of the column total

column.total.function

function used to calculate all numeric values per row

column.total.bold

bold formatting of column total

align

default is "c", which aligns everything centrally. Use "r", "l", "c" and "j"/"u" (justify/align) to change alignment. Can be a vector or a character (like "lrrrcc")

align.part

part of the table where the alignment should take place ("all", "header", "body", "footer")

caption

table caption

na

text for missing values

logicals

vector with two values that replace TRUE and FALSE

round.numbers

number of decimal places to round up for numbers

round.percent

number of decimal places to round to when using columns.percent

format.dates

see format2()

decimal.mark

decimal separator, defaults to dec_mark()

big.mark

thousands separator, defaults to big_mark()

font.family

table font family

font.size

table font size

font.size.header

font size of header

values.colour, values.fill, values.bold, values.italic

values to be formatted

values.colour.picker, values.fill.picker

the colour for values in values.colour or values.fill, is evaluated with colourpicker()

autofit

format table in width automatically

autofit.fullpage

display table across width of page

autofit.fullpage.width

set number of centimetres to width of table

vline

indices of columns to have a vertical line to their right

vline.border

function to define vertical line

vline.part

part of the table where the vertical lines should be placed ("all", "header", "body", "footer")

print

forced printing (required in a `for`` loop)

...

not used

Value

flextable object

Details

Run tbl_markdown() on a flextable object to transform it into markdown for use in Quarto or R Markdown reports. If print = TRUE in non-interactive sessions (Quarto or R Markdown), the flextable object will also be printed in markdown.

See also

Examples

# generate a data.frame
df <- data.frame(text = LETTERS[1:10],
                 `decimal numbers` = runif(10, 0, 10),
                 `whole numbers` = as.integer(runif(10, 0, 10)),
                 `logical values` = as.logical(round(runif(10, 0, 1))),
                 dates = today() - runif(10, 200, 2000),
                 stringsAsFactors = FALSE)

# default
tbl_flextable(df)      # dataset has no row names
#> a flextable object.
#> col_keys: `A`, `B`, `C`, `D`, `E` 
#> header has 1 row(s) 
#> body has 10 row(s) 
#> original dataset sample: 
#>   A    B C D           E
#> 1 A 5.31 9    2 nov 2020
#> 2 B 4.02 9    9 mei 2017
#> 3 C 2.70 3 X 27 jun 2018
#> 4 D 8.07 4 X  4 jul 2019
#> 5 E 0.80 8   15 jan 2021
tbl_flextable(mtcars)  # dataset has row names
#> a flextable object.
#> col_keys: `A`, `B`, `C`, `D`, `E`, `F`, `G`, `H`, `I`, `J`, `K`, `L` 
#> header has 1 row(s) 
#> body has 32 row(s) 
#> original dataset sample: 
#>                   A    B C     D   E    F    G     H I J K L
#> 1         Mazda RX4 21.0 6 160.0 110 3.90 2.62 16.46 0 1 4 4
#> 2     Mazda RX4 Wag 21.0 6 160.0 110 3.90 2.88 17.02 0 1 4 4
#> 3        Datsun 710 22.8 4 108.0  93 3.85 2.32 18.61 1 1 4 1
#> 4    Hornet 4 Drive 21.4 6 258.0 110 3.08 3.21 19.44 1 0 3 1
#> 5 Hornet Sportabout 18.7 8 360.0 175 3.15 3.44 17.02 0 0 3 2

# print in markdown
df |> 
  tbl_flextable() |> 
  tbl_markdown()
#> 
#> 
#> ```{=html}
#> <div class="tabwid"><style>.cl-4f0d0d14{}.cl-4f0518f2{font-family:'Calibri';font-size:10pt;font-weight:bold;font-style:normal;text-decoration:none;color:rgba(0, 0, 0, 1.00);background-color:transparent;}.cl-4f0518fc{font-family:'Calibri';font-size:9pt;font-weight:normal;font-style:normal;text-decoration:none;color:rgba(0, 0, 0, 1.00);background-color:transparent;}.cl-4f086548{margin:0;text-align:center;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);padding-bottom:5pt;padding-top:5pt;padding-left:5pt;padding-right:5pt;line-height: 1;background-color:transparent;}.cl-4f087f10{width:0.616in;background-color:transparent;vertical-align: middle;border-bottom: 2pt solid rgba(102, 102, 102, 1.00);border-top: 2pt solid rgba(102, 102, 102, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-4f087f24{width:1.673in;background-color:transparent;vertical-align: middle;border-bottom: 2pt solid rgba(102, 102, 102, 1.00);border-top: 2pt solid rgba(102, 102, 102, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-4f087f25{width:1.524in;background-color:transparent;vertical-align: middle;border-bottom: 2pt solid rgba(102, 102, 102, 1.00);border-top: 2pt solid rgba(102, 102, 102, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-4f087f26{width:1.391in;background-color:transparent;vertical-align: middle;border-bottom: 2pt solid rgba(102, 102, 102, 1.00);border-top: 2pt solid rgba(102, 102, 102, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-4f087f2e{width:1.096in;background-color:transparent;vertical-align: middle;border-bottom: 2pt solid rgba(102, 102, 102, 1.00);border-top: 2pt solid rgba(102, 102, 102, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-4f087f38{width:0.616in;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-4f087f39{width:1.673in;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-4f087f3a{width:1.524in;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-4f087f42{width:1.391in;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-4f087f43{width:1.096in;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-4f087f4c{width:0.616in;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-4f087f4d{width:1.673in;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-4f087f56{width:1.524in;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-4f087f57{width:1.391in;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-4f087f60{width:1.096in;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-4f087f61{width:0.616in;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-4f087f6a{width:1.673in;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-4f087f6b{width:1.524in;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-4f087f6c{width:1.391in;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-4f087f74{width:1.096in;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-4f087f75{width:0.616in;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-4f087f7e{width:1.673in;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-4f087f7f{width:1.524in;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-4f087f88{width:1.391in;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-4f087f89{width:1.096in;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-4f087f8a{width:0.616in;background-color:transparent;vertical-align: middle;border-bottom: 2pt solid rgba(102, 102, 102, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-4f087f92{width:1.673in;background-color:transparent;vertical-align: middle;border-bottom: 2pt solid rgba(102, 102, 102, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-4f087f93{width:1.524in;background-color:transparent;vertical-align: middle;border-bottom: 2pt solid rgba(102, 102, 102, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-4f087f9c{width:1.391in;background-color:transparent;vertical-align: middle;border-bottom: 2pt solid rgba(102, 102, 102, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-4f087f9d{width:1.096in;background-color:transparent;vertical-align: middle;border-bottom: 2pt solid rgba(102, 102, 102, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}</style><table class='cl-4f0d0d14'><thead><tr style="overflow-wrap:break-word;"><td class="cl-4f087f10"><p class="cl-4f086548"><span class="cl-4f0518f2">text</span></p></td><td class="cl-4f087f24"><p class="cl-4f086548"><span class="cl-4f0518f2">decimal.numbers</span></p></td><td class="cl-4f087f25"><p class="cl-4f086548"><span class="cl-4f0518f2">whole.numbers</span></p></td><td class="cl-4f087f26"><p class="cl-4f086548"><span class="cl-4f0518f2">logical.values</span></p></td><td class="cl-4f087f2e"><p class="cl-4f086548"><span class="cl-4f0518f2">dates</span></p></td></tr></thead><tbody><tr style="overflow-wrap:break-word;"><td class="cl-4f087f38"><p class="cl-4f086548"><span class="cl-4f0518fc">A</span></p></td><td class="cl-4f087f39"><p class="cl-4f086548"><span class="cl-4f0518fc">5.31</span></p></td><td class="cl-4f087f3a"><p class="cl-4f086548"><span class="cl-4f0518fc">9</span></p></td><td class="cl-4f087f42"><p class="cl-4f086548"><span class="cl-4f0518fc"></span></p></td><td class="cl-4f087f43"><p class="cl-4f086548"><span class="cl-4f0518fc">2 nov 2020</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-4f087f4c"><p class="cl-4f086548"><span class="cl-4f0518fc">B</span></p></td><td class="cl-4f087f4d"><p class="cl-4f086548"><span class="cl-4f0518fc">4.02</span></p></td><td class="cl-4f087f56"><p class="cl-4f086548"><span class="cl-4f0518fc">9</span></p></td><td class="cl-4f087f57"><p class="cl-4f086548"><span class="cl-4f0518fc"></span></p></td><td class="cl-4f087f60"><p class="cl-4f086548"><span class="cl-4f0518fc">9 mei 2017</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-4f087f61"><p class="cl-4f086548"><span class="cl-4f0518fc">C</span></p></td><td class="cl-4f087f6a"><p class="cl-4f086548"><span class="cl-4f0518fc">2.70</span></p></td><td class="cl-4f087f6b"><p class="cl-4f086548"><span class="cl-4f0518fc">3</span></p></td><td class="cl-4f087f6c"><p class="cl-4f086548"><span class="cl-4f0518fc">X</span></p></td><td class="cl-4f087f74"><p class="cl-4f086548"><span class="cl-4f0518fc">27 jun 2018</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-4f087f61"><p class="cl-4f086548"><span class="cl-4f0518fc">D</span></p></td><td class="cl-4f087f6a"><p class="cl-4f086548"><span class="cl-4f0518fc">8.07</span></p></td><td class="cl-4f087f6b"><p class="cl-4f086548"><span class="cl-4f0518fc">4</span></p></td><td class="cl-4f087f6c"><p class="cl-4f086548"><span class="cl-4f0518fc">X</span></p></td><td class="cl-4f087f74"><p class="cl-4f086548"><span class="cl-4f0518fc">4 jul 2019</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-4f087f61"><p class="cl-4f086548"><span class="cl-4f0518fc">E</span></p></td><td class="cl-4f087f6a"><p class="cl-4f086548"><span class="cl-4f0518fc">0.80</span></p></td><td class="cl-4f087f6b"><p class="cl-4f086548"><span class="cl-4f0518fc">8</span></p></td><td class="cl-4f087f6c"><p class="cl-4f086548"><span class="cl-4f0518fc"></span></p></td><td class="cl-4f087f74"><p class="cl-4f086548"><span class="cl-4f0518fc">15 jan 2021</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-4f087f75"><p class="cl-4f086548"><span class="cl-4f0518fc">F</span></p></td><td class="cl-4f087f7e"><p class="cl-4f086548"><span class="cl-4f0518fc">8.23</span></p></td><td class="cl-4f087f7f"><p class="cl-4f086548"><span class="cl-4f0518fc">3</span></p></td><td class="cl-4f087f88"><p class="cl-4f086548"><span class="cl-4f0518fc"></span></p></td><td class="cl-4f087f89"><p class="cl-4f086548"><span class="cl-4f0518fc">13 sep 2019</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-4f087f75"><p class="cl-4f086548"><span class="cl-4f0518fc">G</span></p></td><td class="cl-4f087f7e"><p class="cl-4f086548"><span class="cl-4f0518fc">4.58</span></p></td><td class="cl-4f087f7f"><p class="cl-4f086548"><span class="cl-4f0518fc">2</span></p></td><td class="cl-4f087f88"><p class="cl-4f086548"><span class="cl-4f0518fc"></span></p></td><td class="cl-4f087f89"><p class="cl-4f086548"><span class="cl-4f0518fc">21 sep 2021</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-4f087f75"><p class="cl-4f086548"><span class="cl-4f0518fc">H</span></p></td><td class="cl-4f087f7e"><p class="cl-4f086548"><span class="cl-4f0518fc">4.40</span></p></td><td class="cl-4f087f7f"><p class="cl-4f086548"><span class="cl-4f0518fc">7</span></p></td><td class="cl-4f087f88"><p class="cl-4f086548"><span class="cl-4f0518fc">X</span></p></td><td class="cl-4f087f89"><p class="cl-4f086548"><span class="cl-4f0518fc">9 aug 2021</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-4f087f75"><p class="cl-4f086548"><span class="cl-4f0518fc">I</span></p></td><td class="cl-4f087f7e"><p class="cl-4f086548"><span class="cl-4f0518fc">1.98</span></p></td><td class="cl-4f087f7f"><p class="cl-4f086548"><span class="cl-4f0518fc">7</span></p></td><td class="cl-4f087f88"><p class="cl-4f086548"><span class="cl-4f0518fc"></span></p></td><td class="cl-4f087f89"><p class="cl-4f086548"><span class="cl-4f0518fc">22 sep 2019</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-4f087f8a"><p class="cl-4f086548"><span class="cl-4f0518fc">J</span></p></td><td class="cl-4f087f92"><p class="cl-4f086548"><span class="cl-4f0518fc">4.32</span></p></td><td class="cl-4f087f93"><p class="cl-4f086548"><span class="cl-4f0518fc">1</span></p></td><td class="cl-4f087f9c"><p class="cl-4f086548"><span class="cl-4f0518fc"></span></p></td><td class="cl-4f087f9d"><p class="cl-4f086548"><span class="cl-4f0518fc">23 nov 2018</span></p></td></tr></tbody></table></div>
#> ```
#> 
#> 
#> 
#> 
#> 
  
# transform a gtsummary to a flextable
iris |>
  tbl_gtsummary(Species, add_p = TRUE) |>
  tbl_flextable()
#> a flextable object.
#> col_keys: `label`, `stat_1`, `stat_2`, `stat_3`, `p.value` 
#> header has 1 row(s) 
#> body has 4 row(s) 
#> original dataset sample: 
#>          label          stat_1          stat_2          stat_3 p.value
#> 1 Sepal.Length 5.0 (4.8 – 5.2) 5.9 (5.6 – 6.3) 6.5 (6.2 – 6.9)  <0.001
#> 2  Sepal.Width 3.4 (3.2 – 3.7) 2.8 (2.5 – 3.0) 3.0 (2.8 – 3.2)  <0.001
#> 3 Petal.Length 1.5 (1.4 – 1.6) 4.3 (4.0 – 4.6) 5.6 (5.1 – 5.9)  <0.001
#> 4  Petal.Width 0.2 (0.2 – 0.3) 1.3 (1.2 – 1.5) 2.0 (1.8 – 2.3)  <0.001
  
# extra formatting
tbl_flextable(df,
              logicals = c("X", "-"),     # replaces TRUE en FALSE
              values.colour = "X",
              values.fill = "X",
              row.names = "S. aureus",
              columns.italic = 1,
              format.dates = "ddd dd-mm-yy",
              round.numbers = 3)
#> a flextable object.
#> col_keys: `A`, `B`, `C`, `D`, `E`, `F` 
#> header has 1 row(s) 
#> body has 10 row(s) 
#> original dataset sample: 
#>           A B     C D E           F
#> 1 S. aureus A 5.312 9 - ma 02-11-20
#> 2 S. aureus B 4.024 9 - di 09-05-17
#> 3 S. aureus C 2.702 3 X wo 27-06-18
#> 4 S. aureus D 8.066 4 X do 04-07-19
#> 5 S. aureus E 0.797 8 - vr 15-01-21

# row totals
tbl_flextable(df,
              row.total = TRUE,           # add row total
              row.total.function = max,   # instead of sum()
              row.total.name = "Maximum", # also works with dates
              columns.percent = 2,        # 2nd column as percentages
              round.percent = 0)          # rounding percentages
#> a flextable object.
#> col_keys: `A`, `B`, `C`, `D`, `E` 
#> header has 1 row(s) 
#> body has 10 row(s) 
#> original dataset sample: 
#>   A    B C D           E
#> 1 A 531% 9    2 nov 2020
#> 2 B 402% 9    9 mei 2017
#> 3 C 270% 3 X 27 jun 2018
#> 4 D 807% 4 X  4 jul 2019
#> 5 E  80% 8   15 jan 2021

# column names
tbl_flextable(df,
              column.names = c("1" = "Column 1",
                               "2" = "Column 2",
                               dates = "DATES!"))
#> a flextable object.
#> col_keys: `A`, `B`, `C`, `D`, `E` 
#> header has 1 row(s) 
#> body has 10 row(s) 
#> original dataset sample: 
#>   A    B C D           E
#> 1 A 5.31 9    2 nov 2020
#> 2 B 4.02 9    9 mei 2017
#> 3 C 2.70 3 X 27 jun 2018
#> 4 D 8.07 4 X  4 jul 2019
#> 5 E 0.80 8   15 jan 2021
tbl_flextable(df,
              column.names = LETTERS)
#> Warning: Only the first 5 items of column.names will be used
#> a flextable object.
#> col_keys: `A`, `B`, `C`, `D`, `E` 
#> header has 1 row(s) 
#> body has 10 row(s) 
#> original dataset sample: 
#>   A    B C D           E
#> 1 A 5.31 9    2 nov 2020
#> 2 B 4.02 9    9 mei 2017
#> 3 C 2.70 3 X 27 jun 2018
#> 4 D 8.07 4 X  4 jul 2019
#> 5 E 0.80 8   15 jan 2021

# vertical lines, alignment and row names
tbl_flextable(df,
              align = "lrrcc", # also works: c("l", "r", "r", "c", "c")
              font.size = 12,
              vline = c(2, 4),
              vline.part = "all",
              row.names = paste("Experiment", 1:10))
#> a flextable object.
#> col_keys: `A`, `B`, `C`, `D`, `E`, `F` 
#> header has 1 row(s) 
#> body has 10 row(s) 
#> original dataset sample: 
#>              A B    C D E           F
#> 1 Experiment 1 A 5.31 9    2 nov 2020
#> 2 Experiment 2 B 4.02 9    9 mei 2017
#> 3 Experiment 3 C 2.70 3 X 27 jun 2018
#> 4 Experiment 4 D 8.07 4 X  4 jul 2019
#> 5 Experiment 5 E 0.80 8   15 jan 2021

# width of cells and table
tbl_flextable(data.frame(test1 = "A", test2 = "B"),
              vline = 1,
              autofit.fullpage.width = 16, # default values in cm
              columns.width = c(1, 3))     # ratio; cells become 4 and 12 cm
#> a flextable object.
#> col_keys: `A`, `B` 
#> header has 1 row(s) 
#> body has 1 row(s) 
#> original dataset sample: 
#>   A B
#> 1 A B

tbl_flextable(data.frame(test1 = "A", test2 = "B"),
              vline = 1,
              autofit.fullpage = FALSE,    # no fullpage autofit
              columns.width = c(1, 3))     # cells become 1 and 3 cm
#> a flextable object.
#> col_keys: `A`, `B` 
#> header has 1 row(s) 
#> body has 1 row(s) 
#> original dataset sample: 
#>   A B
#> 1 A B
              
# adding extra header or footer
tbl_flextable(data.frame(test1 = "A", test2 = "B"),
              row.extra.header = list(values = c("Header", "Header"),
                                      widths = c(1, 1)),
              row.extra.footer = list(values = c("Footer", "Footer"),
                                      widths = c(1, 1)))
#> a flextable object.
#> col_keys: `A`, `B` 
#> header has 2 row(s) 
#> body has 1 row(s) 
#> original dataset sample: 
#>   A B
#> 1 A B