Export Data and Plots

These functions can be used to export data and plots. They invisibly return the object itself again, allowing for usage in pipes. The functions work closely with the certeprojects package and support Trello card numbers.

export(
  object,
  fn,
  filename = NULL,
  card_number = project_get_current_id(ask = FALSE),
  ...
)

export.R(
  object,
  filename = NULL,
  card_number = project_get_current_id(ask = FALSE),
  export_qry = TRUE,
  ...
)

export.excel(
  object,
  filename = NULL,
  card_number = project_get_current_id(ask = FALSE),
  export_qry = TRUE,
  sheet_names = NULL,
  autofilter = TRUE,
  rows_zebra = TRUE,
  cols_zebra = FALSE,
  ...
)

export.xlsx(
  object,
  filename = NULL,
  card_number = project_get_current_id(ask = FALSE),
  export_qry = TRUE,
  sheet_names = NULL,
  autofilter = TRUE,
  rows_zebra = TRUE,
  cols_zebra = FALSE,
  ...
)

export.csv(
  object,
  filename = NULL,
  card_number = project_get_current_id(ask = FALSE),
  export_qry = TRUE,
  na = "",
  ...
)

export.csv2(
  object,
  filename = NULL,
  card_number = project_get_current_id(ask = FALSE),
  export_qry = TRUE,
  na = "",
  ...
)

export.tsv(
  object,
  filename = NULL,
  card_number = project_get_current_id(ask = FALSE),
  export_qry = TRUE,
  na = "",
  ...
)

export.spss(
  object,
  filename = NULL,
  card_number = project_get_current_id(ask = FALSE),
  export_qry = TRUE,
  ...
)

export.pdf(
  plot,
  filename = NULL,
  card_number = project_get_current_id(ask = FALSE),
  size = "A5",
  portrait = FALSE,
  ...
)

export.png(
  plot,
  filename = NULL,
  card_number = project_get_current_id(ask = FALSE),
  width = 1000,
  height = 800,
  text.factor = 1.2,
  ...
)

export.clipboard(
  object,
  sep = "\t",
  na = "",
  header = TRUE,
  quote = FALSE,
  decimal.mark = ",",
  ...
)

Arguments

object, plot

the R object to export

fn

a manual export function, such as haven::write_sas to write SAS files

filename

the full path of the exported file

card_number

a Trello card number

...

arguments passed on to methods

export_qry

export the underlying SQL query as well

sheet_names

sheet names

autofilter

create autofilter on columns in first row

rows_zebra

create banded rows

cols_zebra

create banded columns

na

replacement character for empty values (default: "")

size

paper size, defaults to A5. Can be A0 to A7.

portrait

portrait mode, defaults to FALSE (i.e., landscape mode)

width

required width of the PNG file in pixels

height

required height of the PNG file in pixels

text.factor

text factor for the exported plot. Defaults to 1.2, which loosely equals a PDF file in A5 format when it comes to text sizes.

sep

(for export.clipboard()) separator for values in a row (default: tab)

header

(for export.clipboard()) use column names as header (default: TRUE)

quote

(for export.clipboard()) use quotation marks (default: FALSE)

decimal.mark

(for export.clipboard()) character to use for decimal numbers

Details

The export() function can export to any file format, as long as the export function is passed on to the fn argument. This function has to have the object as first argument and the future file location as second argument.

RDS files as created by export.R() are compatible with R3 and R4.

The export.excel() uses as_excel() internally.

For export.csv(), export.csv2() and export.tsv(), files will be saved in UTF-8 encoding and values NA will be exported as "" at default. Like other *.csv and *.csv2 functions, csv is comma (,) separated and csv2 is semicolon (;) separated.

Exporting as SPSS files with export.spss() requires the haven package to be installed.

If the filename is left blank in export.pdf() or export.png(), the title of plot will be used if available (and a timestamp otherwise).

The export.clipboard() function requires the clipr package to be installed. The function allows any object (also other than data.frames) and is only limited to the available amount of RAM memory.

Examples

if (FALSE) {
library(dplyr)

# export two files: 'whole_file.rds' and 'first_ten_rows.xlsx'
starwars %>%
  export.R("whole_file") %>%
  slice(1:10) %>%
  export.excel("first_ten_rows")
}