POST Data (Encrypted)

Securely transfer any R object with full attribute preservation and cross-platform JSON compatibility. This uses the same approach as the mmbi.epi package of the UMCG.

post_data(
  object,
  url,
  authorization_header = NULL,
  compress = TRUE,
  encrypt = TRUE,
  key = read_secret("tools.encryption_password")
)

create_json_body(
  object,
  compress = TRUE,
  encrypt = TRUE,
  key = read_secret("tools.encryption_password")
)

Arguments

object

Any object of any size, preferably a data set

url

A character string specifying the target URL for the HTTP POST request. Must include the full scheme (e.g., "https://" or "http://"), hostname, and path.

authorization_header

A character string specifying the value of the Authorization header to include in the POST request, e.g. "Bearer <token>". Use NULL to omit the header.

compress

Should the object be compressed/decompressed? At least allowed: "gzip" (or TRUE), "bzip2", "xz", see base::memCompress(). Use FALSE to not compress/decompress.

encrypt

Should the object be encrypted/decrypted? This applies AES-GCM via openssl::aes_gcm_encrypt(), providing authenticated encryption. This guarantees both confidentiality and integrity: the file cannot be read without the correct key, and any tampering will be detected automatically during decryption. The initialization vector (iv) will be a length-12 random raw vector.

key

A character to be used as the encryption key. Internally, this is converted using openssl::sha256() to ensure a raw high-entropy key of length 32, suitable for AES-GCM encryption. The default is the system environment variable: mmbi_epi_encryption_key.

Examples

if (FALSE) { # \dontrun{

post_data(iris,
          url = "https://some-server:8000/post",
          compress = TRUE,
          encrypt = TRUE)
} # }

# MANUAL WAY -----------------------------------------------------------

# use create_json_body() to make a(n encrypted) JSON of an object

iris_json <- iris |> create_json_body(compress = TRUE, encrypt = TRUE)
#> Warning: In read_secret(): environmental variable 'secrets_file' not set

# curl -X POST https://some-server:8000/post
#      -H "Content-Type: application/json"
#      -d '...'

# replace the "..." with the outcome of create_json_body():
iris_json
#> {"data":"zU8/3MBOf/NRF/knjkDv2m24LH9TBcDn+FCnCTyLLNQRnd4TKq1auN/cMbmlW8KsBmRikZk5\n9yWpD1iG7AMNWimH64gJVWW/2FACst4bdXnn+kf3HHz3vuM5dNPafDwbKTNQGdz7AgZAnpyr\n3Hcg+D1/3vZGyjOSA8bD716/aS3cdT4yASfe6vBWLBM2zOkGwStGJFa6OVGeiaTxoldmW4LH\npf3QQCV3KBv5FDNQguvH2rYv0U5YBLaJbIuF2tsXOR2U3T3WsFEYGQ489j1G63CY3HX8biPK\n6iqpyO2b1Dvrk8qo3t90DSnIny+5HKA/yTxrtg94vczGs6iyKhH3cMXPk/jk0HNGz0rtDFMy\nlv7R9pi0aVF/ZqBPNaVWe+mZAJeWHhYI3Phdg1e9dP7rEkF3tKsRJYgUkkya3WeLj7X92ZvH\nH6xYt6YmFedhd3sEWbloSv/GkH/2znBNTPbFcw/uwJ7c0GEmr8HzOcnQg5pUUP2B0PJUUgdd\nTtP4w74upwglwiXj2+5Jl9+0CD2ubFHARizzF8S/w7YAchRNbU+VdFWCgOGH+wepfCPHilTV\nL2S8Ygio4D7xNngJdnkvrESVUJQCazImiyXC7SIaCcMabF4jQcr2XsBZI9UhJwsAOjph4L87\n4r/I+w94xo9cltpeNcoCzHOmbZYySjWKWaO0NUyfflkIalEfEMPsp6+v7H4q7FkEDMj3JDqn\nr1OKel6OI84yUoMEn/hTg6jwA3rrhfC0IRB4wOSp8CsbNhkqf5zqNAcH3Upv/csH5cVxBpYC\nn0DefDXm8KxEJgpwCMiTAI0MlwaBGqhASKc6O7om9wXJIAyRlJO7VkAHTYRDOM70Emnvm9CD\n2+nvTE1Af/MnfVi5lyid427f+IYtpcDlKixhOvHc/572SGSoKwfT6K4w4T1ge7hheXQH3p+L\nTknFxixBIWk25l5mtfCB/J0YXFi3eHhbgpSHBfHk3AAe2+gVbVmIEjLEg/fGemCgDTho6J9A\nR7EuWZUk2/Akqq82c/aqlrJHULDekesi8OcGXnwEAP4qVM6sPHDRflXdwx0YjbFcA220qnpV\nPdGeB0meQ3qg92H4BZqY3CIW4wO7N57SquykeUsVVbKAjpV+qtn6Besi6jtjozTkyITp5Yjg\nHG9o4aURN+AtKG/Anit7Rpg1hfnipkmxCJQF3lcFJa03yDHsxau3Hr+Vz96Ylb993ROaU2lo\ndZ7KOGGRs8UQYe7Iidn0Kvdb9P3z+Myfj7vu3h1uIhi2K/ZypuRRjywjTJKhzd9fQcFKKBRR\n1xS9Gl2PSaKQHva9fTNVJSQNPFQX7aD2Mwxuf14TdDtmAfVhJ3ZIZ4UKPF+dDfziNJZzmlUZ\n3LsTwD1qutz8hKrksQ1LJHJ46cs+oOMYou8igWqNGhkPLHeJpJ2NKe0BZRZ5NHzbkmwOUJl1\nb3GfzAhk3ZKQ6BOXGqmE31bB1DvG8OsIUcxhsJKLI8tnnQJJQC/I+wIWIQH6ri/b8AgJwMHT\n0IM+KWRNMh7Q6CH0vvY9vwBbi65CQ9urm8YzND3rKQSNZYD5mHL6G4Omn7fDZSP+h/pq81vo\nyvrJTwp/2md2gCacdlz8aay7vNl3GF4dtIdvYPlf2hYx5gSg3ex4bt/6mXYIMoEyrn738Uah\nwaZw1/oylbb20XjUh9qWaTQhBa1gnif/rhPbk32+R1UddLbgpb7mIWcgtD9ztuCnhUq54XId\n5PYb/WjvJXH8dXqZfPMbcq0KVFU2iLmlyz3a7yrp//2ov9S3qA5GhxV9gmmNeKH1m4DsjBec\nl9LYaH7bJxBihXZbfsjKmAomZxkBzDYEyLBLVE1SQvoI+Av2v8woRgdJ0TlbTIhj1Ajwa+Is\nIt8C1qzgVJjy1hx6kBK6hnAUAgVYz/OK1B6gWY9GPMagqYaXGxmGJ4gdIp5QMbFeiSKKyx0A\nsVWaW9i0x1T8BuU/60csoS3k5It3iESUjO6yrYzoZAAh","iv":"EZZvMuI621306L9E"}