POST Data (Encrypted)

Securely transfer any R object with full attribute preservation and cross-platform JSON compatibility. This uses the same approach as the mmbi.epi package of the UMCG.

post_data(
  object,
  url,
  authorization_header = NULL,
  compress = TRUE,
  encrypt = TRUE,
  key = read_secret("tools.encryption_password")
)

create_json_body(
  object,
  compress = TRUE,
  encrypt = TRUE,
  key = read_secret("tools.encryption_password")
)

Arguments

object

Any object of any size, preferably a data set

url

A character string specifying the target URL for the HTTP POST request. Must include the full scheme (e.g., "https://" or "http://"), hostname, and path.

authorization_header

A character string specifying the value of the Authorization header to include in the POST request, e.g. "Bearer <token>". Use NULL to omit the header.

compress

Should the object be compressed/decompressed? At least allowed: "gzip" (or TRUE), "bzip2", "xz", see base::memCompress(). Use FALSE to not compress/decompress.

encrypt

Should the object be encrypted/decrypted? This applies AES-GCM via openssl::aes_gcm_encrypt(), providing authenticated encryption. This guarantees both confidentiality and integrity: the file cannot be read without the correct key, and any tampering will be detected automatically during decryption. The initialization vector (iv) will be a length-12 random raw vector.

key

A character to be used as the encryption key. Internally, this is converted using openssl::sha256() to ensure a raw high-entropy key of length 32, suitable for AES-GCM encryption. The default is the system environment variable: mmbi_epi_encryption_key.

Examples

if (FALSE) { # \dontrun{

post_data(iris,
          url = "https://some-server:8000/post",
          compress = TRUE,
          encrypt = TRUE)
} # }

# MANUAL WAY -----------------------------------------------------------

# use create_json_body() to make a(n encrypted) JSON of an object

iris_json <- iris |> create_json_body(compress = TRUE, encrypt = TRUE)
#> Warning: In read_secret(): environmental variable 'secrets_file' not set

# curl -X POST https://some-server:8000/post
#      -H "Content-Type: application/json"
#      -d '...'

# replace the "..." with the outcome of create_json_body():
iris_json
#> {"data":"v69fe0Dg58MB0k4OXI6nA8P1e4+4UsfJ5JqF6qSHv+wCUx6qSKIe186lh3kKNX/vGbPx8fWv\nKm8SGUfpmp8b/wkAidiCr27AwnY8xnYosUxjOjuoaqPChOfYGhQmUdN6eOh3EpSP/eQf2sT8\ni7sw47i1VncleAomRJ1AFnhfLey+iZZo/uwEmRIofCRmvvaIBXVIXugjcf/60N3pj3e9Ojhm\n2H6ueRnsaVWiPpYCm09gDzeTm845eP5IDsSQTn8ncCGXEalfF9IgRKk+Z1QGZjE0o4k/em3Z\n1C4ouSe32X/qOFnbIuRxaBcEbbOOxueZHhZmsvbtIH5D+HDtM7/AuWVl8DoLeFPkBtcFWTNa\noMuj3srS+/RPibS4RO/+Q54SUwH/xYapxkRJEkETCyDbaQwOHMke0nxmkYJ0XoySkUQeUDpd\nrka3DKWPK8kAW7F/2fbPZOr00EydFKFS1pzZA6wrReLk3up4MT5BgL/o+D2k6aBod3M+aF38\nxU2tVaEnM0GWfURv1Z/crMi/wY91ZzhEpjl0qeWW5sWsqk6iPBpvFbZc63Lo54vt3EOxPpXf\n1XDoSlVp4XmwTpWBuQ4FGzx7oaYm5jRLPaUIswXyWR8OrtCfylhTnRLiYK5wgj8sdIQz5KkA\npPzJlDJi8GyVbezXK30X34uWIl65+xeLnNY/iN6u/zxqcOSBNK89/5el27ucheDrnBWc3W1I\n7QtK6spSzL5c5JRbGn6QO3rl1xIw7DcV0Tbs09xF9KkKoiPR0qjIaQJpGWDUYkkBtOluNzO/\nxizGjthg5oYev10khwssCOlGePySCIk39V7WpYIShKMIfy6e/CN3pg6VT375k6yRbTPwo0j5\nO595BniiXepnrcLy9NP3GIzNWjIVW1LlVwKmuRG53LWu+rUhejFqlaOlGCY4F0yfbnAGj8bn\nkxa2vnQ1hoWfqscG6o6+2F2aIMThzkFUYwSLxkvq3V60DW1ksw7Ph/27Vtb6EqW33ZMCVt1P\njrsZ+7Bqg39PDTXD0b4rTRR3XAnM2b66f3cJCbJyIa5pcNemsXu/DP+zaKxkIbUJqoR95sEO\n+cuyijCSWQ+RToMmQpPh8Y9gt1RRYsgw0MhVLGuxqzh6XyMxL/4fJKR+iPqm14FUjFOQZ/B1\nA+L3G3c+HLvf/un9WvC7UNTOW2yJgc7npcXLHvh1/JaekHwFOvPJvpvyJi6hPjB0goqyyydE\nyskT2DHFGb/H+NL/nP2U/7FhoACMHPgBTt0E7OSyluAQnJ8/0YHV+zlYJhuwrcgmcM6hYyM+\nPjWq718gsmudGNYxHeCciaZ899p2eAnu3LC2Qinbic1ukVH6iuXJXmat554XDHTzqHhDoBVF\n50IJ2BgR4CHGbE4WgCgOW6BxA0VUrALu2XbXqwWBZQ/pvdDAILBHYES/0qFN/uwbm9v93haK\nsK42kS9AgnYWvKj/nU8j2Rb30kHjyIQNU+tIlkrzkb0ewGb2y1UfTw5nK3AjQw2wfu3Q8xHR\nL/mqxXyK8mv5F1JXTyFmLCJ2aMDxESdnYxu5l+nlgbKIVUCzK09jcnTHYnLRIps7YUllV4yV\n9eEaTEa3f7o9A0wu/CpztM33vpgrcltyT/g1KVKi0U/5KJ2+cGmPEuP803Uy1vZBSXmnhn4S\n1HiGqOU28LIyL/t9jWzzzyewvQx6PPR8BZWVC0aZkoGVA4qITM3l6M8r0RJxOZM2F8M/XFDR\n6XeDUiZq7J9X9NAXsE3L82RGfZaoareG4NcY1FY+8K0M404NvzAbVndfCyRK5+QThwLBFjmW\nfkxAD10tebRurI5cHnxR2cUx8Ffyp7poppTtFYUpPjfTRaPWTPQ0wN2y5FVcc5ZiJ+wxkeww\nYx5cHCNujmudXPphIXwO8dUheFQDd8xJ2VwnGhqE0wop/8tPldz9TifI0bmI0xZniekP7Gzh\nS5/vXgZzUI9Sme+FPlKkBgL/Xj8s5eTgYp7idhBPlxHb","iv":"9IA2oi3K8ZK479bs"}