POST Data (Encrypted)

Securely transfer any R object with full attribute preservation and cross-platform JSON compatibility. This uses the same approach as the mmbi.epi package of the UMCG.

post_data(
  object,
  url,
  authorization_header = NULL,
  compress = TRUE,
  encrypt = TRUE,
  key = read_secret("tools.encryption_password")
)

create_json_body(
  object,
  compress = TRUE,
  encrypt = TRUE,
  key = read_secret("tools.encryption_password")
)

Arguments

object

Any object of any size, preferably a data set

url

A character string specifying the target URL for the HTTP POST request. Must include the full scheme (e.g., "https://" or "http://"), hostname, and path.

authorization_header

A character string specifying the value of the Authorization header to include in the POST request, e.g. "Bearer <token>". Use NULL to omit the header.

compress

Should the object be compressed/decompressed? At least allowed: "gzip" (or TRUE), "bzip2", "xz", see base::memCompress(). Use FALSE to not compress/decompress.

encrypt

Should the object be encrypted/decrypted? This applies AES-GCM via openssl::aes_gcm_encrypt(), providing authenticated encryption. This guarantees both confidentiality and integrity: the file cannot be read without the correct key, and any tampering will be detected automatically during decryption. The initialization vector (iv) will be a length-12 random raw vector.

key

A character to be used as the encryption key. Internally, this is converted using openssl::sha256() to ensure a raw high-entropy key of length 32, suitable for AES-GCM encryption. The default is the system environment variable: mmbi_epi_encryption_key.

Examples

if (FALSE) { # \dontrun{

post_data(iris,
          url = "https://some-server:8000/post",
          compress = TRUE,
          encrypt = TRUE)
} # }

# MANUAL WAY -----------------------------------------------------------

# use create_json_body() to make a(n encrypted) JSON of an object

iris_json <- iris |> create_json_body(compress = TRUE, encrypt = TRUE)
#> Warning: In read_secret(): environmental variable 'secrets_file' not set

# curl -X POST https://some-server:8000/post
#      -H "Content-Type: application/json"
#      -d '...'

# replace the "..." with the outcome of create_json_body():
iris_json
#> {"data":"5s84iMgiAQPSP9Qv5xv2hWIqhJiJKuTZlHDT2D0Uud24EJHYNMxrz90nwPqncEKRbkOZzQ+2\nCwRypSZB4z0KpDZ80Q1RTycBTGd6hnyW9o6F4qUWpx2+hvkQna++rTGACzbpnGG9QoiFXGoc\nuo4JAP53Ihw3H/s4tQcQ1NcHtaEp7GmO8qdovV915N5399wq3+nJuSRl7meEfIX8XzkQ6HEB\nR6F1I2O9SU2zCW6J5jy8e9U2+DeLbMOPzHMPCqpTz1u2GTTSudXmYok75hc75eYj/l5GBB2M\n2U5P2USBhTmUDBYcLWdJ3xI3ulilVekIMFcOX0WSZ8h/fUw9L+UFDcd4Wm7jhMV3nhmqeQgd\nwcRBF1lgoh+H0gqwya/0+qdUtlS/tgzlopCRAQSx80pR11y/umZEGo8CRpMdYOemAinYwz61\nHiazs/KUOvtidQTH79JFHpMc7ao7GKQX+AUxJraSAlQrG9ebirt/9JPESM6GjfXaYO0cbPPZ\nwSm1vRS0CJ2HtDPFUc54qNRnAovkH47oQq+0sIQnsiBhEZ79znEhLCk/MzHTylV8MlYuuKSF\n+sNAwCWD44BYSzWakRoH7OY0BZEQRrGUT5+oWWHE6KU42PljR4OAgbzfNzVZ6gyY/18RYMdP\nuJC8TJaFEGPBs9H6QhlVk82Sg8SfBZYkX6IKVdvCGpC0B8GmOuAfuALQZPUIAxG3dJfr5MXp\nG81sUVUjfwI1ixt+MsObt4PyuOn7MNAECZTiNMjE1q2fnUSus7seEwyquim3TOe9KfeCEkGQ\nZdHbJF8JsR2yM4FC+N/vVGIypj2yyJII0HnCdcuM3bCKqoY1XIA+OeQWflP+kz2Vs1OZbjTQ\nFw6jp5Tk+u52X+WgNaLCe1B4nsXjnivPi/3KpMH/RksBf9Fyt3uqFGoL+LXu9gNfB/09gjs6\nP/W4oSak+K11c8i63dNE9IoVW9zRsSWte7hSF7ZJNe+nEsWm7cDAXvy6HCeoAfemMxLAMiok\n9LIY3cUkDSBQ3e4aN41SrVhzzrS7idh2EdQzTPXB25B5ZAPGOGJXpAM3CZvIqIx6mhwC/sa/\nLGIhYP9EEZSFTwnmzRyQfc4jwVZ1xGH43rTO43il+2f8oMVcrqALEPrcsNsPm5G5ScK+tGz3\nlO6faNvB3EcXFEN+g1s6Pv/qh1WSgZm6Zpmaz3lTNIXlCeBtaHOX/ESV8pz3O+IV4oqHH2ki\nZ/L3shFGuhgq3bBhzmx5nTF9PSpTb2HH93R7mdKgPAkVNmgTeS9OAUoi+tstskrESYJ2lusz\nNboLFJXI7WRDgLaj/gb/bqu0vz73RY4viFo5WaAERqHg96L3TFN/XTgnfgsCnxdMUZhPMkkS\n7LFNxxTfrrV6ji1pSq0Xeqkj9Yc44ZngdlRplKPRjBm57TZebOn59DMw61u6DewMdZfRmUKV\nOHZYCGBpPtmXxQcGviGbojcfgO20xmQJrGodclRYjRgMCenCTQGEIS5XODBbrXUxgTp+4Qe8\nROALIhblAGcMTZvc0ieg9EHGzrii7MREzrhj73jqtYa0MRatMZlRIGwvBSjQkvdDKuJ0Jy4W\nBaqS5H3s5MvP23EIojRZTSeMSsyWhG/GiKp1jl9oB8iDBW1QWnA7k05wg54f08RMhVCxZ7yL\ngz405eh2Y+HnD8Ud5g9MBRbD8oFVA/D4VnDo+MJeN5A1b2+yTL9vQvfoUpetfSZlz3HVMBeN\nmZSwpUZyIQK3l4bUzqIA6MKDuVL9DM0i7JOyAx0t1QufEYiWPdcjVcR95ywVhFuVGEFaotFS\n9DXmHjGU/3wlX/TNEjVFh/q0HCMSYuGXEI8M0hLp/LpN/6NNrYlVSK87euoEaPzVR39p8QKq\nuyfFFn2m/PciI8EdrrgZt2Ocf9A3GMGmdLd32vQpZ5Rm2rMNip/h498gB07tRr9HKJgVH2rU\nAHcR5PK4kzEWLIVn/t8E14+KFZuEHVFzBDHr3Ek4k0qi","iv":"8kF0JW9bmqfIR0lu"}