POST Data (Encrypted)

Securely transfer any R object with full attribute preservation and cross-platform JSON compatibility. This uses the same approach as the mmbi.epi package of the UMCG.

post_data(
  object,
  url,
  authorization_header = NULL,
  compress = TRUE,
  encrypt = TRUE,
  key = read_secret("tools.encryption_password")
)

create_json_body(
  object,
  compress = TRUE,
  encrypt = TRUE,
  key = read_secret("tools.encryption_password")
)

Arguments

object

Any object of any size, preferably a data set

url

A character string specifying the target URL for the HTTP POST request. Must include the full scheme (e.g., "https://" or "http://"), hostname, and path.

authorization_header

A character string specifying the value of the Authorization header to include in the POST request, e.g. "Bearer <token>". Use NULL to omit the header.

compress

Should the object be compressed/decompressed? At least allowed: "gzip" (or TRUE), "bzip2", "xz", see base::memCompress(). Use FALSE to not compress/decompress.

encrypt

Should the object be encrypted/decrypted? This applies AES-GCM via openssl::aes_gcm_encrypt(), providing authenticated encryption. This guarantees both confidentiality and integrity: the file cannot be read without the correct key, and any tampering will be detected automatically during decryption. The initialization vector (iv) will be a length-12 random raw vector.

key

A character to be used as the encryption key. Internally, this is converted using openssl::sha256() to ensure a raw high-entropy key of length 32, suitable for AES-GCM encryption. The default is the system environment variable: mmbi_epi_encryption_key.

Examples

if (FALSE) { # \dontrun{

post_data(iris,
          url = "https://some-server:8000/post",
          compress = TRUE,
          encrypt = TRUE)
} # }

# MANUAL WAY -----------------------------------------------------------

# use create_json_body() to make a(n encrypted) JSON of an object

iris_json <- iris |> create_json_body(compress = TRUE, encrypt = TRUE)
#> Warning: In read_secret(): environmental variable 'secrets_file' not set

# curl -X POST https://some-server:8000/post
#      -H "Content-Type: application/json"
#      -d '...'

# replace the "..." with the outcome of create_json_body():
iris_json
#> {"data":"+kT7+1vII6mLYolxLZPM37g9FHbe9itSyR5dKMV7LGhocVMPOBaRTpDdfjksnvqAzFQy+eYy\n1qzMiFdJjcCy9y+Zo+HLq9jsB7/1nbBxxqaH66mOog2RpdZB+j821h1kutj3S9RETLiXXE77\nkv0DbSrUGodOlEZiXxR26R1uc1/2Zg/FW2tOYbcWYZk3xGzNDqXIVDYV8BHbqA8AZO3rMkSP\nJfRxyskC8zoxZR3APaDnP6V/1eAn67ncBBIisBQzIm3FSmCTVArrt5h8D4FghnLViRXEbRDM\nn9061bkCIoeUxG3YL6X2Ul4vO6OVR+A82KkWukZs0hRgQHRYc+Wbe8gcdKrNtYRsFKzFX9Th\nP266mbbPRvopPZsVoLHPz1V+jGJnaFRI1C4XEoU5cyu9AzpvQSROS+OvM83BUmGRGnVNxEsH\n224UFGMpBuOOaVUGhdqh2jDrF+0jpP0kkoRmYpPzTKq4hWRA7g6dRegMg6leyg4rXskGMppt\nsGbrRegwsGq28zWTJPFH8gp+tN2IvZSd4CxZGMZTFVgq/QevsgFdque+nCKfgk0cUSE2l3WI\nn/01xjldJVcYovbzWThFazwdOPqFEtDqE7hCXFWE7hn9Jn8wvxEurZt1KTf7XMcFg+VdaQmu\nZpWlg4BbYH2SHpg5y/L0iEDEv4xQTrsz/haoDyscEhY6BSZamZiZYAGVE0Kd/hNrhXuK5gWy\nyOXWzKansRXGhHFU9YJizwomQ0KJ6GeF7vhORx6HDN+XKtoK7s9RSLQ4k/m2jR13Xtsc0Cqt\n4FwjlbtXsQKyY2MuAew3uhNRUZlZGMLA9xmgGKHkmnWmySRfLMQM5Xa7orYGGJ2Ms5di+ZmZ\n5Ci5poRyf6rH5+LdMq8rGz6oQ/3CuWK5D/cV69SwwJw94Cg6eJ9XFIm3xQ+CYgUMFDEcmwsp\ncHA+X/NzyuJaFSnQWBeYJKXZ/e4TXq5mLd5TiJj8gIA8FYvnIVO1NdgXGp6DMcxWJCXsQtBY\noknhPGaOeV13NeLsBf9i3HtdHTMa/q6+DwkrcAbpn+UylL4lkw08/LFpygmVV1Df3iq5d/sm\nhp2c/N0A47Iph78UYE1Yq2wMbHWI07FfdQqq4cM8vy3XU9CbsRNiVJ/hxHU8meGcsalTMaNV\nbFt8R66wBLSzGj/bLrn5Z7dOaGHhQXQXkvWBMpGlcM0j7q9XWDUb+9ln1HzFplTPcwcjA0uK\n636OkAIuC28V/TDgYrxy820Aj/my9loF9cus8LX/LMAU0IgFtXtxASBRc+C+A5Ux4GVaPfl+\nmf0LpHd46JOAGRnEvZOupinYGAx8P0HCZdf/hMH1gE5sgEqBI/wtkxBCSwDa6o7yy9CMJ9uW\nYIgeL6iUzUeZfRkGm4swBmH0aN2aAa88tv3ykIDaIXnLi1lKvM9do0BuwUZupYt6GRXTN47+\npMX8DybWBiNd+fqPxWA2coIun5mQaHB4ZrCwYp8hesfgZzC/fNHcqejPUvmnCsHjcUcO/z6C\ng7//y9k4o9jBOG1vJGub+wAEPoz8WZ2z3PaHFl8ZtPa83jcEp6l75fmBS8mztYOP4Dm9Ixbc\nO00rHLCXjeBZPCHwvkJBGYa1eRWXnkZ5RjF5QQZSNIyIfUPE3eZvnHJ37hZqjj2gjRkmHYyB\n2jljoIq+qr9gm+kex8QbWVcG92/Qtxlj5Q6r5rnTF4AgsU0pJQfsjPrYIErVV1/+RPELdvY7\n+WVHUlMQz0nUsEh0qHyp4hnLwZtpfH5cZMbE2a2UJQsWsxkMbe/n7m48m1mApin4rhIMhids\nryfGgoRW+sZ2s5v5ndqda5wRF6bpnxEOszx9/O/e8n7q4xiyh2Rahso0N0ySD2cLjzlbsUsK\nQpUwzvtxPLTyR+bZx0fmTqXhYBcLvef09410Csm9j6g5XbdfD+YEKDD9ObaIPLeE/H5U9d4u\nYKGeV0apniN5wr6Hh7xydQWhCrVpg4IBWuvQJ7oeJc7k","iv":"W2kpL82mtl+qaQYW"}