POST Data (Encrypted)

Securely transfer any R object with full attribute preservation and cross-platform JSON compatibility. This uses the same approach as the mmbi.epi package of the UMCG.

post_data(
  object,
  url,
  authorization_header = NULL,
  compress = TRUE,
  encrypt = TRUE,
  key = read_secret("tools.encryption_password")
)

create_json_body(
  object,
  compress = TRUE,
  encrypt = TRUE,
  key = read_secret("tools.encryption_password")
)

Arguments

object

Any object of any size, preferably a data set

url

A character string specifying the target URL for the HTTP POST request. Must include the full scheme (e.g., "https://" or "http://"), hostname, and path.

authorization_header

A character string specifying the value of the Authorization header to include in the POST request, e.g. "Bearer <token>". Use NULL to omit the header.

compress

Should the object be compressed/decompressed? At least allowed: "gzip" (or TRUE), "bzip2", "xz", see base::memCompress(). Use FALSE to not compress/decompress.

encrypt

Should the object be encrypted/decrypted? This applies AES-GCM via openssl::aes_gcm_encrypt(), providing authenticated encryption. This guarantees both confidentiality and integrity: the file cannot be read without the correct key, and any tampering will be detected automatically during decryption. The initialization vector (iv) will be a length-12 random raw vector.

key

A character to be used as the encryption key. Internally, this is converted using openssl::sha256() to ensure a raw high-entropy key of length 32, suitable for AES-GCM encryption. The default is the system environment variable: mmbi_epi_encryption_key.

Examples

if (FALSE) { # \dontrun{

post_data(iris,
          url = "https://some-server:8000/post",
          compress = TRUE,
          encrypt = TRUE)
} # }

# MANUAL WAY -----------------------------------------------------------

# use create_json_body() to make a(n encrypted) JSON of an object

iris_json <- iris |> create_json_body(compress = TRUE, encrypt = TRUE)
#> Warning: In read_secret(): environmental variable 'secrets_file' not set

# curl -X POST https://some-server:8000/post
#      -H "Content-Type: application/json"
#      -d '...'

# replace the "..." with the outcome of create_json_body():
iris_json
#> {"data":"yMrN4mEWC6HjfaGMPnPpokEF7Yn3zbjgG38rl3gNTPwCx3OeBVangHzsvrAxyCJ0SIpiGpqe\ngkH10j/DVBLiy8fyQD8r/S07DzDL/6EmR0itQrx1rReAKO5V3DBT/iW8ph4GMNNTeEgRX7en\nGsQsXCYQLlnagQqgnn/X6qt8mbQtVc9Vuo2lJAvxauZBOF62Cr9SLwKVEz+Od7Yzw8yu0AvF\nthXL2vhDLwxn9AOAYikayN5USMLqxmrC6eqLLnnIiE4dIfGxBf2hd+3fEeiRwtlwjgycuTcE\nucpuhPgmXeCPH6fBBILllaK+OaRFSDpR2brj6VaQbHPvyUhAttq9vcnEt1YJBZ/yAfqPJxEV\n23Vg060oRwk+yq8jwcLSP7JveYlURiaZeWVERp69z5rsX3/kCpgs7OrcES1pTzFW210gSBoI\nrE/uyobe1YgTObQ7A9/4pTjgIcekuo3Wa12Gct0RWznlvzl2Y9kkc8L8trfHKIBgyxLIM8zy\nuhl8fC7Zp6DFMzpDs0Gfzynrfw8/vv7MuFMRndODfSeJznA6CCXM6a7IKD2+dBb0fyB8u3Z3\nl+nIQ2pwGztu4536zI+B2iWDoUXW4Wh2jZGrmK2XGtggpJQSu4DFQAz/37dzZBHb5x8J4eDM\nn1Ai9BuD3pXTmZK1KBsYpkhA4QhWdYkydR/UtCa6zPI68mbFeeJy2NBhhFoJL6UhRfroZ22p\nKnTjBrB5m98vYTcEM0fCDVAayUFbSVRd5ZaiOr8z0jO7lVOMDzy/E6YEE4K0v2oF/YyYhU/U\nbd1tIQDOoAOGDbvh6ygtVC+lrWVO29aymM4EWd+p2wUXkL2CqOi2mqtYzGnejL6D2Of3UA4m\nugiBVgIgpWSpAMEVJ3TKyToRLDsdjNtbeTzU64jQ2kQpAA7a/kY1U0o3JksCJCi+bberG26j\nzn1DbUKN14zYPx7d2nzPkVo6XidtOQUNrd7RgNh1PAn62d27EQ5XQihXEmaouYwis2H9hDZW\n0Vyr0tKqe5nJy+mXZCvjvwtuLShiCYWz1uKZNEgbH7Cqs6N87urlLZ/sJkZv3ueBw2huJ0p1\no94iA2vFUINADzYhuDA4wsKRwZQ0vFjSGJ+axxpI2vdapin017wH8z6TfOpOxjPPBg0bzP9i\ncDyGJPBn/UwCRX1YFydJHLAJsckWnrst5DQQDqeZZmine0QTKMc18Im3viBADnUJhZoSJKEs\n2gw0tY9GuxP0eCxpQ2YtCBGfB7LlfOR8jivQKeXWZS1icjSnb2KCLkctdTBxTGSkUwCNImkf\nW/Jl4qSq6WclS8SDK4VEXmm4UmJwPOTrcTuL+3oc7hLAdqYJ4J/etlqRqO8Knv9NF//Iie3u\n0ySQFAti+cVk2R8sntq9E7oakJZacjD+gWFPSwDq8JfIXgoBTeC5xS4+ciMaxYqYQ+Hq3CZ7\n8hcxvRxqH9JxlReVPwVnhSEBewC81D1SCQvnLwEQTHec0Po3JLyXCK/W/UTNs97RL0zFWjOn\nRS47+ObPx0iKsIXlv5Fnrs2NORSE1BPQUCICJaxkIcVrHsttxPV2qZmTR89J/+m0cZ4wHq5M\nekrnZDh5zsfitJYCEKw2V7OtbcdEPcI89OZxRM+nuQYaBeaLM6l0LiKQJaxWs8GWgMrtKEd5\ne6KLK4Z0utpgGxHfjhtR8Q48TlmJqmNcqKGHLXvSOoNi6uVqgozwKV9Im0kdyElQdzw03tSy\nFigYN0FPGWy8+frIhcrB2Xq+zBoKEPCKtwGUFPCZLuLccEoUjIqlGhWKwfdIrivDjdqLu5Sf\naZr7vq7eOytjLgYykABkg4aJ1OkDuO1GB2sDdBe9v1WZfHTQseddv6eW6e0Eqkj5X9LBfB68\nx2JzhXYf1CK0/V/KF4xkxmN60gScKYwiBzxKcaelJL/iP78GTMzEAotiPfqGPMivG1IN13Ya\nZVThFw7kr0euEA8Z+9/1Qm7AgAOepCo7qOsHxLQPSeze","iv":"0twXfcXTjQzaQbw8"}