POST Data (Encrypted)

Securely transfer any R object with full attribute preservation and cross-platform JSON compatibility. This uses the same approach as the mmbi.epi package of the UMCG.

post_data(
  object,
  url,
  authorization_header = NULL,
  compress = TRUE,
  encrypt = TRUE,
  key = read_secret("tools.encryption_password")
)

create_json_body(
  object,
  compress = TRUE,
  encrypt = TRUE,
  key = read_secret("tools.encryption_password")
)

Arguments

object

Any object of any size, preferably a data set

url

A character string specifying the target URL for the HTTP POST request. Must include the full scheme (e.g., "https://" or "http://"), hostname, and path.

authorization_header

A character string specifying the value of the Authorization header to include in the POST request, e.g. "Bearer <token>". Use NULL to omit the header.

compress

Should the object be compressed/decompressed? At least allowed: "gzip" (or TRUE), "bzip2", "xz", see base::memCompress(). Use FALSE to not compress/decompress.

encrypt

Should the object be encrypted/decrypted? This applies AES-GCM via openssl::aes_gcm_encrypt(), providing authenticated encryption. This guarantees both confidentiality and integrity: the file cannot be read without the correct key, and any tampering will be detected automatically during decryption. The initialization vector (iv) will be a length-12 random raw vector.

key

A character to be used as the encryption key. Internally, this is converted using openssl::sha256() to ensure a raw high-entropy key of length 32, suitable for AES-GCM encryption. The default is the system environment variable: mmbi_epi_encryption_key.

Examples

if (FALSE) { # \dontrun{

post_data(iris,
          url = "https://some-server:8000/post",
          compress = TRUE,
          encrypt = TRUE)
} # }

# MANUAL WAY -----------------------------------------------------------

# use create_json_body() to make a(n encrypted) JSON of an object

iris_json <- iris |> create_json_body(compress = TRUE, encrypt = TRUE)
#> Warning: In read_secret(): environmental variable 'secrets_file' not set

# curl -X POST https://some-server:8000/post
#      -H "Content-Type: application/json"
#      -d '...'

# replace the "..." with the outcome of create_json_body():
iris_json
#> {"data":"9IB7iYJxujlyl5TS03q6H1fqfL5UVi9uuB4mzCDwQTzfcp9EueuCoDu6SVJyBAhBvFY9V/Ef\nHGeakU1t2pEp6i+u5z+mV9sW5hSFnMw7LEo5AXub7s5afiCLrSKJo3dZTqVlt2WXe9ZRxBdS\nCyGoCfvIooXItSHLbN1Mzr14QLSXjfPAZ5Ntz4P4saSHBuuaMrksQ8Q2j/v5OXK91kHR3WX+\nOABC46sBxL9l4e4pj0+S54W66Y9h5dNISWKGdKf2CNM9+246O76vcWVtrdo1KPwhYEA/TsDF\n/v9QbymmjoEMJFDzG9JeuJD8EXapymiSRAvtFtSIiePzUktos7d1IkuDPZzfqysF9wIYTgAS\nQrNQ2I8NjOB+aUx9jfk9zru/Zc0UaP+8EATTNBpU9dpKr7dGcW46IprmzqlRYRYz5QrveK8/\nkJeUQ5oiDcy+gygeUaulHzhh841ftwndpNEik2Wr0mU5OwtQW4dzNU1dqaj4uEwIHQ2DzCqz\n9quAFYdQIfi9JyLuTYRZOvr8rBLQj8BLMKX7D3JVWTegsIyVKglcDEW8+TDHTlMZAa52PTUY\nynbX34W54ZKWcbZ5DwKOlUoKlYlbyoQ8iHhu/9IYCZbkMxehLCvFt7gxU5QPRcROfFxFU2xC\nmI0iSSm/+OwnMRS/KMc4hQWeK1Rv/dK9IbFnQdSChOYimb8YlfeePcNe4GnH5nxM6otBkNmj\nJPogX6u98s7lBTI5jzm0WlQ5y6W5fdTOCkQGpWWViHTE/Rs9HeJs0NPdYy/3jL+/1EjkgyDI\n5Bb6eMjxBYDLbWB+vY3w3+aNneGHFVbO5uNK9s9tGnxqQkgaKsLNoXmYtWcZTp32sLqgLAPC\ndO1+mVuV8tVDeCLbKpOB5D7rfMURtVpRn9BXlWMSAhU8cX3NOt3n2bWAlx1NRup/s6wTmhUf\n8SndANXNR/UqDd1UYLPLSzpQ6mGo0M4kRhIBiHYmt56my+FJ1sMp0eKOiy5lKhNYP+bGrcUo\nELMl4sxb5gzNLFPV7JM1Ed8GmcyNCXWP0YLhtKaog0xGJ+VQvBC4ThhID0d2V9hziHjdwS1D\nrjZHgUnwgnRnVLtu5wL9O/lcE/0u/uzJt9Fnr4oi1KVRJ89XgVJ7bOsvldB4WLZ5gWhuj6uI\npr2lO7Bp8SeZcJO2QMfDggZZH5NBIfwoBye69hAWR2aTlGEmDnDRIwuSsKCaUXd1lNxuk3zn\nOPnWkKZJ2rzXBfDq7aVuqBxhbb8DFn1KNMPBsi5UC1ArVRv0pqTwmei77Lkp/c3Hq3Py1OAm\ntEx+7wY5hFuRYAVPccHgSTnyk5NLnaMFDtrlhmSuBYdQlwpOu5KSC0iYHmrPf+FOXaR2xiWv\nsi8x/QG1I4ySwHp3Kk9c2lJ89UgjsVw+EHfMnAEV8652gBVPOTDj5wx8fqzhcfQmHPH1dxqT\n0Mt/F3RQo4sem/o8iSGAsbnxMPeUeS2XpS3NnMUL1ci315EJeX3w7JQdWdoMJBLmcl5rvAnj\n9EJTKJ2t3KvMiRrDJWcpTFYwitdWppDYM6qQRRAI39IRJDI0ZMLznNlWWRufiBqegMEarUgI\nqMXOIM4GXdVQRKfB8h9DBX82G7cfyOR0Al/DsGQh98ZH/HeGQhT3VJRn8hU5K5f2O1zLzeKN\nlEiDddSMyO4tXrpxTLPEKC98zvZcvMm0ehhSfOssbK6J/gJK88x3ZbSY1KI8hAaSDEDaFPdr\nq4/nZHO149x4CeZWkS7Drpq83ppFb3fgc/uxuWx12BLa/R8+FHk/hXwaHUxtUDJ79aCrV6vN\n7iNkjoHfWDxkodZoNvIPBQ7/t6BgSiJgLWtj0HYNHIoay15RA6yh2AnksiGL5w94EUUSZG/E\nJQTpIhEFOyk3JbcUJMKaq7R9gfaHwdrVdY5xrj0nPSMIsVXiodyjguZH9A2wdSqbdU2OvdY4\n5McyOtRctFR+ZasrI4BhDiKfFXtI+e2dvlv22u2w8ETS","iv":"DZUSalHY2K7YNBSE"}