POST Data (Encrypted)

Securely transfer any R object with full attribute preservation and cross-platform JSON compatibility. This uses the same approach as the mmbi.epi package of the UMCG.

post_data(
  object,
  url,
  authorization_header = NULL,
  compress = TRUE,
  encrypt = TRUE,
  key = read_secret("tools.encryption_password")
)

create_json_body(
  object,
  compress = TRUE,
  encrypt = TRUE,
  key = read_secret("tools.encryption_password")
)

Arguments

object

Any object of any size, preferably a data set

url

A character string specifying the target URL for the HTTP POST request. Must include the full scheme (e.g., "https://" or "http://"), hostname, and path.

authorization_header

A character string specifying the value of the Authorization header to include in the POST request, e.g. "Bearer <token>". Use NULL to omit the header.

compress

Should the object be compressed/decompressed? At least allowed: "gzip" (or TRUE), "bzip2", "xz", see base::memCompress(). Use FALSE to not compress/decompress.

encrypt

Should the object be encrypted/decrypted? This applies AES-GCM via openssl::aes_gcm_encrypt(), providing authenticated encryption. This guarantees both confidentiality and integrity: the file cannot be read without the correct key, and any tampering will be detected automatically during decryption. The initialization vector (iv) will be a length-12 random raw vector.

key

A character to be used as the encryption key. Internally, this is converted using openssl::sha256() to ensure a raw high-entropy key of length 32, suitable for AES-GCM encryption. The default is the system environment variable: mmbi_epi_encryption_key.

Examples

if (FALSE) { # \dontrun{

post_data(iris,
          url = "https://some-server:8000/post",
          compress = TRUE,
          encrypt = TRUE)
} # }

# MANUAL WAY -----------------------------------------------------------

# use create_json_body() to make a(n encrypted) JSON of an object

iris_json <- iris |> create_json_body(compress = TRUE, encrypt = TRUE)
#> Warning: In read_secret(): environmental variable 'secrets_file' not set

# curl -X POST https://some-server:8000/post
#      -H "Content-Type: application/json"
#      -d '...'

# replace the "..." with the outcome of create_json_body():
iris_json
#> {"data":"X9Vv7W1lw0oT42VRIzAmKgk9NkgfBip5HOdN6yYQjI353vUIQU0PDCLtPzvLXnsBkqEZhpaw\noMlAmY+jLbHqqeG3+V2FRD3mPyPY9WvziJEQszLsJotkNJgwic9FM9IOR+1LxHrK5jY925zY\nP14eR1TZENA8Ygl9YxOoIQRIH/LvXQY6d52MATK4rCKEbaI7ZAMmjFemOAsG3lXmONYaKsNc\nCR1wAweuxYt93pWUqlj88TLPcFpJCD4NCxLzWKBy0wmkdf1sUK7hyDH8pZD/PBRS6J+cTf+5\nUV3dqIw1ZYll72q7HwNMTbtkFAWAn9bsq7Hr9AUU2X/A1iEsMXU/Zqx2l8VYvWEQ+thXOPkk\nx4VIh6G+WT9prBFROUQh1X7aez+avFIeZ2JDQqmKrxPqdDCIP/L1pMh3T9N//I8MYkBi+RCj\n4E1YsQC/aePh09CAdgPLb4jYz6lGcwhqNhGUv0FEIm8IoLMK3nkEeo3/d9rUjhg7poeZg6ql\nys1XoBG3Jmqc/wsoW5J59lMsRcvov9S03a2+jwBY9PZOTRBsPQykckM/7j35pgOoQLqSnm4u\nDqOql3Wbzm0smwXuJRs2vqyKqJx79Xzyf2prWkOEZzdsBJzzXZCCU5EMWj3WlV9hxhfU3UaZ\nyPRjvXEU36tWqhql2tcmLd41Ux3/ZgPI4MC952jv6f3gzCf83aktcwArgGExismoQeEZsdqv\nsuAY/wd+HfVaOZvj7jlbywanI1PK0+4PBbaKVm5anNPizeQSUvFgXrku4D3WvQbHoGzQ0eip\ncjh2QRKCJAcKhJ7DJyk/37Xn+/Oh9ZVCzk0p3dbjhP1bepSlJgoD0FQS6Gpa+2h6ymMM+mqj\n3p9vn19G7NpFiTkidHGZ24eQANhSPt06NtdN2AjSMBwJcEYIcRiLe1OLBlqBJadMDKeaCBNf\nXZ39ndLAbzh+lPA6PdkoYYksIWJEFtAvb85QGnZ1fXscyq3shAOjoFFBtwKKFUM67rbgjU8U\n//ZFX1K3NMe1n8fsS1y7z3jMpCRAjcDDsw/qKNVq9j+Rg0zyXE3q0ykU3SHbvAEAJGOZNXXx\n8gkDoIIgWwIgZUcwEntlUfEtAXXXhAGq+u5SEcsZ6rxV1FTxsWkaFhlJk2ufx2xqcX7/1xj3\nE7JFvNBAZLWkSbgKBbTE8d71hOgKShG/ZpZ9rYxnWqzekqpLJI84+bHDEwsL188+M4Jm1v+U\n6fT2bpD/x1T328KgINWCwBDCBvHN+8OlK8QBoq/WfJdRCpLi6GjQPlyOgJZekKlmgmnQcAt9\nyDpTZi/uJSaXa3XDFTBv8PWCe7tEjAatuInpmjMdL5anfOab6GBGtMOQpgTCrlOKkrCEqx6G\nHZzftM+vDHt8TnfYJ3N2nGsgJFiU7tozlJTyMVW+X6p0qxfAvyGug+FIpWO8dyFX0dQMHk8m\n419AokuP8G+Lnc4wcvchaNvvHD0VoampFGqLyraHW+37bwvDuCC9YsP50h7g7BRKUqp1Mz2w\nxZD0RBFLj6tZS2OfknfM7cadj7xig/1usckKd5PjZE5K8zgq5Vt+kBLBKnQPwLdh6w24WFFY\nsvjrZvon7v0vpOZL4Cj+wwGxDccapI7XCLzC2zs5aQXPU4WfGQrd7xRdqF/pRS9FMRJpFxWk\nm53et2AZ20XdSa/96KmF+BMS6OPljJKfxdalAStiJWpkdGonAtmpse2dDC3Ss5TZLOWtjJbN\n7w+TkLnqmX1Vk4kZcek3FcTEdQ6IFNJahBw9Zu9PMfiGb/qcEWoZIfscQingbv1UYgPdIq8b\nC7PwowhbK7sRMVhvtIsqK9Im9YkrcrVLQlNA2GA7xqvRIL+8stZIyP7gMTsxul14KmNfPqPM\nK9DG2grMZBg2Koum98cgpYeZgBU06EaGdA+yuvzUUT5CkBoCPvB7CG1g5/XLvFSLlpt1uGDj\naB6fMiiHJXB88qtZOCXOSKpjegm4zQzBrXIv9KDVy/lT","iv":"1XfOvppxMVzVDD/C"}