POST Data (Encrypted)

Securely transfer any R object with full attribute preservation and cross-platform JSON compatibility. This uses the same approach as the mmbi.epi package of the UMCG.

post_data(
  object,
  url,
  authorization_header = NULL,
  compress = TRUE,
  encrypt = TRUE,
  key = read_secret("tools.encryption_password")
)

create_json_body(
  object,
  compress = TRUE,
  encrypt = TRUE,
  key = read_secret("tools.encryption_password")
)

Arguments

object

Any object of any size, preferably a data set

url

A character string specifying the target URL for the HTTP POST request. Must include the full scheme (e.g., "https://" or "http://"), hostname, and path.

authorization_header

A character string specifying the value of the Authorization header to include in the POST request, e.g. "Bearer <token>". Use NULL to omit the header.

compress

Should the object be compressed/decompressed? At least allowed: "gzip" (or TRUE), "bzip2", "xz", see base::memCompress(). Use FALSE to not compress/decompress.

encrypt

Should the object be encrypted/decrypted? This applies AES-GCM via openssl::aes_gcm_encrypt(), providing authenticated encryption. This guarantees both confidentiality and integrity: the file cannot be read without the correct key, and any tampering will be detected automatically during decryption. The initialization vector (iv) will be a length-12 random raw vector.

key

A character to be used as the encryption key. Internally, this is converted using openssl::sha256() to ensure a raw high-entropy key of length 32, suitable for AES-GCM encryption. The default is the system environment variable: mmbi_epi_encryption_key.

Examples

if (FALSE) { # \dontrun{

post_data(iris,
          url = "https://some-server:8000/post",
          compress = TRUE,
          encrypt = TRUE)
} # }

# MANUAL WAY -----------------------------------------------------------

# use create_json_body() to make a(n encrypted) JSON of an object

iris_json <- iris |> create_json_body(compress = TRUE, encrypt = TRUE)
#> Warning: In read_secret(): environmental variable 'secrets_file' not set

# curl -X POST https://some-server:8000/post
#      -H "Content-Type: application/json"
#      -d '...'

# replace the "..." with the outcome of create_json_body():
iris_json
#> {"data":"kuruCbkRvxBO0d/vdP02//GUTXI1VXKKprsvktigl7YN595v0UJ+6HRDu8fZW5v3lCvYEH6Y\ndt0R3ET76TpiQNF7Zw+Kl+SoPuM96ZQw1fzUhRFgUZHG3WfLz/TpuFvpMRHsSjiaRlz+O4Xj\nFaULH6+BlplFhgV3XSrsqgsviO8zXzly99krI5Z7p6mtNXVecabOastCdWXbCiGHfm09B3yd\nUTV4qk+qrITJVdKSvKdRfTyZtFqimhS5pH7haCg0YfmOn8ZYrLYtKRO7IeZtPLDEajQCt+kY\n2WGWvNpMPop1Al2FwOI2KcW7qzMF70lmG4I0RHj6edkayzWq00Ig75am046ZQqmbw+H+0N6L\nPVOe4ur+Vy2fNucp1iSKkeZsp+cY2bGR+XLWJigXYY5M1weC9A70iH5cpnqIScgngYQOAqV0\n0YfVM5f1o5tzY/MQUM+N4cJ5xjsEcBiLV2froXztro2/JbzjkhCvKtNMLRmY8fe6L5faWJsK\nGOxJ1zCBFS5CMl6xjnXu0dekMJquz+gHKK8ArwYb4eO0mjO7fKtcd8j05lpGl53lraCHBZjR\nhUMnpvF1twMR3a0HBJQZZXrW0FIDLxUyvA65iUmDlZmM9saz5yAk4tqD4HEqDRjQ/FWYhQRM\n97dt+BWM5eq1iX7Z80//hIwGPz1MruIiwqyDYRMNJAQ+iSV/VFt9ZykLwj1mU+RGMnb//P+r\nv8NluPtQ45nhcyiw166RryKeFACLsN0lvktBPclMTIpBbWmZCSPr/ERg9CftNcCAQmn7O1Yy\nZQ4uOVKUboTageBnCm3fwpao/txscoCIDbuPNCb+Q7awewZlDPkc8n8xoClSWPQypuHGUXe6\nIXy/FVoQ9gjJEH6X1tjugwOxJeXxF2BhI8BGuwA/3w5yl6rO9y1Xq11Ec70VbSKTYTJN+dFN\nZEKWFd6p+aTS4WNFdWKhsgy9nhOBot4hL2ye+KoZ3ZScyBHXVmTsxgFeSfIq1N3QnPBn2rHI\nJqjirDBw21XiKxxqEGccBclgYnavnhUxCUmm0noCeHdJTCxEInVlHy7Es9Hg/v9aX8P59dOX\nXfML4E36NdMXhgTAxHgqNsT23theVrqidEywPJ3jdVjlsorUHzj4UM8om0LvcvZUb/1gF0Kh\nLmiCNr/ompiROZ4Y2vGKv5ubrhaIHcGOzGydhQkc9804Q0IchgDvcDgM02CBHtrbz4ZLutjO\nmW78KZ0MjRo3VOIvSVXZhK4CDQTRPFmY0BgmgcrOOyZeyL80OpZLtTm30uI+FURd4lSHWxGw\nJjl6ykeuCFvmM4bvICfjFsereChs3I+kJj2WA2Az8IgCgpcrq4rQ15JZEfQGbh5r7QJe3VB/\nvhf9CBDeJjNXjdhQNBVfef5aSIbE6mhIcKndrZoGLMH8cTect/Vi3tcCUnX7HwrX3z5aAmXK\n+/HiBb94KjNoS+paHl5RqLKoRUZGihjJv+/0hdZqMS29KLy0BMprl8UyhD/TkeTxGyzEaO92\nM9RbDmlzonQ5nRgxGUpFxftm52v/nfa748TA1TiaGQwvC056+4RnlU1csjxlrnPx4o7DhKKi\nXH9eMF5Z3T410q1Tl/3yNoBNJVCFD+u+RgdVgpqdUEz3uphXy06QY/2tmF3Tn3T6PviZGu1H\nc6ipxMYJtDa3Ar17Cnm6L6QRaaq5g4uK/c3OpOVHIbzlBDo0rF9iKWSvWrmuPw7kglj96oBa\nxeluaAX/TDKAtIHYo8Aj6SKYoPI9GrB0X+DvwZPOVhQX6FPb5BsdAaSsbF9KFPoPMSFWxSh6\nmnlGUsponXEeYVvOS9Qihc132k7BMh3MGn4tQsMB4fcVVe7O8pZRsdsG8TQXc1GroKpjTpUj\nvbWsK0wiSvhFWoaKkDO/Dr942WnCQPaEENKEtXSnsaL2JMUXdXH6VQ3bMPKYNDtL4jhUFF2t\nbUGC92muXYOZuHKCDWnaARoZXY7I2R6jp/YPC088rvWV","iv":"aWj+lEQQBBp9fkm3"}