POST Data (Encrypted)

Securely transfer any R object with full attribute preservation and cross-platform JSON compatibility. This uses the same approach as the mmbi.epi package of the UMCG.

post_data(
  object,
  url,
  authorization_header = NULL,
  compress = TRUE,
  encrypt = TRUE,
  key = read_secret("tools.encryption_password")
)

create_json_body(
  object,
  compress = TRUE,
  encrypt = TRUE,
  key = read_secret("tools.encryption_password")
)

Arguments

object

Any object of any size, preferably a data set

url

A character string specifying the target URL for the HTTP POST request. Must include the full scheme (e.g., "https://" or "http://"), hostname, and path.

authorization_header

A character string specifying the value of the Authorization header to include in the POST request, e.g. "Bearer <token>". Use NULL to omit the header.

compress

Should the object be compressed/decompressed? At least allowed: "gzip" (or TRUE), "bzip2", "xz", see base::memCompress(). Use FALSE to not compress/decompress.

encrypt

Should the object be encrypted/decrypted? This applies AES-GCM via openssl::aes_gcm_encrypt(), providing authenticated encryption. This guarantees both confidentiality and integrity: the file cannot be read without the correct key, and any tampering will be detected automatically during decryption. The initialization vector (iv) will be a length-12 random raw vector.

key

A character to be used as the encryption key. Internally, this is converted using openssl::sha256() to ensure a raw high-entropy key of length 32, suitable for AES-GCM encryption. The default is the system environment variable: mmbi_epi_encryption_key.

Examples

if (FALSE) { # \dontrun{

post_data(iris,
          url = "https://some-server:8000/post",
          compress = TRUE,
          encrypt = TRUE)
} # }

# MANUAL WAY -----------------------------------------------------------

# use create_json_body() to make a(n encrypted) JSON of an object

iris_json <- iris |> create_json_body(compress = TRUE, encrypt = TRUE)
#> Warning: In read_secret(): environmental variable 'secrets_file' not set

# curl -X POST https://some-server:8000/post
#      -H "Content-Type: application/json"
#      -d '...'

# replace the "..." with the outcome of create_json_body():
iris_json
#> {"data":"MvXXXxUzO9kXcP81XzSXd1OnjJ0XkYTaY9pKhHohxrYo40Hf4GLgITwfLDigoPebE2Hmw921\n1JNhM+yk6z3s6VckNcl65xBN4dUlCdd81MnqDlPUq6MzrqtwLqSQLxlCQCHkR7D+4wc5CK/V\nX6yKUOrKdAeXa5pM5lwreKzZtNwT182Y4dcWf279f5cLkmb+RgVpawZlGK0ccbzMSlDAzKKW\ngx0pLcOdE0yi6WIBtU6kOMrqAGqjalIDwnBsu8IXBLUTzVYCev+p1oyC27/RDDQQpWUtJGxr\ne9fN7B5Rt0p4DQeSNAIJMVPKLMcJULNIG4RMszYj6wR2vlj5roEH0RyuM6BxXVR8bcYhf59B\n8r8G9oFzgCM6PiD0NY4Jzloan/u+AOqCCyRFEzueoEJ1psFnDZ1Og/0bAYirLeG1QAa7JIaY\nEgl7Bo6M8VUIKtJrP68336IomFBL/Vwl3dtbSVVpChl9fvoDNQUCL7fQSJ461XEHP3XaS0no\ni3jGbnTyvDmNkSUVxva35cV4GxEuXd1y3QlyNx+y7TIxYtqFOQ3kqZ8efLetlpQJfsEQsMoH\nEDYuD4CWNdRRzVaRJEyXUaDKxHf3siEODLS4DPcvC9tw++9vB8SXJsAOs4WoBs9ynv9Zheyv\nUY708Qj94SPGBJULL/apjD7MQt1C4moMBp0vSEhjHz2DM8Cf977EuYDvkk69h4mMlOg8xcdB\nQ9L9zrBwcrV/TVDfpFXTQ5TUWrOmpMgI0slfAm2jbZxdAvulnDbgGdSyi7e8URalLtTYLrBT\n2e63fKKxySEAGkyqZ6rClDuvs5d4qskMzMTbP2fSfEUwR9qz45b4ixdDCBIEPpL7k9ExKzKK\nViSc8NTmkCQarhf9yKHFe29hi/g9H2DCBm9yyafEVQ09mmSVuubWPH6MmKEGCG83gNNB/pr0\nkDM3LR+43srqbiWi3wHKrGlHOisCX+6u/cvphyB0IqoM9mXxaWyD7NiEVvq26faFB45rx4So\n9RcnyDHdLTsB0baVYEgUJDup1LODQtEk7IR4PpTJKBPIdZBAZLtGpSRafy1ejT6sfIo9/yJ8\n77he5kqsfuIhyGCBS2vKWeHpWO0nD6U/o0aATO+XpltzOq2tOoj44zh6Dbh9c+SwzyJT611b\nglW1MC7VUu7IReSDpztELLofQaElO6aMKatXn7O8TstGclybWy5j6krfH0PfpTte8+eia0Ac\njJM6NgP6hGHdcf1fBQm0TiuBSOnuqi7EXZkn0f37njjBsAUfU6rGH5EdtU4lGgTi3M+BRhOY\nSHdiTHmw2g0TQdwcPeDOYk3GMiAPBb1fkNn5o/cA7cmi12WtfuY3/3Z1wkr3eNFDkrjmi3vV\nbJJZeG/P0SlZD+ZjRiuwOd5s9jYdOqD3ELcX4ZqZQ7J/KTPCHT+1VZ018H1FGNpmTPrL0MtP\nvHlvWGdA5Y2+KMbH0X9WRwuzs4TbgcFg69damF2vNuRtVQNLbw26Bp/IQDvtOAW4hI4Y214g\n35U7alVqKql5qP9SUvJSeF6j8sPKdInNAPz97wJeOAJgewOrnr/c2cjD64N1bUhxedqv5I0/\njkHQeQEEZ0uwWn1r40D6MEPrvrjOwPZn6by/DDwOiHa44dxqFiiw+YAzMp3MuKcQ9PsFl+Wv\nhhQtKQHZzpjM1EE2F+zYI8bZXlVVD9fDJb2UAdE5EcNH7MVNDQVuQpmjQW4GjFOYNZUCZrx0\n/tOAJ7QQ+iQWZBEJYxkBihTOMCWUkyZOE2uR+Kk2lYMG24+Hdrf4UXXMAqhqXme5SLbQ7fVw\n4C6r6eCwVVXWSBL2JGNC5Z6RnFqCny/dznlXJXG8u3SRA6xIeUBkMfyn1YqRJB2rj2feAUD6\nAZm4uC8czq2KrvWobz1efY4FjeGHdgBWP9/TlbAfug/H9iX/pRuQNpkghWExo9QGsXKpoEJy\nFQaVV+7Iety0obb2zIulpq7bbgNQv8tpnHd2vqlsNNo/","iv":"UUrcw0I64H+WFQ8S"}