POST Data (Encrypted)

Securely transfer any R object with full attribute preservation and cross-platform JSON compatibility. This uses the same approach as the mmbi.epi package of the UMCG.

post_data(
  object,
  url,
  authorization_header = NULL,
  compress = TRUE,
  encrypt = TRUE,
  key = read_secret("tools.encryption_password")
)

create_json_body(
  object,
  compress = TRUE,
  encrypt = TRUE,
  key = read_secret("tools.encryption_password")
)

Arguments

object

Any object of any size, preferably a data set

url

A character string specifying the target URL for the HTTP POST request. Must include the full scheme (e.g., "https://" or "http://"), hostname, and path.

authorization_header

A character string specifying the value of the Authorization header to include in the POST request, e.g. "Bearer <token>". Use NULL to omit the header.

compress

Should the object be compressed/decompressed? At least allowed: "gzip" (or TRUE), "bzip2", "xz", see base::memCompress(). Use FALSE to not compress/decompress.

encrypt

Should the object be encrypted/decrypted? This applies AES-GCM via openssl::aes_gcm_encrypt(), providing authenticated encryption. This guarantees both confidentiality and integrity: the file cannot be read without the correct key, and any tampering will be detected automatically during decryption. The initialization vector (iv) will be a length-12 random raw vector.

key

A character to be used as the encryption key. Internally, this is converted using openssl::sha256() to ensure a raw high-entropy key of length 32, suitable for AES-GCM encryption. The default is the system environment variable: mmbi_epi_encryption_key.

Examples

if (FALSE) { # \dontrun{

post_data(iris,
          url = "https://some-server:8000/post",
          compress = TRUE,
          encrypt = TRUE)
} # }

# MANUAL WAY -----------------------------------------------------------

# use create_json_body() to make a(n encrypted) JSON of an object

iris_json <- iris |> create_json_body(compress = TRUE, encrypt = TRUE)
#> Warning: In read_secret(): environmental variable 'secrets_file' not set

# curl -X POST https://some-server:8000/post
#      -H "Content-Type: application/json"
#      -d '...'

# replace the "..." with the outcome of create_json_body():
iris_json
#> {"data":"XTDysvbIQy68v23lOwabjVZsXnFidnGbSxMm5qsPgwUqu9uDBrL/V5TrFYyQw0p1ivhYVxd9\nQqcXlWx5l69PGSF7j5l0QzHZ5sicqH+CedIVTB/deUwn1ZPPg7Lv6GqpjH8t1Ivq8tpVNA+e\nSgWzr9ESUHnwrKIuYLRD7NJK21baCLUEr6pzY0kI8GDiaCf1os5irGW6ih1EF4iMSkMqSRP3\nyCJHkjIiaw4SaQX8JicrkXDIE9Chkgv/bByPck1U2GYI7BXG9/HTkswJsZR8mLM3B8J5GxPP\n3JWfaqaIPxhS/KP01lvh1AJL4FA8BVWDlkfV3ZKzz5dOMNcjUouhApTahdomfxRhC/QVIwWC\nLvuHFYF5Iewp3glvwSA4w0OaLZxUBewQWcOWruBfJLk607YJsu0mN6Z1sbcwptII2JvDv9pY\nUWuP1YEOgvOSsxYt2Q0W4wR3G/idpvgIFKwHEkGngvASJHi4fHcUpaCBTQ+TVa7PHO0V/nl2\nMDHotas57uB5oGQMGAdA/CRmN714qlqrt2h6MlzzMIeklHZ8BN55dPmnlQgVxhY6l49jSR2+\n0TjGs4CNTUvAra0Kp2OhVYXc9OUgQFXSqMquYVw07/wsomIpq9X/eyfc3aocCj/U2H1dNeKh\nEYxPmE5O7/bTbNZYAMO2PuOgxFNqyeC97MYHeqpzsCGsTeIDjcaxic5/RKi3tGgwo003ir72\nXC6WwOAtCB8aaw75oKTwxFwkgdDhiBl4GE7hgkbdy0THd3g4lB7gkpzjKp7E7HC39djf/gif\nc/y8eDLsJEh7hTb5NOJAIEc/ig+BOcC2rwC+4+10l3Iuj6t+q+LgI1iV8XGvLhewavtno9NY\n8ueF2M4ivO32gisauObH/QOr0UNzDZkOZ/JFCAu2acdZ9t6hnW5r2Impi2+jdjPv/IZyy7Lv\ntZSuz67e4GgR7RSxDsx/In3EhVstXKLXMBi/4bB6O9dH8UAXxIdayVWaFCCnujO58aisCc6N\nT7Nn/84aLkVKcQF1d2jHklcfFCijO28A601ZzQyksW1G2WKPvZGsJZMhY/YrpUYGP/LpGmlV\n1n6+j7qu0UjrSE7/VraOTvScXSYxGNbQ207NyXnH0tCqu5txe/joqa+WMaP6iZokBrFb3wIS\niH/JboxBJzgbpwVC+0Th5DCXt8Ja5hVCm7DaKnIs3d9CcTS7kLQ44EYG3V0zmEbljuUTWTPh\nRbkK46Hnn6bftn5jvUQ1WB6a704B6XO9e58+h8b7KxGd5ZqD8vThcbP5Rf64MBeQ9fMMYqhf\nbrU1kuaf/luKTysDYDy47yL9PKLxQttkUEEC8naFIstwBJTx1rPjOdiGwpx/n93VcMF2D3Hk\nyjQmmByynyeUcLoHCyFGmHJ/XY1ot7SyhJiCTLk+QhvPnJvq12fL8xC67HQzuCvTZvmmrCHF\nkBC6A+RF7tHKH7W9s0gkpoj7TG5Kei91pt1R8aO6Gwcep3CDZ1EJMpzZAQ0NqFByh2C95Fuw\nwU1rLmOb+CJbm4EVVtSoPZdCk2PuhaF46ImENHon5hT8BHSkzMghVLy0epmY8Ym5vSUb44Yi\nVTw+7N9qAkUuOnJdh4E83LNlMWNGYBv+ipU1vOO/aFDpQ12GNVKIeBcAkZ8a4bEsVJ0QuET5\n7JRy06nv8RGJsYJl9WJaYgPOzaR3i4/V9P2nfRa8P2ezSxc5WLntIDK7kbRQHF+XtqP0t4Ac\no8d+8hshc71WLzLkgm51DPp58HDaUEmnNXVYiRJTf5mYp7lZAJHFLnpDZ7tJ+zDgJqggFCwf\nSIb2VyozcA4PywEbabvqoVeqQdeb/fkNwItCbRuDMwk7SfCUIND+irGFniIlwTpnI3PNcu1H\nNbctthFMSskxtlnIh5IywxgfdK8O78ogUA/GXggydafEKZsMseqeLYeyOmxzxDEwi4Gvu5rK\naYAV0HpNNuoE0yDgTl55mIwqXYJxWVAVfeXOX+LDJeTm","iv":"570hIX7yFLppq1oj"}