POST Data (Encrypted)

Securely transfer any R object with full attribute preservation and cross-platform JSON compatibility. This uses the same approach as the mmbi.epi package of the UMCG.

post_data(
  object,
  url,
  authorization_header = NULL,
  compress = TRUE,
  encrypt = TRUE,
  key = read_secret("tools.encryption_password")
)

create_json_body(
  object,
  compress = TRUE,
  encrypt = TRUE,
  key = read_secret("tools.encryption_password")
)

Arguments

object

Any object of any size, preferably a data set

url

A character string specifying the target URL for the HTTP POST request. Must include the full scheme (e.g., "https://" or "http://"), hostname, and path.

authorization_header

A character string specifying the value of the Authorization header to include in the POST request, e.g. "Bearer <token>". Use NULL to omit the header.

compress

Should the object be compressed/decompressed? At least allowed: "gzip" (or TRUE), "bzip2", "xz", see base::memCompress(). Use FALSE to not compress/decompress.

encrypt

Should the object be encrypted/decrypted? This applies AES-GCM via openssl::aes_gcm_encrypt(), providing authenticated encryption. This guarantees both confidentiality and integrity: the file cannot be read without the correct key, and any tampering will be detected automatically during decryption. The initialization vector (iv) will be a length-12 random raw vector.

key

A character to be used as the encryption key. Internally, this is converted using openssl::sha256() to ensure a raw high-entropy key of length 32, suitable for AES-GCM encryption. The default is the system environment variable: mmbi_epi_encryption_key.

Examples

if (FALSE) { # \dontrun{

post_data(iris,
          url = "https://some-server:8000/post",
          compress = TRUE,
          encrypt = TRUE)
} # }

# MANUAL WAY -----------------------------------------------------------

# use create_json_body() to make a(n encrypted) JSON of an object

iris_json <- iris |> create_json_body(compress = TRUE, encrypt = TRUE)
#> Warning: In read_secret(): environmental variable 'secrets_file' not set

# curl -X POST https://some-server:8000/post
#      -H "Content-Type: application/json"
#      -d '...'

# replace the "..." with the outcome of create_json_body():
iris_json
#> {"data":"kqwoDvlST3ETgLTf1kv/JUJsdDdY/0vZ++9nnLxyN8I5XvTlzxhTrGgXFC94qdij6tNjFMkt\n5jtrbEhdr0HePv3oY4yXuDD3v43kUhCxFDOm6eBDV8E9B3k+Z2KUVfj67pq99Cds33C+VdR7\nFb/I3GWZyBwtPVNf9V9bVrHOdOqSB3YRmOlnlmMGG+xyng7xXRk8Ewm3XbCcPdyg0sO8rWmJ\njk3w3xVXj9e2s6YExVE893yoJD/CdGkZ0gQcFOPjRON3DhBTFYXKNG7v449+YHGkZ90EAiC0\nmFMRrSUHsmSOT5SWaQRlpe7D4qxmPkWWynNNMc5bCW41ABPn6LpDnu59ZBX2KYYGEsPmY/mU\n6eiOPawyc58ORjswmuN0BB+LlpoAz5450aEmm1EfnUwvMYCSt00Z5/mywIjtrpO39KpcERTe\nVJ+9cpNNKmoB0p1BRpj4W1mNcTbRiS/kE5+bewudKQJEznn/pxbsMiuGySCQ5iwJ/zAMO5BB\n0wMVCXzkrOW6qwFLGEQMBrBo+t6Sxldw4oUZHwMBBhBIXb/qbPs6w1/cs+ZIYIahAVlT2B9D\nvEUJLD1BuKl55oCNnnTZpObiYL1gh82molNyvOZc8o3ukLYLkCubCPQ29ZKJG8cF8CBAyXMN\nFUiNUful+CFwt7NN/yfsJ32Ps0aAaS4UM98fgrvkPafjPIyoA32/Sv6ohbTEENl3E0UsgxyI\nulxJxtHSujdXWT+zK14OtDC7BJLrfeoJJnnz5pp45PK4RnursHxx7JkzkPs3xWz8FzgbLepk\nEcMLM5mJHwPHZ+U/TgSf7Fl+ET/cB88bKUfpxG6D1hNJSTUOi18KhBVMiYelCvZ4GGiaY+p9\nxESAEuFPyVnzykZPfitBktzwFenCyl2+0vI5PvvD7bOZFn8VVIbgdcDXCo0t5LIj71P/eqtA\nuHfu/D9UfG/pFukUjs6MbMItvk/VpRIVL3+UbSlvNG1yrPseUSXHDzNjWoPZzURrvJDOcSt6\ng6SygkEEqSkC2YVenVESjWViEdsbBM/kfdA4vTwnp1Fy9J85HFV4CRzMc4dQjARKvHXpLTBv\nIy2Tki06mDGGEuQsFDOHATYoTJ29SgpInlPhfqOOBsYXvv1wcCEDMpZ8ZSHuP7nQyJxNfEVk\n8YhQOIssboXrH0FeipHcgx2+X57pVQLkLPAX5/VYbJdjYzcYfZdCaIwMBUQizBXQDLJKkqt4\n4+dUh84Yt0BV69hr8mzC/cwWZJ6pFGKzgLn6gCL0DquiddyGWwwZc4NyT1EQk0T/Byv8RgEe\nVqjWpEkm++JoiC9HEpIMLcaKDIFvXhzPy4mbORGsuWXbSNvAXRsyEd4bvkGKFUhZt6498sNR\nvMRipfs8u1+O8GvnUDm1RTUzvM/0f4qqi+hiO7OE+dHTy6EYluNdOx1xZmtY/2Na3IqFB74x\nib9dA0og00G2Lf/bdjvONmAl2RK0wB6daq5y7Ok2I8+ZR1Ly9QdYprx0Swz/ERDnN/gW8D/z\n3UZAWcdydY58bHfvV3tmeNERqxOUNrEhKBJzEciplRfzT/WIEKDYYEy9+KU6x+fE1hFMlMjm\nZyl3ZnN0CMFrN7JIqPv/I1c+uU72qvqac/+a/FVwS/WsoyjulwHIGPdqWHlp3BPvRFJBfWoh\nUAzdgPbMn3aUGHxlCDUOdFAqTjWUVof/6I+BKBIhLHP73RKCR+fXVAhfe874C6haucmWw3fx\nlK8pNOviL0jjBV3PU+hgmm3dtRImdQmh08XCyBXFsSS79Ze6ZYHJbcPJ/rhNnq1HJeMnSuuL\nIevzgx/j/f4syltv1Ywj7qpkxVHGBmnMgOkmYxFqWDy6n4r1SBTb80X9SWJL230UAf8D5FIY\n+bCReDa7fu8b5mm8dTOQtZwuRvVj1841iiCH0qMobWqmj6Iq3b9w51X7YWWXQv19BacAL2mD\ncrvgYgIRd8NBcJJ1HPtTpm3uzwJetnp8PvuyD0bnUqV3","iv":"Uvcf30IyKdPu99nu"}