POST Data (Encrypted)

Securely transfer any R object with full attribute preservation and cross-platform JSON compatibility. This uses the same approach as the mmbi.epi package of the UMCG.

post_data(
  object,
  url,
  authorization_header = NULL,
  compress = TRUE,
  encrypt = TRUE,
  key = read_secret("tools.encryption_password")
)

create_json_body(
  object,
  compress = TRUE,
  encrypt = TRUE,
  key = read_secret("tools.encryption_password")
)

Arguments

object

Any object of any size, preferably a data set

url

A character string specifying the target URL for the HTTP POST request. Must include the full scheme (e.g., "https://" or "http://"), hostname, and path.

authorization_header

A character string specifying the value of the Authorization header to include in the POST request, e.g. "Bearer <token>". Use NULL to omit the header.

compress

Should the object be compressed/decompressed? At least allowed: "gzip" (or TRUE), "bzip2", "xz", see base::memCompress(). Use FALSE to not compress/decompress.

encrypt

Should the object be encrypted/decrypted? This applies AES-GCM via openssl::aes_gcm_encrypt(), providing authenticated encryption. This guarantees both confidentiality and integrity: the file cannot be read without the correct key, and any tampering will be detected automatically during decryption. The initialization vector (iv) will be a length-12 random raw vector.

key

A character to be used as the encryption key. Internally, this is converted using openssl::sha256() to ensure a raw high-entropy key of length 32, suitable for AES-GCM encryption. The default is the system environment variable: mmbi_epi_encryption_key.

Examples

if (FALSE) { # \dontrun{

post_data(iris,
          url = "https://some-server:8000/post",
          compress = TRUE,
          encrypt = TRUE)
} # }

# MANUAL WAY -----------------------------------------------------------

# use create_json_body() to make a(n encrypted) JSON of an object

iris_json <- iris |> create_json_body(compress = TRUE, encrypt = TRUE)
#> Warning: In read_secret(): environmental variable 'secrets_file' not set

# curl -X POST https://some-server:8000/post
#      -H "Content-Type: application/json"
#      -d '...'

# replace the "..." with the outcome of create_json_body():
iris_json
#> {"data":"/FUA3iYmmSmybokJh3SUIjOZOM5ym3EJoC+hzpz3NbH3G0FCZjqFpjTdexF9qWQVlZTy1CGs\nuQxzeSEc2nP4tikip3OZYJNE20tq61uglWKDeu64+PUrMwp+Egxh9m8mT8v1VbG8bBqBmqxO\nt5xJi5PG/6QzHnEMhSgKL/iljyANOt6nFz741LdLyu6xraDcSqTKsoU9TZ6y3dYBR9qcPlqu\n5BCJHzkpxVpR40q0SN4Ke+G3J2Wk657ikIBVhGUSolxAarHVew6YnIwFbnxNuFsYuOJJf0MK\n9e+7R9Sf/2tZf4oFlllivpGIGrYvNhOPWwJ4Vky1AVLCKdzbd1U/RTeF/e/QDh/niwlrw8LF\nP7Qcx6DEnmxTF9xRUjRL03lxq+aPumE1uKG3v4n6LZsYhSxxC5hL4GOn8weispaYBaqs+O+x\nFTu2+m2elCFinbUlgrcMJGX8J9xvGs2CXUDsOw4IIOQ9eYVGDZoaFSp1qQCaisc29l6eiGKc\n2T32uOUb5y0wtgfMi00WF94hkAVE2lzvpk2R7f2OsElsPRUeXtFpvM9XS84sjAoa9esJurCu\nx183RTvxo9UXCG/+CrtNn99DfvQhGcZ+LYdO+1q1oFk1/uHjv24yW7YOLhMNyHj5mrSWkEQ6\nmIwPqH+Jgq48WVZ/8ETMkYkTfCn4JUAN9TyyUF7lvxxxXs+UZnOUBeiUeifG+cB8B68cu3Oo\naY0RWn3nF42dm6LvejQwB7F1rbZdT9HA0/p74vssZKdJFjq4XMJP/egDpu1GzuvfJuviy50b\n8E4v9klz1E/srkqJtL02Ao5n0+I7HXuFRY4xJqCjS/reeBIbKngACfJea/V5cz+c+lspmPiT\n9hnzK0MLtp9CVuXSamlCWOuU+owgB5OpPZ0YQhQewtQmYWnSyAYiuyhXd+zzK8ErcCljDxe6\nSv1nHjRr+f3i7M0KVgGlZdm5Ufjxab/jSAICvZj6suKauJW6+Celh4iOKvRnHka1QYHfNbCA\nIeYMvPaLrDAvC/lgIJLKgOoYjYDdE3fl+IN79Myl515kvxUwWGMkK4KT53+lNzE95nyAfXud\ntQp8gWKiSOqLHcEctW3EWiH0sFhiSX8KNtR0daahA7ybbgyGK4OFtt30Ey5c48W1rQeXUPO+\nh7ObzLFHB6xBIwFE61bSilUux96LiEoIh2bN9Q2ruqjeOmQ5Wf4JLjPLVPakmAggsGeaALNB\nsEuAnkJxlTeDIn9bsg0zHgcKa7F2f28htb0aMiiX7j4aovXgH3sg705/oafNuaDDI03ZmIse\nC4jnU9xtISjaGQT9gdguZ9hkQy69fJwU440HMcGSCZbXgCoXZNSt7lg3wQDouVRPAPrkbfd3\nnJ4OvycRZVp50TtiUQwLd71X3UQcw1rAwNJRIeEqVtY6HQFP1L6yp/fUFdXn+QgOl/Nq7WG2\nrHsDU/I6WMSE8i9byELRBQC/QP3AGhgvl0WpkUyqRcAsHX0ZpOKRGbY7HR5CgOl945zPKHpK\ni+xYVNi1wEEpnIchWC2v9nzupA79yF0Edu6OmPr33zCDqHhh2EDy6w1+ddpVw7BneMXlg8Jn\ntG/ttAuhBneEQyWDDcOtstOTl8xTZKg6SVC96Fdgsayd15WyJ728L3FwiVOOPAlzNILcntsS\nxWGRzMagMZFj9Dhhq3KRGintcMgzM8w4APDlUMOcUagXClZIiCk8FHmHIWHbXVTrFlErMmUM\ngCz4xsGci+4/pO1o97cQbWnGm0HKbKYJ4sz6tbPUU/e3wrtF+OFO8Zua5CXUB+C1duT1WaF4\n3VJIHB5XFEUon6oCHnFXTozFifHllG8gcRxUH9l5Cl7QQWeJAdvoLU0NuLxrPofJ1OdzmH2k\nq4vOTNkcjEmCGnYmrwmRt5+hz8RK7GH3vSry/x1e1KIPJqwWRbsZcMgJ06NGuvexC4Qlm7yb\nN1v9XJXnOIKrbegDDV3RLaEw8BczgU8YGAwNPE1LH183","iv":"nc+JmEux7OhjOCgD"}