POST Data (Encrypted)

Securely transfer any R object with full attribute preservation and cross-platform JSON compatibility. This uses the same approach as the mmbi.epi package of the UMCG.

post_data(
  object,
  url,
  authorization_header = NULL,
  compress = TRUE,
  encrypt = TRUE,
  key = read_secret("tools.encryption_password")
)

create_json_body(
  object,
  compress = TRUE,
  encrypt = TRUE,
  key = read_secret("tools.encryption_password")
)

Arguments

object

Any object of any size, preferably a data set

url

A character string specifying the target URL for the HTTP POST request. Must include the full scheme (e.g., "https://" or "http://"), hostname, and path.

authorization_header

A character string specifying the value of the Authorization header to include in the POST request, e.g. "Bearer <token>". Use NULL to omit the header.

compress

Should the object be compressed/decompressed? At least allowed: "gzip" (or TRUE), "bzip2", "xz", see base::memCompress(). Use FALSE to not compress/decompress.

encrypt

Should the object be encrypted/decrypted? This applies AES-GCM via openssl::aes_gcm_encrypt(), providing authenticated encryption. This guarantees both confidentiality and integrity: the file cannot be read without the correct key, and any tampering will be detected automatically during decryption. The initialization vector (iv) will be a length-12 random raw vector.

key

A character to be used as the encryption key. Internally, this is converted using openssl::sha256() to ensure a raw high-entropy key of length 32, suitable for AES-GCM encryption. The default is the system environment variable: mmbi_epi_encryption_key.

Examples

if (FALSE) { # \dontrun{

post_data(iris,
          url = "https://some-server:8000/post",
          compress = TRUE,
          encrypt = TRUE)
} # }

# MANUAL WAY -----------------------------------------------------------

# use create_json_body() to make a(n encrypted) JSON of an object

iris_json <- iris |> create_json_body(compress = TRUE, encrypt = TRUE)
#> Warning: In read_secret(): environmental variable 'secrets_file' not set

# curl -X POST https://some-server:8000/post
#      -H "Content-Type: application/json"
#      -d '...'

# replace the "..." with the outcome of create_json_body():
iris_json
#> {"data":"U5GV8DZ8OLzckgkEh/Zne6bszzEQiuRi4J2ZZ6jYrFZ159wQYyh6PmT4Jw99dZQ3MPSfHjCN\nmQCe66DtxuayFXV9YhoiD/lZDnwQac8j8einXlARl2SOrCAQoVUCZGc7wKqOxttqrKsClTDN\nZgUMjHf0u1Gn6iIBcxmDa5H7+dhoeMSw+LcTA36YF749NK9Z68NdDgJo8zAVIqDgqxKoWpr3\niJLfo4wniLp5v5iWlZt+OwOwUa71ifTQZ2h/PYvXA1SLR8tdgg4PaUMoUVW5g7HhfrKaIy4e\nq1JiTZivZkAsMGwnk4aAKAdXJm/W+M2rTgMMR0izmG2iMJonARTdMtRhkrYzLTPLXBQ04NKI\nSSZw51k7A4tJH3jFKzCkR7Wl4ww+H9CV+xHUVf4P65YvK725SgnwTQoRhKkRGJgq7ADup3+z\nFHnxoQ2dIpN0fB8QkqrxVLRKEB+8yUYpTNYXidYEXNNefrMdhtFyaaDOpzMr4ExDk7PNK/B2\n307JEUjkdBFSoLjRx3joWjuI0ZtfgelapHyUVpzimDCHYyhVccAkqTDTn1vexU2LUszFfYRo\nZzNy4/uXKFWkL3SL30VB8x53G+Jceo/SmUK3wY6XK31Az4gKeSphCsSFuZ3XzuhOP2Tzj7dH\n2/w8aPXUKwAVO8QfxHHy849QE4xbGwchxr56nPuYmdflTp6Vy22D3q2dML419k/91qr9Sq6k\nAzwG7bovw1vDJZJeGYzLIFwzZHSPnioVEYWN24+OtH/lzj2jWObngYx7n7smE6Q5+t7vjetb\n/kzCtztjByt7daXygEztFAa0+MdxFn3rX/1vdhx3bgWyKbZ4QTMPUndVAA7XmoykN0EmQSy+\nbw1g0lCylTk3bySq7KjaylaMmj2rw0jZzXIdJ/2D9saoxa4pkKZhjksPuYUt53FwaJCdTecn\nMtFAL3xaNxNaSnLrVQP4uawaxHI2a7nWVK7HwDiKDvyrL9nSvi84yWd4i6VCgabkrI09EeVL\n8CoPwGJHIiP/ac/dCi6TLry1tj2PdbLyM0qztmcQl4zCdstJDw4LOL/f749zE1aZhMKfcPmP\n2M/Y1kYuoDgxnqVeffrW8L9PnjRxxXFYvgXfQiI21UnhptAvAYFxmQIGlvL9OZha0wiFWxhQ\n5BiY5PxOX2A4JksS5zqaPjLJVsi5XFUQgA5YRbTreAFiJhWEz9GqFsyvIcHDjjGgdyRH+Zmx\nCTlTscWFQk1iESHxV/eW8WdSkIR0EJdflDceAt/zMgd7yDwx/2ny/9yDfgNHIAZ/9dv2pklp\n29sB6bxs8d2CH+q0OQc4Y4n+ErlXagP7XijwizPsymZKUUJiIDynsXA7Aka8lRJS+sX+5K39\nC5LGM6iZKwJ1v/kvaB2vysWqJngp32qPPm9adTaZAaaRdeF7HZnllNf/nqER5Coi+aUp+W2s\nAGVUqFLZzF75IlTH2xofoCzETINSCP7CGcPUkUSqgjP4K56+hLd5AMglVEij6/Gz/Fn+a/Jj\nzM6Cc46nwPnc8CkcE1R2JswpcPzuGexc7MTBIt35qosPMnvdffmt5WUONcui3c6YlWvdHYZk\nEt+qM+hUmjCrxEsmOs+1K10U1N9ttBGSIL5p1eIUuQiyd5/8TAGknkoBKjRyIhQbtHtg3bmQ\neLRk+CAUnnM7F8eS8NYuUp7jwABIGTNNpRuEREsDBPMD+yBnptkZBj44IV2yJ9GDbMlLdFoQ\n0JdoPLHvVqSSC8trH35VUYmpZww5hcMdM81LE+eUdaOLVKE8SKPB3hujxjJ12OnUGltZfK21\nMZcalV/eV3Vi/5YLxjJTpWb8Y1sehHJvKaZd+HCONK61b12W9HP6lAeQuqDmA+uMou1U94Qx\n1S2fBwJkO01hZXTkd1XTnCmbNcoPY5cTVx1zgYM+Ix1R959gUP+Qfviwc/BMYrFl7vsXQveL\nv0YLOsg2FI4YkE0o/ZcTfnTdj+bnMww6UxRto5RC1fiA","iv":"QcrVYBDy3omKmJLL"}