POST Data (Encrypted)

Securely transfer any R object with full attribute preservation and cross-platform JSON compatibility. This uses the same approach as the mmbi.epi package of the UMCG.

post_data(
  object,
  url,
  authorization_header = NULL,
  compress = TRUE,
  encrypt = TRUE,
  key = read_secret("tools.encryption_password")
)

create_json_body(
  object,
  compress = TRUE,
  encrypt = TRUE,
  key = read_secret("tools.encryption_password")
)

Arguments

object

Any object of any size, preferably a data set

url

A character string specifying the target URL for the HTTP POST request. Must include the full scheme (e.g., "https://" or "http://"), hostname, and path.

authorization_header

A character string specifying the value of the Authorization header to include in the POST request, e.g. "Bearer <token>". Use NULL to omit the header.

compress

Should the object be compressed/decompressed? At least allowed: "gzip" (or TRUE), "bzip2", "xz", see base::memCompress(). Use FALSE to not compress/decompress.

encrypt

Should the object be encrypted/decrypted? This applies AES-GCM via openssl::aes_gcm_encrypt(), providing authenticated encryption. This guarantees both confidentiality and integrity: the file cannot be read without the correct key, and any tampering will be detected automatically during decryption. The initialization vector (iv) will be a length-12 random raw vector.

key

A character to be used as the encryption key. Internally, this is converted using openssl::sha256() to ensure a raw high-entropy key of length 32, suitable for AES-GCM encryption. The default is the system environment variable: mmbi_epi_encryption_key.

Examples

if (FALSE) { # \dontrun{

post_data(iris,
          url = "https://some-server:8000/post",
          compress = TRUE,
          encrypt = TRUE)
} # }

# MANUAL WAY -----------------------------------------------------------

# use create_json_body() to make a(n encrypted) JSON of an object

iris_json <- iris |> create_json_body(compress = TRUE, encrypt = TRUE)
#> Warning: In read_secret(): environmental variable 'secrets_file' not set

# curl -X POST https://some-server:8000/post
#      -H "Content-Type: application/json"
#      -d '...'

# replace the "..." with the outcome of create_json_body():
iris_json
#> {"data":"DTUbTHTvhbQdPKHc/3UmLeAh05Cyp7XPmCvJWkuEFiS4RtS+W2tASsL6RJa09Io1M/xJK3xP\nctsRNSV2BIfBi6BBF+sMBHQDLCXiC94CQnh0hLen6KRB8oUi4lkROsII2vWF/He+TPdetmPb\np/PUcoS6dPe9VTrRBjxootbTwtrZqjQcXA8cldBbf5DC//9ofPXNOeq8Vs7L1evVdAAdg1TT\nNJa7jWB6AIfa6T/o3HEazjv9XLW/soJSy8bsGzBq4hkGZnNAOrTYVofOBL+mBv3UGkPLGVXS\n1Pe50qnca+koQgAgxPaxdQDG4QQbResiBQSItnpM8Q1yG61rqpk/tEXGqTOQQVSHitEkc0Yf\nsdmyV7KhhN9pvKyfVMsQekX8P3StEi1CBsNOcmubGF7l53h7kZ/biHmbfaXsqvIyjeWgzXxf\nTuvVaTTKsuz2wh09VbRJKTThEFJdnPshQVVY6XgYUQqLgQkK1pDEqfHHHn5DrPWgoGJ/SR0C\n7PN2BSChhCnKwa+5rpT8naPldvKNMjdz5lvGi6zVeOtJKPvI6DHJ77maEVSMqtrpSb7Blzqu\ned0e7chhsJIsM0ubbbP/JaiFP97R7giDfVAnbDnQFyDeZD+eqBl4ULXm9XMoCYLIx0SzZkXW\nCPrFk6gDcNlvkA0S4GJ6KSYUvEYNkcP93U6uqATre70xah+JSsQcKB10IOBilCzl7FO9YpIM\nvgSz8+MD/9hf+bzD4bQTn4SWjiENvvdaJebYxBfZBmbhG8BAFBLT934XHfRm0SJ0AOAUrJvl\ndCEV0MprFeAaOc2GXEKaCC3mp2KYr7gmBOb1nO4t48vmU299yEqaAg+lLzt5SYy0drFl9mfR\naRCDhox951ycaPnd38F25oII7kkKigfm9x1hrv4tHyAPRscD102BiQM83525RWGGqsC10EQ3\njoxiZfaGfIhXxdgs1Ut+siBGG1h96YOZU5sK7XdR4D9lkHReISpvgY3pzaA/eS/Wdv3Y4oNz\nzdfLdRml4OWV5ADGRExocLwguf1e6GwVkFVSiyY/x1F4EbtznJVuLh9gwHjKY/fguAwXPU9J\nxXD1y+wzNFVXcYcYFiY7wgW4f+pWCo/Qbg5sR6v/PvtEDU91bIhS5MwlPzm3NgdsRHfAem1x\nJZmfNshxXGu6RqIwZyJ6E4PW6L4sIcspgJTFAVL3PY+9+QEN7dq5MNZ99azkd5hHiWTOotvB\no55/yexauRzrDw3SnGpbSXF/zmkoME1YQeQ1naBl6sucpBJI7gkDl1RpP9nADKHdWFzZZ8Bd\no5cVFf36aXoTMB3osC4qOs2CnR+tRNIch8WIVF16lzXGvAYsIWTdnv3rdZ4HGmBi763oLwKl\nDrbDPe8scmlfCyT4GH+EQkhYLkyDYCA0Pb3BFUuGGGQ9V7B03jHF0e0kQvV3w+gV00zCZUGZ\ndmqa56giRASy2gHY4Yie8t2UgT6gLOTPMmg+my0CLV+AkpK11q0vEbIoRbPCTlakWCKZuF7c\nLKzRtVEIu7LNIjhipnvYe797+6alxyskSx15MOLfGt0aLiW0EbSwCo3Iw6W7t5rTmX6T+WKw\n76kIYV09/HGzzpfrxFLjy2c/UA30Hf7KOaN1WdlLOzBJlUzVQAy77/rYj8fDjzzK+j9hKPaP\n4WoJvtKfrD+CJNS6Z2Vj/Jp6hOk5RTdneChmHrvHWxCyx+lsTWhUm2XINIaNOjJff0ST2t03\nMNHIRJkU9eNUqDW3veKj0gkr77olN0kVI/NgRzcu9FtZkzxvo3jSKFQ0D5PI85qx7WQYt3nD\nP3218RGZnYDMAnnjsENTYVkXaNT4lsl+0o4WP0Cp/lTJAxePy56bFOP4MusfQuWf9ugZeAnH\nAky5xgNlgZ1QLC5Vp64to1nKMLfHn15ag30SRdhib2iL7dNaViiMCIKv4jANQDZtrrDWOlWa\n+bX1L+4iMEjFPY/ME1nOd+o90K2ykeJXp1lpZb5GUmP7","iv":"5rVmWWFGN/+NxgrM"}