Format Table as Flextable

Format a data.frame as flextable() with Certe style, bold headers and Dutch number formats.

tbl_flextable(
  x,
  row.names = rownames(x),
  row.names.bold = TRUE,
  rows.italic = NULL,
  rows.bold = NULL,
  rows.height = NULL,
  rows.fill = NULL,
  rows.fill.picker = "certeblauw5",
  rows.zebra = FALSE,
  row.total = FALSE,
  row.total.name = "Totaal",
  row.total.function = sum,
  row.total.widths = NULL,
  row.total.bold = TRUE,
  row.extra.header = list(values = NULL, widths = 1),
  row.extra.footer = list(values = NULL, widths = 1),
  column.names = colnames(x),
  column.names.bold = TRUE,
  columns.width = NULL,
  columns.percent = NULL,
  columns.italic = NULL,
  columns.bold = NULL,
  columns.fill = NULL,
  columns.fill.picker = "certeblauw5",
  columns.zebra = FALSE,
  column.total = FALSE,
  column.total.name = "Totaal",
  column.total.function = sum,
  column.total.bold = TRUE,
  align = "c",
  align.part = "all",
  caption = "",
  na = "",
  logicals = c("X", ""),
  round.numbers = 2,
  round.percent = 1,
  format.dates = "d mmm yyyy",
  decimal.mark = ",",
  big.mark = ".",
  font.family = "Calibri",
  font.size = 9,
  font.size.header = font.size + 1,
  values.colour = NULL,
  values.colour.picker = "certeroze",
  values.fill = NULL,
  values.fill.picker = "certeroze3",
  values.bold = NULL,
  values.italic = NULL,
  autofit = is.null(columns.width) & is.null(rows.height),
  autofit.fullpage = TRUE,
  autofit.fullpage.width = 16,
  vline = NULL,
  vline.border = officer::fp_border(),
  vline.part = c("body", "footer"),
  print = FALSE,
  ...
)

Arguments

x
a data.frame

row.names
row names to be displayed. Will be 1:nrow(x) if set to TRUE, but can be a vector of values.

row.names.bold
display row names in bold

rows.italic
column indexes of rows in italics

rows.bold
column indexes of rows in bold

rows.height
height of the rows in centimetres

rows.fill
the column indices of rows to be shaded

rows.fill.picker
the colour for values in rows.fill, is evaluated with certestyle::colourpicker()

rows.zebra
banded rows in the body - equivalent to rows.fill = seq(2, nrow(x), 2)

row.total
add a row total (at the bottom of the table)

row.total.name
name of the row total

row.total.function
function used to calculate all numeric values per column (non-numeric columns are skipped)

row.total.widths
cell width in row total

row.total.bold
bold formatting of row total

row.extra.header
an extra header to be displayed above the table

row.extra.footer
an extra footer to show below the table

column.names
column names to be displayed. Can also be a named vector where the names are existing columns, or indices of columns. When this vector is smaller than ncol(x), only the first length(column.names) are replaced. When this vector is longer than ncol(x), all column names are replaced

column.names.bold
display column names in bold

columns.width
width of columns. For autofit.fullpage = TRUE, these are proportions to autofit.fullpage.width. For autofit.fullpage = FALSE, these are centimeters

columns.percent
display the column indices as percentages using certestyle::format2() - example: columns.percent = c(2, 3)

columns.italic
column indices of columns to be displayed in italics

columns.bold
column indices of columns in bold

columns.fill
the column indices of rows to be shaded

columns.fill.picker
the colour for values in columns.fill, is evaluated with certestyle::colourpicker()

columns.zebra
banded columns - equivalent to columns.fill = seq(2, ncol(x), 2)

column.total
adding a column total (to the right of the table)

column.total.name
name of the column total

column.total.function
function used to calculate all numeric values per row

column.total.bold
bold formatting of column total

align
default is "c", which aligns everything centrally. Use "r", "l", "c" and "j"/"u" (justify/align) to change alignment. Can be a vector or a character (like "lrrrcc")

align.part
part of the table where the alignment should take place ("all", "header", "body", "footer")

caption
table caption

na
text for missing values

logicals
vector with two values that replace TRUE and FALSE

round.numbers
number of decimal places to round up for numbers

round.percent
number of decimal places to round to when using columns.percent

format.dates
see certestyle::format2()

decimal.mark
decimal separator

big.mark
thousands separator

font.family
table font family

font.size
table font size

font.size.header
font size of header

values.colour, values.fill, values.bold, values.italic
values to be formatted

values.colour.picker, values.fill.picker
the colour for values in values.colour or values.fill, is evaluated with certestyle::colourpicker()

autofit
format table in width automatically

autofit.fullpage
display table across width of page

autofit.fullpage.width
set number of centimetres to width of table

vline
indices of columns to have a vertical line to their right

vline.border
function to define vertical line

vline.part
part of the table where the vertical lines should be placed ("all", "header", "body", "footer")

print
forced printing (required in a `for`` loop)

...
not used

Value

list

See also

Examples

# generate a data.frame
df <- data.frame(text = LETTERS[1:10],
                 `decimal numbers` = runif(10, 0, 10),
                 `whole numbers` = as.integer(runif(10, 0, 10)),
                 `logical values` = as.logical(round(runif(10, 0, 1))),
                 dates = today() - runif(10, 200, 2000),
                 stringsAsFactors = FALSE)

# default
tbl_flextable(df)      # dataset heeft geen rijnamen

#> Warning: OS reports request to set locale to "nl_NL.UTF-8" cannot be honored

#> a flextable object.
#> col_keys: `A`, `B`, `C`, `D`, `E` 
#> header has 1 row(s) 
#> body has 10 row(s) 
#> original dataset sample: 
#>   A    B C D           E
#> 1 A 2,42 6   12 Oct 2018
#> 2 B 6,27 1 X 25 Jan 2021
#> 3 C 1,68 3   13 Mar 2017
#> 4 D 4,45 6    7 Nov 2019
#> 5 E 1,65 4 X 16 Jan 2020

tbl_flextable(mtcars)  # dataset heeft wel rijnamen

#> a flextable object.
#> col_keys: `A`, `B`, `C`, `D`, `E`, `F`, `G`, `H`, `I`, `J`, `K`, `L` 
#> header has 1 row(s) 
#> body has 32 row(s) 
#> original dataset sample: 
#>                   A    B C     D   E    F    G     H I J K L
#> 1         Mazda RX4 21,0 6 160,0 110 3,90 2,62 16,46 0 1 4 4
#> 2     Mazda RX4 Wag 21,0 6 160,0 110 3,90 2,88 17,02 0 1 4 4
#> 3        Datsun 710 22,8 4 108,0  93 3,85 2,32 18,61 1 1 4 1
#> 4    Hornet 4 Drive 21,4 6 258,0 110 3,08 3,21 19,44 1 0 3 1
#> 5 Hornet Sportabout 18,7 8 360,0 175 3,15 3,44 17,02 0 0 3 2

# extra formatting
tbl_flextable(df,
              logicals = c("X", "-"),     # replaces TRUE en FALSE
              values.colour = "X",
              values.fill = "X",
              row.names = "S. aureus",
              columns.italic = 1,
              format.dates = "ddd dd-mm-yy",
              round.numbers = 3)

#> Warning: OS reports request to set locale to "nl_NL.UTF-8" cannot be honored

#> a flextable object.
#> col_keys: `A`, `B`, `C`, `D`, `E`, `F` 
#> header has 1 row(s) 
#> body has 10 row(s) 
#> original dataset sample: 
#>           A B     C D E            F
#> 1 S. aureus A 2,425 6 - Fri 12-10-18
#> 2 S. aureus B 6,272 1 X Mon 25-01-21
#> 3 S. aureus C 1,684 3 - Mon 13-03-17
#> 4 S. aureus D 4,448 6 - Thu 07-11-19
#> 5 S. aureus E 1,647 4 X Thu 16-01-20

# row totals
tbl_flextable(df,
              row.total = TRUE,           # add row total
              row.total.function = max,   # instead of sum()
              row.total.name = "Maximum", # also works with dates
              columns.percent = 2,        # 2nd column as percentages
              round.percent = 0)          # rounding percentages

#> Warning: OS reports request to set locale to "nl_NL.UTF-8" cannot be honored

#> Warning: OS reports request to set locale to "nl_NL.UTF-8" cannot be honored

#> a flextable object.
#> col_keys: `A`, `B`, `C`, `D`, `E` 
#> header has 1 row(s) 
#> body has 10 row(s) 
#> original dataset sample: 
#>   A    B C D           E
#> 1 A 242% 6   12 Oct 2018
#> 2 B 627% 1 X 25 Jan 2021
#> 3 C 168% 3   13 Mar 2017
#> 4 D 445% 6    7 Nov 2019
#> 5 E 165% 4 X 16 Jan 2020

# column names
tbl_flextable(df,
              column.names = c("1" = "Kolom 1",
                               "2" = "Kolom 2",
                               dates = "DATES!"))

#> Warning: OS reports request to set locale to "nl_NL.UTF-8" cannot be honored

#> a flextable object.
#> col_keys: `A`, `B`, `C`, `D`, `E` 
#> header has 1 row(s) 
#> body has 10 row(s) 
#> original dataset sample: 
#>   A    B C D           E
#> 1 A 2,42 6   12 Oct 2018
#> 2 B 6,27 1 X 25 Jan 2021
#> 3 C 1,68 3   13 Mar 2017
#> 4 D 4,45 6    7 Nov 2019
#> 5 E 1,65 4 X 16 Jan 2020

tbl_flextable(df,
              column.names = LETTERS)

#> Warning: OS reports request to set locale to "nl_NL.UTF-8" cannot be honored

#> Warning: Only the first 5 items of column.names will be used

#> a flextable object.
#> col_keys: `A`, `B`, `C`, `D`, `E` 
#> header has 1 row(s) 
#> body has 10 row(s) 
#> original dataset sample: 
#>   A    B C D           E
#> 1 A 2,42 6   12 Oct 2018
#> 2 B 6,27 1 X 25 Jan 2021
#> 3 C 1,68 3   13 Mar 2017
#> 4 D 4,45 6    7 Nov 2019
#> 5 E 1,65 4 X 16 Jan 2020

# vertical lines, alignment and row names
tbl_flextable(df,
              align = "lrrcc", # also works: c("l", "r", "r", "c", "c")
              font.size = 12,
              vline = c(2, 4),
              vline.part = "all",
              row.names = paste("Experiment", 1:10))

#> Warning: OS reports request to set locale to "nl_NL.UTF-8" cannot be honored

#> Warning: Only the first 5 columns have set alignment - the remaining 1 will be centred

#> a flextable object.
#> col_keys: `A`, `B`, `C`, `D`, `E`, `F` 
#> header has 1 row(s) 
#> body has 10 row(s) 
#> original dataset sample: 
#>              A B    C D E           F
#> 1 Experiment 1 A 2,42 6   12 Oct 2018
#> 2 Experiment 2 B 6,27 1 X 25 Jan 2021
#> 3 Experiment 3 C 1,68 3   13 Mar 2017
#> 4 Experiment 4 D 4,45 6    7 Nov 2019
#> 5 Experiment 5 E 1,65 4 X 16 Jan 2020

# width of cells and table
tbl_flextable(data.frame(test1 = "A", test2 = "B"),
              vline = 1,
              autofit.fullpage.width = 16, # default values in cm
              columns.width = c(1, 3))     # ratio; cells become 4 and 12 cm

#> a flextable object.
#> col_keys: `A`, `B` 
#> header has 1 row(s) 
#> body has 1 row(s) 
#> original dataset sample: 
#>   A B
#> 1 A B

tbl_flextable(data.frame(test1 = "A", test2 = "B"),
              vline = 1,
              autofit.fullpage = FALSE,    # no fullpage autofit
              columns.width = c(1, 3))     # cells become 1 and 3 cm

#> a flextable object.
#> col_keys: `A`, `B` 
#> header has 1 row(s) 
#> body has 1 row(s) 
#> original dataset sample: 
#>   A B
#> 1 A B