Format Table as Flextable

Format a data.frame as flextable() with Certe style, bold headers and Dutch number formats.

tbl_flextable(
  x,
  row.names = rownames(x),
  row.names.bold = TRUE,
  rows.italic = NULL,
  rows.bold = NULL,
  rows.height = NULL,
  rows.fill = NULL,
  rows.fill.picker = "certeblauw5",
  rows.zebra = FALSE,
  row.total = FALSE,
  row.total.name = "Totaal",
  row.total.function = sum,
  row.total.widths = NULL,
  row.total.bold = TRUE,
  row.extra.header = list(values = NULL, widths = 1),
  row.extra.footer = list(values = NULL, widths = 1),
  column.names = colnames(x),
  column.names.bold = TRUE,
  columns.width = NULL,
  columns.percent = NULL,
  columns.italic = NULL,
  columns.bold = NULL,
  columns.fill = NULL,
  columns.fill.picker = "certeblauw5",
  columns.zebra = FALSE,
  column.total = FALSE,
  column.total.name = "Totaal",
  column.total.function = sum,
  column.total.bold = TRUE,
  align = "c",
  align.part = "all",
  caption = "",
  na = "",
  logicals = c("X", ""),
  round.numbers = 2,
  round.percent = 1,
  format.dates = "d mmm yyyy",
  decimal.mark = dec_mark(),
  big.mark = big_mark(),
  font.family = "Calibri",
  font.size = 9,
  font.size.header = font.size + 1,
  values.colour = NULL,
  values.colour.picker = "certeroze",
  values.fill = NULL,
  values.fill.picker = "certeroze3",
  values.bold = NULL,
  values.italic = NULL,
  autofit = is.null(columns.width) & is.null(rows.height),
  autofit.fullpage = TRUE,
  autofit.fullpage.width = 16,
  vline = NULL,
  vline.border = fp_border_default("black"),
  vline.part = c("body", "footer"),
  print = FALSE,
  ...
)

tbl_gtsummary_flextable(
  data,
  by,
  label = NULL,
  digits = 2,
  ...,
  decimal.mark = dec_mark()
)

gtsummary_as_flextable(gtsummary)

Arguments

x

a data.frame

row.names

row names to be displayed. Will be 1:nrow(x) if set to TRUE, but can be a vector of values.

row.names.bold

display row names in bold

rows.italic

column indexes of rows in italics

rows.bold

column indexes of rows in bold

rows.height

height of the rows in centimetres

rows.fill

the column indices of rows to be shaded

rows.fill.picker

the colour for values in rows.fill, is evaluated with colourpicker()

rows.zebra

banded rows in the body - equivalent to rows.fill = seq(2, nrow(x), 2)

row.total

add a row total (at the bottom of the table)

row.total.name

name of the row total

row.total.function

function used to calculate all numeric values per column (non-numeric columns are skipped)

row.total.widths

cell width in row total

row.total.bold

bold formatting of row total

row.extra.header

an extra header to be displayed above the table

row.extra.footer

an extra footer to show below the table

column.names

column names to be displayed. Can also be a named vector where the names are existing columns, or indices of columns. When this vector is smaller than ncol(x), only the first length(column.names) are replaced. When this vector is longer than ncol(x), all column names are replaced

column.names.bold

display column names in bold

columns.width

width of columns. For autofit.fullpage = TRUE, these are proportions to autofit.fullpage.width. For autofit.fullpage = FALSE, these are centimeters

columns.percent

display the column indices as percentages using format2() - example: columns.percent = c(2, 3)

columns.italic

column indices of columns to be displayed in italics

columns.bold

column indices of columns in bold

columns.fill

the column indices of rows to be shaded

columns.fill.picker

the colour for values in columns.fill, is evaluated with colourpicker()

columns.zebra

banded columns - equivalent to columns.fill = seq(2, ncol(x), 2)

column.total

adding a column total (to the right of the table)

column.total.name

name of the column total

column.total.function

function used to calculate all numeric values per row

column.total.bold

bold formatting of column total

align

default is "c", which aligns everything centrally. Use "r", "l", "c" and "j"/"u" (justify/align) to change alignment. Can be a vector or a character (like "lrrrcc")

align.part

part of the table where the alignment should take place ("all", "header", "body", "footer")

caption

table caption

na

text for missing values

logicals

vector with two values that replace TRUE and FALSE

round.numbers

number of decimal places to round up for numbers

round.percent

number of decimal places to round to when using columns.percent

format.dates

see format2()

decimal.mark

decimal separator, defaults to dec_mark()

big.mark

thousands separator, defaults to big_mark()

font.family

table font family

font.size

table font size

font.size.header

font size of header

values.colour, values.fill, values.bold, values.italic

values to be formatted

values.colour.picker, values.fill.picker

the colour for values in values.colour or values.fill, is evaluated with colourpicker()

autofit

format table in width automatically

autofit.fullpage

display table across width of page

autofit.fullpage.width

set number of centimetres to width of table

vline

indices of columns to have a vertical line to their right

vline.border

function to define vertical line

vline.part

part of the table where the vertical lines should be placed ("all", "header", "body", "footer")

print

forced printing (required in a `for`` loop)

...

Arguments passed on to gtsummary::tbl_summary()

data

A data frame

by

A column name (quoted or unquoted) in data. Summary statistics will be calculated separately for each level of the by variable (e.g. by = trt). If NULL, summary statistics are calculated using all observations. To stratify a table by two or more variables, use tbl_strata()

label

List of formulas specifying variables labels, e.g. list(age ~ "Age", stage ~ "Path T Stage"). If a variable's label is not specified here, the label attribute (attr(data$age, "label")) is used. If attribute label is NULL, the variable name will be used.

digits

List of formulas specifying the number of decimal places to round summary statistics. If not specified, tbl_summary guesses an appropriate number of decimals to round statistics. When multiple statistics are displayed for a single variable, supply a vector rather than an integer. For example, if the statistic being calculated is "{mean} ({sd})" and you want the mean rounded to 1 decimal place, and the SD to 2 use digits = list(age ~ c(1, 2)). User may also pass a styling function: digits = age ~ style_sigfig

gtsummary

a gtsummary object created with gtsummary::tbl_summary()

Value

flextable object

Details

Run tbl_markdown() on a flextable object to transform it into R Markdown and use it for Quarto or R Markdown reports.

tbl_gtsummary_flextable() creates a summary table with gtsummary::tbl_summary(), transforms it to a flextable and runs tbl_flextable() to apply the Certe style. The table will be Dutch if decimal.mark is a comma (default on Dutch systems). To turn a manual gtsummary into a 'Certe flextable', run gtsummary_as_flextable() on a gtsummary object.

See also

Examples

# generate a data.frame
df <- data.frame(text = LETTERS[1:10],
                 `decimal numbers` = runif(10, 0, 10),
                 `whole numbers` = as.integer(runif(10, 0, 10)),
                 `logical values` = as.logical(round(runif(10, 0, 1))),
                 dates = today() - runif(10, 200, 2000),
                 stringsAsFactors = FALSE)

# default
tbl_flextable(df)      # dataset has no row names
#> a flextable object.
#> col_keys: `A`, `B`, `C`, `D`, `E` 
#> header has 1 row(s) 
#> body has 10 row(s) 
#> original dataset sample: 
#>   A    B C D           E
#> 1 A 6.20 4   21 feb 2018
#> 2 B 9.99 2   27 mei 2019
#> 3 C 7.40 8 X  5 feb 2019
#> 4 D 1.93 4   19 jun 2019
#> 5 E 2.57 4    2 nov 2018
tbl_flextable(mtcars)  # dataset has row names
#> a flextable object.
#> col_keys: `A`, `B`, `C`, `D`, `E`, `F`, `G`, `H`, `I`, `J`, `K`, `L` 
#> header has 1 row(s) 
#> body has 32 row(s) 
#> original dataset sample: 
#>                   A    B C     D   E    F    G     H I J K L
#> 1         Mazda RX4 21.0 6 160.0 110 3.90 2.62 16.46 0 1 4 4
#> 2     Mazda RX4 Wag 21.0 6 160.0 110 3.90 2.88 17.02 0 1 4 4
#> 3        Datsun 710 22.8 4 108.0  93 3.85 2.32 18.61 1 1 4 1
#> 4    Hornet 4 Drive 21.4 6 258.0 110 3.08 3.21 19.44 1 0 3 1
#> 5 Hornet Sportabout 18.7 8 360.0 175 3.15 3.44 17.02 0 0 3 2

# print in markdown
df |> 
  tbl_flextable() |> 
  tbl_markdown()
#> 
#> 
#> ```{=html}
#> <div class="tabwid"><style>.cl-a34e37ba{}.cl-a3474ab8{font-family:'Calibri';font-size:10pt;font-weight:bold;font-style:normal;text-decoration:none;color:rgba(0, 0, 0, 1.00);background-color:transparent;}.cl-a3474acc{font-family:'Calibri';font-size:9pt;font-weight:normal;font-style:normal;text-decoration:none;color:rgba(0, 0, 0, 1.00);background-color:transparent;}.cl-a3476246{margin:0;text-align:center;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);padding-bottom:5pt;padding-top:5pt;padding-left:5pt;padding-right:5pt;line-height: 1;background-color:transparent;}.cl-a347a5e4{width:109.6pt;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-a347a5f8{width:79.3pt;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-a347a5f9{width:120.3pt;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-a347a602{width:44.3pt;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-a347a603{width:100pt;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-a347a60c{width:109.6pt;background-color:transparent;vertical-align: middle;border-bottom: 2pt solid rgba(102, 102, 102, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-a347a60d{width:120.3pt;background-color:transparent;vertical-align: middle;border-bottom: 2pt solid rgba(102, 102, 102, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-a347a60e{width:79.3pt;background-color:transparent;vertical-align: middle;border-bottom: 2pt solid rgba(102, 102, 102, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-a347a616{width:100pt;background-color:transparent;vertical-align: middle;border-bottom: 2pt solid rgba(102, 102, 102, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-a347a617{width:44.3pt;background-color:transparent;vertical-align: middle;border-bottom: 2pt solid rgba(102, 102, 102, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-a347a620{width:109.6pt;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-a347a621{width:44.3pt;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-a347a62a{width:120.3pt;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-a347a62b{width:79.3pt;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-a347a634{width:100pt;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-a347a635{width:120.3pt;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-a347a63e{width:79.3pt;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-a347a63f{width:44.3pt;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-a347a648{width:109.6pt;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-a347a649{width:100pt;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-a347a652{width:44.3pt;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-a347a65c{width:79.3pt;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-a347a65d{width:120.3pt;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-a347a666{width:109.6pt;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-a347a667{width:100pt;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-a347a67a{width:44.3pt;background-color:transparent;vertical-align: middle;border-bottom: 2pt solid rgba(102, 102, 102, 1.00);border-top: 2pt solid rgba(102, 102, 102, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-a347a684{width:120.3pt;background-color:transparent;vertical-align: middle;border-bottom: 2pt solid rgba(102, 102, 102, 1.00);border-top: 2pt solid rgba(102, 102, 102, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-a347a685{width:109.6pt;background-color:transparent;vertical-align: middle;border-bottom: 2pt solid rgba(102, 102, 102, 1.00);border-top: 2pt solid rgba(102, 102, 102, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-a347a686{width:100pt;background-color:transparent;vertical-align: middle;border-bottom: 2pt solid rgba(102, 102, 102, 1.00);border-top: 2pt solid rgba(102, 102, 102, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-a347a68e{width:79.3pt;background-color:transparent;vertical-align: middle;border-bottom: 2pt solid rgba(102, 102, 102, 1.00);border-top: 2pt solid rgba(102, 102, 102, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}</style><table class='cl-a34e37ba'>
#> ```
#> 
#> ```{=html}
#> <thead><tr style="overflow-wrap:break-word;"><td class="cl-a347a67a"><p class="cl-a3476246"><span class="cl-a3474ab8">text</span></p></td><td class="cl-a347a684"><p class="cl-a3476246"><span class="cl-a3474ab8">decimal.numbers</span></p></td><td class="cl-a347a685"><p class="cl-a3476246"><span class="cl-a3474ab8">whole.numbers</span></p></td><td class="cl-a347a686"><p class="cl-a3476246"><span class="cl-a3474ab8">logical.values</span></p></td><td class="cl-a347a68e"><p class="cl-a3476246"><span class="cl-a3474ab8">dates</span></p></td></tr></thead><tbody><tr style="overflow-wrap:break-word;"><td class="cl-a347a602"><p class="cl-a3476246"><span class="cl-a3474acc">A</span></p></td><td class="cl-a347a5f9"><p class="cl-a3476246"><span class="cl-a3474acc">6.20</span></p></td><td class="cl-a347a5e4"><p class="cl-a3476246"><span class="cl-a3474acc">4</span></p></td><td class="cl-a347a603"><p class="cl-a3476246"><span class="cl-a3474acc"></span></p></td><td class="cl-a347a5f8"><p class="cl-a3476246"><span class="cl-a3474acc">21 feb 2018</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-a347a602"><p class="cl-a3476246"><span class="cl-a3474acc">B</span></p></td><td class="cl-a347a5f9"><p class="cl-a3476246"><span class="cl-a3474acc">9.99</span></p></td><td class="cl-a347a5e4"><p class="cl-a3476246"><span class="cl-a3474acc">2</span></p></td><td class="cl-a347a603"><p class="cl-a3476246"><span class="cl-a3474acc"></span></p></td><td class="cl-a347a5f8"><p class="cl-a3476246"><span class="cl-a3474acc">27 mei 2019</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-a347a602"><p class="cl-a3476246"><span class="cl-a3474acc">C</span></p></td><td class="cl-a347a5f9"><p class="cl-a3476246"><span class="cl-a3474acc">7.40</span></p></td><td class="cl-a347a5e4"><p class="cl-a3476246"><span class="cl-a3474acc">8</span></p></td><td class="cl-a347a603"><p class="cl-a3476246"><span class="cl-a3474acc">X</span></p></td><td class="cl-a347a5f8"><p class="cl-a3476246"><span class="cl-a3474acc">5 feb 2019</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-a347a621"><p class="cl-a3476246"><span class="cl-a3474acc">D</span></p></td><td class="cl-a347a62a"><p class="cl-a3476246"><span class="cl-a3474acc">1.93</span></p></td><td class="cl-a347a620"><p class="cl-a3476246"><span class="cl-a3474acc">4</span></p></td><td class="cl-a347a634"><p class="cl-a3476246"><span class="cl-a3474acc"></span></p></td><td class="cl-a347a62b"><p class="cl-a3476246"><span class="cl-a3474acc">19 jun 2019</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-a347a63f"><p class="cl-a3476246"><span class="cl-a3474acc">E</span></p></td><td class="cl-a347a635"><p class="cl-a3476246"><span class="cl-a3474acc">2.57</span></p></td><td class="cl-a347a648"><p class="cl-a3476246"><span class="cl-a3474acc">4</span></p></td><td class="cl-a347a649"><p class="cl-a3476246"><span class="cl-a3474acc"></span></p></td><td class="cl-a347a63e"><p class="cl-a3476246"><span class="cl-a3474acc">2 nov 2018</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-a347a652"><p class="cl-a3476246"><span class="cl-a3474acc">F</span></p></td><td class="cl-a347a65d"><p class="cl-a3476246"><span class="cl-a3474acc">4.69</span></p></td><td class="cl-a347a666"><p class="cl-a3476246"><span class="cl-a3474acc">2</span></p></td><td class="cl-a347a667"><p class="cl-a3476246"><span class="cl-a3474acc">X</span></p></td><td class="cl-a347a65c"><p class="cl-a3476246"><span class="cl-a3474acc">13 aug 2019</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-a347a652"><p class="cl-a3476246"><span class="cl-a3474acc">G</span></p></td><td class="cl-a347a65d"><p class="cl-a3476246"><span class="cl-a3474acc">4.03</span></p></td><td class="cl-a347a666"><p class="cl-a3476246"><span class="cl-a3474acc">0</span></p></td><td class="cl-a347a667"><p class="cl-a3476246"><span class="cl-a3474acc"></span></p></td><td class="cl-a347a65c"><p class="cl-a3476246"><span class="cl-a3474acc">11 apr 2020</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-a347a652"><p class="cl-a3476246"><span class="cl-a3474acc">H</span></p></td><td class="cl-a347a65d"><p class="cl-a3476246"><span class="cl-a3474acc">7.09</span></p></td><td class="cl-a347a666"><p class="cl-a3476246"><span class="cl-a3474acc">8</span></p></td><td class="cl-a347a667"><p class="cl-a3476246"><span class="cl-a3474acc">X</span></p></td><td class="cl-a347a65c"><p class="cl-a3476246"><span class="cl-a3474acc">21 aug 2020</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-a347a602"><p class="cl-a3476246"><span class="cl-a3474acc">I</span></p></td><td class="cl-a347a5f9"><p class="cl-a3476246"><span class="cl-a3474acc">4.44</span></p></td><td class="cl-a347a5e4"><p class="cl-a3476246"><span class="cl-a3474acc">1</span></p></td><td class="cl-a347a603"><p class="cl-a3476246"><span class="cl-a3474acc">X</span></p></td><td class="cl-a347a5f8"><p class="cl-a3476246"><span class="cl-a3474acc">12 mei 2019</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-a347a617"><p class="cl-a3476246"><span class="cl-a3474acc">J</span></p></td><td class="cl-a347a60d"><p class="cl-a3476246"><span class="cl-a3474acc">2.98</span></p></td><td class="cl-a347a60c"><p class="cl-a3476246"><span class="cl-a3474acc">0</span></p></td><td class="cl-a347a616"><p class="cl-a3476246"><span class="cl-a3474acc"></span></p></td><td class="cl-a347a60e"><p class="cl-a3476246"><span class="cl-a3474acc">5 apr 2020</span></p></td></tr></tbody></table></div>
#> ```
#> 
#> 
#> 
  
# create a gtsummary as flextable
iris |> 
  tbl_gtsummary_flextable(Species, decimal.mark = ".")
#> a flextable object.
#> col_keys: `label`, `stat_1`, `stat_2`, `stat_3` 
#> header has 1 row(s) 
#> body has 4 row(s) 
#> original dataset sample: 
#>          label            stat_1            stat_2            stat_3
#> 1 Sepal.Length 5.00 (4.80, 5.20) 5.90 (5.60, 6.30) 6.50 (6.23, 6.90)
#> 2  Sepal.Width 3.40 (3.20, 3.68) 2.80 (2.53, 3.00) 3.00 (2.80, 3.18)
#> 3 Petal.Length 1.50 (1.40, 1.58) 4.35 (4.00, 4.60) 5.55 (5.10, 5.88)
#> 4  Petal.Width 0.20 (0.20, 0.30) 1.30 (1.20, 1.50) 2.00 (1.80, 2.30)
iris |> 
  tbl_gtsummary_flextable(Species, decimal.mark = ",")
#> a flextable object.
#> col_keys: `label`, `stat_1`, `stat_2`, `stat_3` 
#> header has 1 row(s) 
#> body has 4 row(s) 
#> original dataset sample: 
#>          label            stat_1            stat_2            stat_3
#> 1 Sepal.Length 5,00 (4,80; 5,20) 5,90 (5,60; 6,30) 6,50 (6,23; 6,90)
#> 2  Sepal.Width 3,40 (3,20; 3,68) 2,80 (2,53; 3,00) 3,00 (2,80; 3,18)
#> 3 Petal.Length 1,50 (1,40; 1,58) 4,35 (4,00; 4,60) 5,55 (5,10; 5,88)
#> 4  Petal.Width 0,20 (0,20; 0,30) 1,30 (1,20; 1,50) 2,00 (1,80; 2,30)
  
# extra formatting
tbl_flextable(df,
              logicals = c("X", "-"),     # replaces TRUE en FALSE
              values.colour = "X",
              values.fill = "X",
              row.names = "S. aureus",
              columns.italic = 1,
              format.dates = "ddd dd-mm-yy",
              round.numbers = 3)
#> a flextable object.
#> col_keys: `A`, `B`, `C`, `D`, `E`, `F` 
#> header has 1 row(s) 
#> body has 10 row(s) 
#> original dataset sample: 
#>           A B     C D E           F
#> 1 S. aureus A 6.195 4 - wo 21-02-18
#> 2 S. aureus B 9.992 2 - ma 27-05-19
#> 3 S. aureus C 7.395 8 X di 05-02-19
#> 4 S. aureus D 1.931 4 - wo 19-06-19
#> 5 S. aureus E 2.572 4 - vr 02-11-18

# row totals
tbl_flextable(df,
              row.total = TRUE,           # add row total
              row.total.function = max,   # instead of sum()
              row.total.name = "Maximum", # also works with dates
              columns.percent = 2,        # 2nd column as percentages
              round.percent = 0)          # rounding percentages
#> a flextable object.
#> col_keys: `A`, `B`, `C`, `D`, `E` 
#> header has 1 row(s) 
#> body has 10 row(s) 
#> original dataset sample: 
#>   A    B C D           E
#> 1 A 620% 4   21 feb 2018
#> 2 B 999% 2   27 mei 2019
#> 3 C 740% 8 X  5 feb 2019
#> 4 D 193% 4   19 jun 2019
#> 5 E 257% 4    2 nov 2018

# column names
tbl_flextable(df,
              column.names = c("1" = "Column 1",
                               "2" = "Column 2",
                               dates = "DATES!"))
#> a flextable object.
#> col_keys: `A`, `B`, `C`, `D`, `E` 
#> header has 1 row(s) 
#> body has 10 row(s) 
#> original dataset sample: 
#>   A    B C D           E
#> 1 A 6.20 4   21 feb 2018
#> 2 B 9.99 2   27 mei 2019
#> 3 C 7.40 8 X  5 feb 2019
#> 4 D 1.93 4   19 jun 2019
#> 5 E 2.57 4    2 nov 2018
tbl_flextable(df,
              column.names = LETTERS)
#> Warning: Only the first 5 items of column.names will be used
#> a flextable object.
#> col_keys: `A`, `B`, `C`, `D`, `E` 
#> header has 1 row(s) 
#> body has 10 row(s) 
#> original dataset sample: 
#>   A    B C D           E
#> 1 A 6.20 4   21 feb 2018
#> 2 B 9.99 2   27 mei 2019
#> 3 C 7.40 8 X  5 feb 2019
#> 4 D 1.93 4   19 jun 2019
#> 5 E 2.57 4    2 nov 2018

# vertical lines, alignment and row names
tbl_flextable(df,
              align = "lrrcc", # also works: c("l", "r", "r", "c", "c")
              font.size = 12,
              vline = c(2, 4),
              vline.part = "all",
              row.names = paste("Experiment", 1:10))
#> a flextable object.
#> col_keys: `A`, `B`, `C`, `D`, `E`, `F` 
#> header has 1 row(s) 
#> body has 10 row(s) 
#> original dataset sample: 
#>              A B    C D E           F
#> 1 Experiment 1 A 6.20 4   21 feb 2018
#> 2 Experiment 2 B 9.99 2   27 mei 2019
#> 3 Experiment 3 C 7.40 8 X  5 feb 2019
#> 4 Experiment 4 D 1.93 4   19 jun 2019
#> 5 Experiment 5 E 2.57 4    2 nov 2018

# width of cells and table
tbl_flextable(data.frame(test1 = "A", test2 = "B"),
              vline = 1,
              autofit.fullpage.width = 16, # default values in cm
              columns.width = c(1, 3))     # ratio; cells become 4 and 12 cm
#> a flextable object.
#> col_keys: `A`, `B` 
#> header has 1 row(s) 
#> body has 1 row(s) 
#> original dataset sample: 
#>   A B
#> 1 A B

tbl_flextable(data.frame(test1 = "A", test2 = "B"),
              vline = 1,
              autofit.fullpage = FALSE,    # no fullpage autofit
              columns.width = c(1, 3))     # cells become 1 and 3 cm
#> a flextable object.
#> col_keys: `A`, `B` 
#> header has 1 row(s) 
#> body has 1 row(s) 
#> original dataset sample: 
#>   A B
#> 1 A B
              
# adding extra header or footer
tbl_flextable(data.frame(test1 = "A", test2 = "B"),
              row.extra.header = list(values = c("Header", "Header"),
                                      widths = c(1, 1)),
              row.extra.footer = list(values = c("Footer", "Footer"),
                                      widths = c(1, 1)))
#> a flextable object.
#> col_keys: `A`, `B` 
#> header has 2 row(s) 
#> body has 1 row(s) 
#> original dataset sample: 
#>   A B
#> 1 A B