Summarise Table as gtsummary

Summarise a data.frame as gtsummary with Dutch defaults. These objects are based on the gt package by RStudio. To provide Certe style and compatibility with MS Word, use tbl_flextable() to transform the gtsummary object.

tbl_gtsummary(
  x,
  by = NULL,
  label = NULL,
  digits = 1,
  ...,
  language = "nl",
  column1_name = "Eigenschap",
  add_n = FALSE,
  add_p = FALSE,
  add_ci = FALSE,
  add_overall = FALSE,
  decimal.mark = dec_mark(),
  big.mark = big_mark()
)

Arguments

x

a data.frame

by

(tidy-select)
A single column from data. Summary statistics will be stratified by this variable. Default is NULL.

label

(formula-list-selector)
Used to override default labels in summary table, e.g. list(age = "Age, years"). The default for each variable is the column label attribute, attr(., 'label'). If no label has been set, the column name is used.

digits

(formula-list-selector)
Specifies how summary statistics are rounded. Values may be either integer(s) or function(s). If not specified, default formatting is assigned via assign_summary_digits(). See below for details.

...

Arguments passed on to gtsummary::tbl_summary()

language

the language to use, defaults to Dutch

column1_name

name to use for the first column

add_n

add the overall N using gtsummary::add_n()

add_p

add the p values gtsummary::add_p() (tests will be determined automatcally)

add_ci

add the confidence interval using gtsummary::add_ci()

add_overall

add the overall statistics using gtsummary::add_overall()

decimal.mark

decimal separator, defaults to dec_mark()

big.mark

thousands separator, defaults to big_mark()

Details

tbl_gtsummary() creates a summary table with gtsummary::tbl_summary(), to which different extra columns can be added e.g. with add_p = TRUE and add_overall = TRUE.

Examples

# These examples default to the Dutch language

iris |>
  tbl_gtsummary()
#> Warning: The `quiet` argument of `set_gtsummary_theme()` is deprecated as of gtsummary
#> 2.0.0.
#> ℹ Argument has been ignored.
#> ℹ The deprecated feature was likely used in the certetoolbox package.
#>   Please report the issue to the authors.

  

Characteristic
      	N = 1501
    

Sepal.Length
	5.8 (5.1, 6.4)

Sepal.Width
	3.0 (2.8, 3.3)

Petal.Length
	4.4 (1.6, 5.1)

Petal.Width
	1.3 (0.3, 1.8)

Species

    setosa
	50.0 (33.3%)

    versicolor
	50.0 (33.3%)

    virginica
	50.0 (33.3%)

1 Median (Q1, Q3); n (%)
    

iris |> 
  tbl_gtsummary(Species, add_p = TRUE)

  

Characteristic
      	setosa

N = 501
      	versicolor

N = 501
      	virginica

N = 501
      	p-value2
    

Sepal.Length
	5.0 (4.8, 5.2)
	5.9 (5.6, 6.3)
	6.5 (6.2, 6.9)
	<0.001

Sepal.Width
	3.4 (3.2, 3.7)
	2.8 (2.5, 3.0)
	3.0 (2.8, 3.2)
	<0.001

Petal.Length
	1.5 (1.4, 1.6)
	4.4 (4.0, 4.6)
	5.6 (5.1, 5.9)
	<0.001

Petal.Width
	0.2 (0.2, 0.3)
	1.3 (1.2, 1.5)
	2.0 (1.8, 2.3)
	<0.001

1 Median (Q1, Q3)
    

2 Kruskal-Wallis rank sum test
    

  
iris |> 
  tbl_gtsummary(Species, add_n = TRUE)

  

Characteristic
      	setosa

N = 501
      	versicolor

N = 501
      	virginica

N = 501
      	N
    

Sepal.Length
	5.0 (4.8, 5.2)
	5.9 (5.6, 6.3)
	6.5 (6.2, 6.9)
	150

Sepal.Width
	3.4 (3.2, 3.7)
	2.8 (2.5, 3.0)
	3.0 (2.8, 3.2)
	150

Petal.Length
	1.5 (1.4, 1.6)
	4.4 (4.0, 4.6)
	5.6 (5.1, 5.9)
	150

Petal.Width
	0.2 (0.2, 0.3)
	1.3 (1.2, 1.5)
	2.0 (1.8, 2.3)
	150

1 Median (Q1, Q3)
    

  
# support strata by providing 
iris2 <- iris
iris2$Category <- sample(LETTERS[1:2], size = 150, replace = TRUE)
head(iris2)
#>   Sepal.Length Sepal.Width Petal.Length Petal.Width Species Category
#> 1          5.1         3.5          1.4         0.2  setosa        B
#> 2          4.9         3.0          1.4         0.2  setosa        A
#> 3          4.7         3.2          1.3         0.2  setosa        B
#> 4          4.6         3.1          1.5         0.2  setosa        A
#> 5          5.0         3.6          1.4         0.2  setosa        B
#> 6          5.4         3.9          1.7         0.4  setosa        A

iris2 |> 
  tbl_gtsummary(c(Category, Species))

  

Characteristic
      	
        
A

      
      	
        
B

      
    

setosa

N = 271
      	versicolor

N = 221
      	virginica

N = 221
      	setosa

N = 231
      	versicolor

N = 281
      	virginica

N = 281
    

Sepal.Length
	5.0 (4.6, 5.1)
	5.8 (5.5, 6.1)
	6.6 (6.2, 7.1)
	5.1 (4.8, 5.3)
	6.1 (5.6, 6.4)
	6.4 (6.3, 6.9)

Sepal.Width
	3.4 (3.1, 3.6)
	2.8 (2.7, 3.0)
	3.0 (2.8, 3.2)
	3.5 (3.3, 3.7)
	2.9 (2.5, 3.0)
	3.0 (2.8, 3.2)

Petal.Length
	1.5 (1.3, 1.6)
	4.3 (4.0, 4.5)
	5.6 (5.3, 5.9)
	1.5 (1.4, 1.5)
	4.4 (4.0, 4.6)
	5.4 (5.1, 5.8)

Petal.Width
	0.2 (0.2, 0.3)
	1.4 (1.2, 1.5)
	2.0 (1.8, 2.1)
	0.2 (0.2, 0.3)
	1.3 (1.2, 1.5)
	2.1 (1.9, 2.3)

1 Median (Q1, Q3)
    

# transform to flextable 
# (formats to Certe style and allows rendering to Word)
iris |> 
  tbl_gtsummary(Species) |> 
  tbl_flextable()
#> Warning: Unknown or uninitialised column: `spanning_header`.
#> [1] "<div class=\"tabwid\"><style>.cl-f7c358dc{}.cl-f7bca10e{font-family:'Source Sans Pro';font-size:10pt;font-weight:bold;font-style:normal;text-decoration:none;color:rgba(255, 255, 255, 1.00);background-color:transparent;}.cl-f7bca122{font-family:'Source Sans Pro';font-size:6pt;font-weight:bold;font-style:normal;text-decoration:none;color:rgba(255, 255, 255, 1.00);background-color:transparent;position: relative;bottom:3pt;}.cl-f7bca12c{font-family:'Source Sans Pro';font-size:9pt;font-weight:normal;font-style:normal;text-decoration:none;color:rgba(0, 0, 0, 1.00);background-color:transparent;}.cl-f7bca12d{font-family:'Source Sans Pro';font-size:5.4pt;font-weight:normal;font-style:normal;text-decoration:none;color:rgba(0, 0, 0, 1.00);background-color:transparent;position: relative;bottom:2.7pt;}.cl-f7bf8464{margin:0;text-align:center;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);padding-bottom:5pt;padding-top:5pt;padding-left:5pt;padding-right:5pt;line-height: 1;background-color:transparent;}.cl-f7bf846e{margin:0;text-align:center;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);padding-bottom:5pt;padding-top:5pt;padding-left:5pt;padding-right:5pt;line-height: 1;background-color:transparent;}.cl-f7bfa2e6{width:1.761in;background-color:rgba(74, 100, 125, 1.00);vertical-align: middle;border-bottom: 2pt solid rgba(74, 100, 125, 1.00);border-top: 2pt solid rgba(74, 100, 125, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-f7bfa2f0{width:1.513in;background-color:rgba(74, 100, 125, 1.00);vertical-align: middle;border-bottom: 2pt solid rgba(74, 100, 125, 1.00);border-top: 2pt solid rgba(74, 100, 125, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-f7bfa2fa{width:1.761in;background-color:rgba(226, 231, 236, 1.00);vertical-align: top;border-bottom: 0.75pt solid rgba(255, 255, 255, 0.00);border-top: 0.75pt solid rgba(255, 255, 255, 0.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-f7bfa304{width:1.513in;background-color:rgba(226, 231, 236, 1.00);vertical-align: top;border-bottom: 0.75pt solid rgba(255, 255, 255, 0.00);border-top: 0.75pt solid rgba(255, 255, 255, 0.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-f7bfa305{width:1.761in;background-color:rgba(246, 247, 248, 1.00);vertical-align: top;border-bottom: 0.75pt solid rgba(255, 255, 255, 0.00);border-top: 0.75pt solid rgba(255, 255, 255, 0.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-f7bfa30e{width:1.513in;background-color:rgba(246, 247, 248, 1.00);vertical-align: top;border-bottom: 0.75pt solid rgba(255, 255, 255, 0.00);border-top: 0.75pt solid rgba(255, 255, 255, 0.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-f7bfa318{width:1.761in;background-color:rgba(246, 247, 248, 1.00);vertical-align: top;border-bottom: 2pt solid rgba(74, 100, 125, 1.00);border-top: 0.75pt solid rgba(255, 255, 255, 0.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-f7bfa319{width:1.513in;background-color:rgba(246, 247, 248, 1.00);vertical-align: top;border-bottom: 2pt solid rgba(74, 100, 125, 1.00);border-top: 0.75pt solid rgba(255, 255, 255, 0.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-f7bfa322{width:1.761in;background-color:transparent;vertical-align: middle;border-bottom: 0.75pt solid rgba(255, 255, 255, 0.00);border-top: 0.75pt solid rgba(255, 255, 255, 0.00);border-left: 0 solid rgba(255, 255, 255, 0.00);border-right: 0 solid rgba(255, 255, 255, 0.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-f7bfa323{width:1.513in;background-color:transparent;vertical-align: middle;border-bottom: 0.75pt solid rgba(255, 255, 255, 0.00);border-top: 0.75pt solid rgba(255, 255, 255, 0.00);border-left: 0 solid rgba(255, 255, 255, 0.00);border-right: 0 solid rgba(255, 255, 255, 0.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.tabwid {\n  font-size: initial;\n  padding-bottom: 1em;\n}\n\n.tabwid table{\n  border-spacing:0px !important;\n  border-collapse:collapse;\n  line-height:1;\n  margin-left:auto;\n  margin-right:auto;\n  border-width: 0;\n  border-color: transparent;\n  caption-side: top;\n}\n.tabwid-caption-bottom table{\n  caption-side: bottom;\n}\n.tabwid_left table{\n  margin-left:0;\n}\n.tabwid_right table{\n  margin-right:0;\n}\n.tabwid td, .tabwid th {\n    padding: 0;\n}\n.tabwid a {\n  text-decoration: none;\n}\n.tabwid thead {\n    background-color: transparent;\n}\n.tabwid tfoot {\n    background-color: transparent;\n}\n.tabwid table tr {\nbackground-color: transparent;\n}\n.katex-display {\n    margin: 0 0 !important;\n}</style><table data-quarto-disable-processing='true' class='cl-f7c358dc'><thead><tr style=\"overflow-wrap:break-word;\"><th class=\"cl-f7bfa2e6\"><p class=\"cl-f7bf8464\"><span class=\"cl-f7bca10e\">Characteristic</span></p></th><th class=\"cl-f7bfa2f0\"><p class=\"cl-f7bf8464\"><span class=\"cl-f7bca10e\">setosa</span><span class=\"cl-f7bca10e\">  </span><br><span class=\"cl-f7bca10e\">N = 50</span><span class=\"cl-f7bca122\">1</span></p></th><th class=\"cl-f7bfa2f0\"><p class=\"cl-f7bf8464\"><span class=\"cl-f7bca10e\">versicolor</span><span class=\"cl-f7bca10e\">  </span><br><span class=\"cl-f7bca10e\">N = 50</span><span class=\"cl-f7bca122\">1</span></p></th><th class=\"cl-f7bfa2f0\"><p class=\"cl-f7bf8464\"><span class=\"cl-f7bca10e\">virginica</span><span class=\"cl-f7bca10e\">  </span><br><span class=\"cl-f7bca10e\">N = 50</span><span class=\"cl-f7bca122\">1</span></p></th></tr></thead><tbody><tr style=\"overflow-wrap:break-word;\"><td class=\"cl-f7bfa2fa\"><p class=\"cl-f7bf846e\"><span class=\"cl-f7bca12c\">Sepal.Length</span></p></td><td class=\"cl-f7bfa304\"><p class=\"cl-f7bf846e\"><span class=\"cl-f7bca12c\">5.0 (4.8, 5.2)</span></p></td><td class=\"cl-f7bfa304\"><p class=\"cl-f7bf846e\"><span class=\"cl-f7bca12c\">5.9 (5.6, 6.3)</span></p></td><td class=\"cl-f7bfa304\"><p class=\"cl-f7bf846e\"><span class=\"cl-f7bca12c\">6.5 (6.2, 6.9)</span></p></td></tr><tr style=\"overflow-wrap:break-word;\"><td class=\"cl-f7bfa305\"><p class=\"cl-f7bf846e\"><span class=\"cl-f7bca12c\">Sepal.Width</span></p></td><td class=\"cl-f7bfa30e\"><p class=\"cl-f7bf846e\"><span class=\"cl-f7bca12c\">3.4 (3.2, 3.7)</span></p></td><td class=\"cl-f7bfa30e\"><p class=\"cl-f7bf846e\"><span class=\"cl-f7bca12c\">2.8 (2.5, 3.0)</span></p></td><td class=\"cl-f7bfa30e\"><p class=\"cl-f7bf846e\"><span class=\"cl-f7bca12c\">3.0 (2.8, 3.2)</span></p></td></tr><tr style=\"overflow-wrap:break-word;\"><td class=\"cl-f7bfa2fa\"><p class=\"cl-f7bf846e\"><span class=\"cl-f7bca12c\">Petal.Length</span></p></td><td class=\"cl-f7bfa304\"><p class=\"cl-f7bf846e\"><span class=\"cl-f7bca12c\">1.5 (1.4, 1.6)</span></p></td><td class=\"cl-f7bfa304\"><p class=\"cl-f7bf846e\"><span class=\"cl-f7bca12c\">4.4 (4.0, 4.6)</span></p></td><td class=\"cl-f7bfa304\"><p class=\"cl-f7bf846e\"><span class=\"cl-f7bca12c\">5.6 (5.1, 5.9)</span></p></td></tr><tr style=\"overflow-wrap:break-word;\"><td class=\"cl-f7bfa318\"><p class=\"cl-f7bf846e\"><span class=\"cl-f7bca12c\">Petal.Width</span></p></td><td class=\"cl-f7bfa319\"><p class=\"cl-f7bf846e\"><span class=\"cl-f7bca12c\">0.2 (0.2, 0.3)</span></p></td><td class=\"cl-f7bfa319\"><p class=\"cl-f7bf846e\"><span class=\"cl-f7bca12c\">1.3 (1.2, 1.5)</span></p></td><td class=\"cl-f7bfa319\"><p class=\"cl-f7bf846e\"><span class=\"cl-f7bca12c\">2.0 (1.8, 2.3)</span></p></td></tr></tbody><tfoot><tr style=\"overflow-wrap:break-word;\"><td  colspan=\"4\"class=\"cl-f7bfa322\"><p class=\"cl-f7bf8464\"><span class=\"cl-f7bca12d\">1</span><span class=\"cl-f7bca12c\">Median (Q1, Q3)</span></p></td></tr></tfoot></table></div>"
#> attr(,"class")
#> [1] "knit_asis"
#> attr(,"knit_cacheable")
#> [1] NA